### 2024-11-07 #671 @mpadge
pkgmatch: Find R Packages Matching Either Descriptions or Other R Packages
Submitting Author Name: Mark Padgham
Submitting Author Github Handle: @mpadge
Repository:  https://github.com/ropensci-review-tools/pkgmatch
Version submitted: 0.4.2
Submission type: Standard

Version accepted: TBD

---

```
Package: pkgmatch
Title:  Find R Packages Matching Either Descriptions or Other R Packages
Version: 0.4.2
    person("Mark", "Padgham", , "mark.padgham@email.com", role = c("aut", "cre"),
    person("Davis", "Vaughan", , "davis@posit.co", role = c("ctb"))
    )
Description: Find R packages matching either descriptions or other R packages.
    https://github.com/ropensci-review-tools/pkgmatch
    brio,
    checkmate,
    cli,
    curl,
    fs,
    httr2,
    memoise,
    pbapply,
    Rcpp,
    rvest,
    tokenizers,
    treesitter,
    treesitter.r,
    vctrs
    gert,
    hms,
    httptest2,
    piggyback,
    pkgbuild,
    rappdirs,
    roxygen2,
    withr,
    Rcpp
NeedsCompilation: yes
Language: en-GB
```

  - [x] rOpenSci tools

Data retrieval, because the package includes code to generate language model (LM) embeddings from all R packages retrieved from both CRAN and rOpenSci package repositories. Wrapper because LM embeddings are generated by wrapping interface to [ollama](https://ollama.com) software. Plus I've inserted a new, one-off category of "rOpenSci tools" for internal, staff-curated packages.

Beyond internal rOpenSci use, target audiences are (1) entirely general audience of those interested in searching R packages using either text or code input, and (2) package developers, who can use this package to identify similar packages or functions to code they might be working on.

No, not at all. There are to my knowledge two other R packages for interfacing with LMs: [tidyllm](https://github.com/edubruell/tidyllm) and [elmer](https://github.com/tidyverse/elmer). Both of these are general interfaces to LM API endpoints, while this package specifically uses LM outputs to identify best-matching packages.
