### 2024-11-08 8c52618
update phrases in good-enough vignette
