### 2024-11-11 33b84df
update readme outputs for #78
