### 2024-11-11 8b1cf5c
update vignette output to close #78
