### 2024-11-11 84331b7
fix cran-update fn when no new pkgs
