### 2024-11-27 ff09184
add options to get/set ollama url for #83
