### 2024-11-27 45b74d9
use 'ollama_get_url' value in ollama fns for #83
