### 2024-11-27 72d7820
describe ollama endpoint for #83 in vignette
