### 2024-11-27 225c19d
add 'check_ollama_url' for #83
