+++ R/zzz.R
@@ -48,9 +48,27 @@ get_ollama_url <- function () {
 #' @export
 set_ollama_url <- function (ollama_url) {
 
+    check_ollama_url (ollama_url)
+
     op <- options ()
     op.pkgmatch <- list (pkgmatch.ollama.url = ollama_url)
     options (op.pkgmatch)
 
     invisible (ollama_url)
+}
+
+check_ollama_url <- function (ollama_url) {
+
+    ollama_url <- ifelse (
+        grepl ("^http", ollama_url),
+        ollama_url,
+        paste0 ("https://", ollama_url)
+    )
+    check <- tryCatch (
+        curl::curl_parse_url (ollama_url),
+        error = function (e) e
+    )
+    if (is (check, "error")) {
+        cli::cli_abort (check$message)
+    }
 }
