### 2024-11-27 5bc5055
test ollama urls to close #83
