+++ tests/testthat/_snaps/ollama-endpoint.md
@@ -1,0 +1,8 @@
+# ollama endpoints
+
+    Code
+      set_ollama_url(new_url)
+    Condition
+      Error in `check_ollama_url()`:
+      ! Failed to parse URL: Bad scheme
+
