### 2024-11-27 ac37f4b
tidy man entries for #83
