+++ man/ollama_check.Rd
@@ -17,7 +17,7 @@ before returning.
 \description{
 Note that the URL of a locally-running ollama instance is presumed by
 default to be "127.0.0.1:11434". Other values can be set using the
-\link{ollama_set_url} function.
+\link{set_ollama_url} function.
 }
 \examples{
 \dontrun{
