+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: pkgmatch
 Title:  Find R Packages Matching Either Descriptions or Other R Packages
-Version: 0.4.2.013
+Version: 0.4.2.014
 Authors@R: c(
     person("Mark", "Padgham", , "mark.padgham@email.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2172-5265")),
@@ -20,6 +20,7 @@ Imports:
     fs,
     httr2,
     memoise,
+    methods,
     pbapply,
     Rcpp,
     rvest,
