### 2024-11-27 8dfbb21
Allow getting ollama endpoint via an env var
