### 2024-12-04 #671 @MargaretSiple-NOAA
- [ ] **Installation instructions:** Are installation instructions clear enough for human users? - *Yes but see notes below.*
- [ ] **Tests**: If the package has some interactivity / HTTP / plot production etc. are the tests using [state-of-the-art tooling](https://devguide.ropensci.org/building.html#testing)? *Yes but see notes below. I think I just can't fully evaluate this until I can install docker and ollama on my personal computer and test the functions*
---

Nice package, @mpadge ! This will be useful for anyone working on package development or review. I wrote a few notes on what I saw during the editor check process-- they're longer than I usually write, but I figure these will probably make the reviewers' lives easier so it doesn't hurt to mention them early.

A few notes of mine from the editor check process:
- The pkgdown page seems to indicate that the user can set the corpus to rOpenSci or to CRAN, but some functions only provide search results for rOpenSci, e.g., `pkgmatch_bm25()` requires the corpus to be rOpenSci-related. Some clarity about which functions apply to which package corpi (?) might be helpful.
- I got 1 note from running `devtools::check()`: "Package suggested but not available for checking: 'piggyback'" -- I don't worry about these too much just wanted to flag it just in case.

- The ollama connection gave me some trouble at the start, so I imagine future users might struggle on this. I recommend revising the documentation a little bit to make it more obvious what needs to happen with ollama before users can get started: a) Consider adding a sentence to the top of the "Get started" article that indicates that docker and ollama installs need to happen first and foremost. Here, I would link to the "ollama" article. b) Change the title of the "ollama" article to say something like "Before you begin: ollama" or "Before you begin: ollama installation", so that when people are browsing, it's very clear to them that they'll need to read that article first.
- The beginning of the [Docker section](https://docs.ropensci.org/pkgmatch/articles/ollama.html#docker) in the ollama article should indicate that users need to have Docker installed as well. I learned from this process that my institution actually doesn't have a Docker license! This may impede editors/reviewers who work at NOAA. It sounds like they're working on getting a license so that may change in the future, so that will probably change in the future.

- I had some trouble getting test coverage my usual way (`covr::package_coverage()` fails with an error about a temp file)... but `devtools::test()` works great). Some tests are failing but I'm pretty sure that's because I don't have ollama properly installed on my machine. It might be nice to have a one-line check in one of the early vignettes that can show people whether they are missing any of the components need to run the fns.

---
