### 2024-12-05 37166de
more bm25 doc updates for #88
