+++ man/pkgmatch_bm25.Rd
@@ -20,7 +20,11 @@ rOpenSci corpus will be automatically downloaded and used.}
 
 \item{corpus}{If \code{txt} is not specified, data for nominated corpus will be
 downloaded to local cache directory, and BM25 values calculated against
-those. Must be one of "ropensci", "ropensci-fns", or "cran".}
+those. Must be one of "ropensci", "ropensci-fns", or "cran". Note that the
+"ropensci-fns" corpus contains entries for every single function of every
+rOpenSci package, and the resulting BM25 values can be used to determine the
+best-matching function. The other two corpora are package-based, and the
+results can be used to find the best-matching package.}
 }
 \value{
 A \code{data.frame} of package names and 'BM25' measures against text
