### 2024-12-05 85abf7b
update docs for similar_fns fn for #88
