### 2024-12-05 6b94d0a
rename & restructure ollama article for #88
