### 2024-12-05 c07416f
explicit min versions of curl and R
Respectively for 'curl_parse_url' and for native pipe
