### 2024-12-05 7340667
add tests/README for #88
