### 2024-12-05 #671 @mpadge
Thanks @MargaretSiple-NOAA for really useful feedback! The issue linked above has details of changes made in response. I'd say the most important of those for reviewers of this package is that I've added [`tests/README`](https://github.com/ropensci-review-tools/pkgmatch/tree/main/tests) explaining that tests can be run without having `ollama` installed or running anywhere.
