### 2024-12-05 00c1b12
start 'pkgmatch_generate_data' fn for #86
