+++ R/data-generate.R
@@ -1,0 +1,136 @@
+#' Generate all data necessary to run "pkgmatch" on a local corpus of R
+#' packages.
+#'
+#' The function generates and stores required data in local cache directory,
+#' determined by environment variable "PKGMATCH_CACHE_DIR" if that exists,
+#' otherwise default of `file.path(rappdirs::user_cache_dir(), "R",
+#' "pkgmatch")`.
+#'
+#' @param path Path to a local directory containing R packages, each in its own
+#' sub-directory.
+#' @param corpus Name of corpus. This is used to name resultant data files.
+#' @param fns Should embeddings also be generated for individual functions of
+#' each package? For large corpora, the resultant files can be prohibitively
+#' large, and so this should only be set to `TRUE` for sufficiently small
+#' corpora. The rOpenSci corpus is under 400 packages, and function-level
+#' embeddings are possible, while CRAN with over 20,000 packages is too large.
+#' @param num_cores Number of cores to be used in parallel computation of data.
+#'
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' path <- "<path>/<to>/<several>/<R-pacakges>"
+#' pkgmatch_generate_data (path)
+#' }
+pkgmatch_generate_data <- function (path,
+                                    corpus = NULL,
+                                    fns = FALSE,
+                                    num_cores = parallel::detectCores ()) {
+
+    checkmate::assert_character (corpus, len = 1L)
+    checkmate::assert_directory_exists (path)
+    packages <- fs::dir_ls (path, type = "directory", recurse = FALSE)
+    cache_dir <- pkgmatch_cache_path ()
+
+    cli::cli_alert_info ("Generating package embeddings ...")
+    embeddings <- pkgmatch_embeddings_from_pkgs (packages)
+    f <- fs::path (cache_dir, paste0 ("embeddings-", corpus, ".Rds"))
+    saveRDS (embeddings, f)
+    cli::cli_alert_success ("Done")
+    cli::cli_inform ("")
+
+    if (fns) {
+        cli::cli_alert_info ("Generating function-level embeddings ...")
+        embeddings_fns <-
+            pkgmatch_embeddings_from_pkgs (packages, functions_only = TRUE)
+        f <- fs::path (cache_dir, paste0 ("embeddings-fns-", corpus, ".Rds"))
+        saveRDS (embeddings_fns, f)
+        cli::cli_alert_success ("Done")
+        cli::cli_inform ("")
+    }
+
+    cli::cli_alert_info ("Generating BM25 values ...")
+    cl <- parallel::makeCluster (num_cores)
+    txt_with_fns <- pbapply::pblapply (
+        packages,
+        function (p) pkgmatch:::get_pkg_text (p),
+        cl = cl
+    )
+    parallel::stopCluster (cl)
+
+    txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
+    idfs <- list (
+        with_fns = bm25_idf (txt_with_fns),
+        wo_fns = bm25_idf (txt_wo_fns)
+    )
+    token_lists <- list (
+        with_fns = bm25_tokens_list (txt_with_fns),
+        wo_fns = bm25_tokens_list (txt_wo_fns)
+    )
+    bm25_data <- list (idfs = idfs, token_lists = token_lists)
+    f <- fs::path (cache_dir, paste0 ("bm25-", corpus, ".Rds"))
+    saveRDS (bm25_data, f)
+
+    if (fns) {
+        txt_fns <- get_all_fn_descs (txt_with_fns)
[…]
