+++ man/pkgmatch_generate_data.Rd
@@ -1,0 +1,32 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data-generate.R
+\name{pkgmatch_generate_data}
+\alias{pkgmatch_generate_data}
+\title{Generate all data necessary to run "pkgmatch" on a local corpus of R
+packages.}
+\usage{
+pkgmatch_generate_data(path, corpus = NULL, fns = FALSE)
+}
+\arguments{
+\item{path}{Path to a local directory containing R packages, each in its own
+sub-directory.}
+
+\item{corpus}{Name of corpus. This is used to name resultant data files.}
+
+\item{fns}{Should embeddings also be generated for individual functions of
+each package? For large corpora, the resultant files can be prohibitively
+large, and so this should only be set to \code{TRUE} for sufficiently small
+corpora. The rOpenSci corpus is under 400 packages, and function-level
+embeddings are possible, while CRAN with over 20,000 packages is too large.}
+}
+\description{
+The function generates and stores required data in local cache directory,
+determined by environment variable "PKGMATCH_CACHE_DIR" if that exists,
+otherwise default of \code{file.path(rappdirs::user_cache_dir(), "R", "pkgmatch")}.
+}
+\examples{
+\dontrun{
+path <- "<path>/<to>/<several>/<R-pacakges>"
+pkgmatch_generate_data (path)
+}
+}
