+++ .github/workflows/test-coverage.yaml
@@ -40,7 +40,8 @@ jobs:
             install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package"),
             line_exclusions = list (
               'R/ollama.R',
-              'R/browse.R'
+              'R/browse.R',
+              'R/data-generate.R'
             )
           )
           covr::to_cobertura(cov)
