+++ NAMESPACE
@@ -9,6 +9,7 @@ export(pkgmatch_bm25_fn_calls)
 export(pkgmatch_browse)
 export(pkgmatch_embeddings_from_pkgs)
 export(pkgmatch_embeddings_from_text)
+export(pkgmatch_generate_data)
 export(pkgmatch_load_data)
 export(pkgmatch_similar_fns)
 export(pkgmatch_similar_pkgs)
