### 2024-12-05 1171221
restructure data-generate.R for #86 into module sub-fns
