### 2024-12-05 2be5278
fix return val of 'generate_data' fn #86
