### 2025-01-08 ropensci-review-tools/pkgmatch#90 @MargaretSiple-NOAA
Errors in devtools::test()
Hi @mpadge -- I think this is the appropriate place to add this issue (with reference to https://github.com/ropensci/software-review/issues/671 ). I recently pulled the updates you made after my initial review and started re-checking everything, and I have run into an issue during `devtools::check()` that wasn't present before. I started with a local build from source (I'm using the Build tab in RStudio).

- Everything is good with the install.
- Everything is good with `devtools::check()`
- When I run `devtools::test()`, I get several errors about functions that are included in `{httptest2}`, like `with_mock_dir()` :
![image](https://github.com/user-attachments/assets/3c3b8418-e921-4db4-b3c0-f686ad4fb326)

I am stumped because the testthat.R script loads `{httptest2}`. Is this a problem with how I'm running the tests? I don't recall this issue happening before, so I tested it again on a different machine and was able to repeat the error.

I am running R 4.4.2 and have replicated this error on a different machine with R 4.4.1. Any guidance you can provide would be much appreciated!
