### 2025-01-08 #671 @MargaretSiple-NOAA
Hi @mpadge -- I was just revising my editor checks above after your revisions, and everything is looking good except some issues I had with `devtools::test()` - these are documented in the issue I logged on the pkgmatch github page (above). I don't know what they mean but I think we should have a resolution before reviewers are running similar checks.
