### 2025-01-09 cfff102
namespace 'with_mock_dir' to fix #90
