### 2025-01-09 ropensci-review-tools/pkgmatch#90 @mpadge
Thanks @MargaretSiple-NOAA. The problem is because both `devtools::test()` and `testthat::test_local()` use default parameters from [`pkgload::load_all()`](https://pkgload.r-lib.org/reference/load_all.html), and these load all *Imports*, but no *Suggests* (other than `testthat`). The `use_mock_dir()` function is from `httptest2`, which is in *Suggests*, so not found using those fns.

Anyway, the above commit fixes everything by just adding explicit namespaces to all `with_mock_dir()` calls. That procedure is a very generally good idea that i'll use from here on - thank you!
