### 2025-01-09 #671 @mpadge
Thanks @MargaretSiple-NOAA, I've fixed the issue linked above, so `devtools::test()` should now work for everybody.
