### 2025-01-14 #671 @MargaretSiple-NOAA
Thank for for addressing that , @mpadge ! The tests now run without issues.

Two more quick things:
- `pkgmatch_similar_pkgs(..., corpus = "cran")` always gives a warning `Error in matrix(emb[[what]], nrow = nrow, ncol = npkgs) :
  non-numeric matrix extent`. This renders fine on the pkgdown page [here](https://docs.ropensci.org/pkgmatch/articles/pkgmatch.html) so I suspect it is a computer-specific issue. Everything works as expected when `corpus = "ropensci"`.
- Several of the tests fail and trace back to `parseNamespaceFile(pkg_name, package.lib = lp)`. I think it's fine to fix this when reviews are in but wanted to give you a heads-up while I'm still searching for a Reviewer 2.
