### 2025-01-15 ropensci-review-tools/pkgmatch#92 @mpadge
similar_pkgs bug
Reported by @MargaretSiple-NOAA in [this review comment](https://github.com/ropensci/software-review/issues/671#issuecomment-2591329359), and able to be reproduced locally:

``` r
library (pkgmatch)
packageVersion ("pkgmatch")
#> [1] '0.4.2.22'
input <- "Package do download open agricultural data including crop time series"
pkgmatch_similar_pkgs (input, corpus = "cran")
#> Error in matrix(this_vec, nrow = nrow, ncol = ncol): non-numeric matrix extent
```
