### 2025-01-15 ropensci-review-tools/pkgmatch#92 @mpadge
Oh dear, it seems the daily data updates have stopped working properly, and the embeddings data for CRAN are no longer properly structured, meaning the whole workflow has been failing to update. Last successful update was Nov 11th 2024. I'll have to re-generate all data to fix and then restart the updating process ...
