### 2025-01-15 ac5d906
add 'latest_only' param to 'update_cran' fn for #92
