### 2025-01-15 602aa3a
more fixes to 'list_new_cran_updates' fn for #92
