### 2025-01-15 af01d84
use curl and not httr2 to download tarballs [ci skip]
https://github.com/r-lib/httr2/issues/487
