### 2025-01-15 fd33323
all 'closeAllConnections' call to cran update loop for #92 [ci skip]
