### 2025-01-15 ropensci-review-tools/pkgmatch#92 @MargaretSiple-NOAA
Oh dear. Let me know if you'd like to pause the review process, if you think those updates will take a while.
