### 2025-01-16 3117d01
run update workflow on Mon and Thurs
