+++ .github/workflows/update.yaml
@@ -3,7 +3,7 @@ on:
     branches:
       - main
   schedule:
-    - cron: 0 0 * * 1
+    - cron: 0 0 * * 1,4
 
 name: Update pkgmatch data
 
