### 2025-01-16 afa5f66
update ropensci fn embeddings for #92
