+++ R/data-update-ropensci.R
@@ -41,17 +41,25 @@ pkgmatch_update_ropensci <- function () {
             extract_data_from_local_dir (pkg_dir),
             error = function (e) NULL
         )
+        fns <- tryCatch (
+            pkgmatch_embeddings_from_pkgs (pkg_dir, functions_only = TRUE),
+            error = function (e) NULL
+        )
         fs::dir_delete (pkg_dir)
 
         pkgmatch_update_progress_message (i, 1, nrow (reg_updated), pt0)
 
-        return (dat)
+        return (list (dat = dat, fns = fns))
     })
+    fns <- do.call (cbind, lapply (res, function (i) i$fns))
+    res <- lapply (res, function (i) i$dat)
     names (res) <- reg_updated$name
 
     append_data_to_embeddings (res, flist, cran = FALSE)
     append_data_to_bm25 (res, flist, cran = FALSE)
     append_data_to_fn_calls (res, flist, cran = FALSE)
+
+    append_data_to_fn_embeddings (fns, flist)
 
     options ("rlib_message_verbosity" = op)
 
@@ -93,4 +101,18 @@ days_in_this_month <- function (today = Sys.Date ()) {
     next_month <- as.Date (paste0 (next_year, "-", next_month, "-", "1"))
     dates <- seq (this_month, next_month, by = "day")
     return (length (dates) - 1L)
+}
+
+append_data_to_fn_embeddings <- function (fns, flist) {
+
+    pkgs <- unique (gsub ("::.*$", "", colnames (fns)))
+    f <- grep ("embeddings\\-fns\\.", flist, value = TRUE)
+    embeddings_fns <- readRDS (f)
+    ptn <- paste0 (paste0 (pkgs, "::"), collapse = "|")
+    index <- which (!grepl (ptn, colnames (embeddings_fns)))
+    embeddings_fns <- cbind (embeddings_fns [, index], fns)
+
+    index <- order (colnames (embeddings_fns))
+    embeddings_fns <- embeddings_fns [, index]
+    saveRDS (f, embeddings_fns)
 }
