### 2025-01-16 b22e8b8
update workflow to daily cron runs
