### 2025-01-16 4f2ada1
fix data save in append_data_to_fn_embeddings for #92
