### 2025-01-16 3e4501b
fix 'opt_is_quiet' reset in similar_fns fn
