### 2025-01-16 c406fa1
fix another verbosity control reset option
