### 2025-01-16 872ce02
only run update workflow on cron, not on push
