+++ .github/workflows/update.yaml
@@ -1,7 +1,7 @@
 on:
-  push:
-    branches:
-      - main
+  # push:
+  #   branches:
+  #     - main
   schedule:
     - cron: 1 0 * * *
 
