### 2025-01-16 109072e
add 'pkgmatch_update_cache' fn for #92
