+++ NAMESPACE
@@ -13,6 +13,7 @@ export(pkgmatch_load_data)
 export(pkgmatch_similar_fns)
 export(pkgmatch_similar_pkgs)
 export(pkgmatch_treesitter_fn_tags)
+export(pkgmatch_update_cache)
 export(pkgmatch_update_data)
 export(set_ollama_url)
 export(text_is_code)
