### 2025-01-16 ropensci-review-tools/pkgmatch#92 @mpadge
You may (and should!) now run `pkgmatch_update_cache()`. After that, you should get the following results:
``` r
library (pkgmatch)
packageVersion ("pkgmatch")
#> [1] '0.4.3.2'
input <- "Package do download open agricultural data including crop time series"
pkgmatch_similar_pkgs (input, corpus = "ropensci")
#> [1] "rsat"          "rdhs"          "rnaturalearth" "smapr"
#> [5] "tacmagic"
pkgmatch_similar_pkgs (input, corpus = "cran")
#> [1] "cropdatape" "ifo"        "bundesbank" "CropScapeR" "dataseries"
pkgmatch_similar_fns (input)
#> [1] "rsnps::fetch_genotypes"           "rgbif::occ_download_datasets"
#> [3] "weatherOz::get_available_imagery" "cde::ea_wbids"
#> [5] "getCRUCLdata::get_CRU_stack"
```

(These are subject to change, as data are updated daily, so no result is guaranteed to be reproducible from one day to the next ...)
