### 2025-01-17 9c53c93
clean verbosity options code for update fns #92
