+++ R/similarity-metrics.R
@@ -20,12 +20,6 @@ similarity_embeddings <- function (input, embeddings, input_is_code) {
     return (dat)
 }
 
-similarity_bm25 <- function (input, bm25, idfs = NULL) {
-
-    b <- pkgmatch_bm25 (input = input, idfs = idfs)
-    this_emb <- get_embeddings (input, code = FALSE)
-}
-
 #' cosine similarity between one input vector and an input matrix with column
 #' names.
 #' @noRd
