### 2025-01-17 b684623
fix 'opt_is_quiet' fn
