### 2025-01-20 4220cc2
turn off all console output on gha
