### 2025-01-23 7ee87a9
do not call 'progress_message' fn at all on gha; closes #99
