### 2025-02-02 #671 @maurolepore
Dear @mpadge this is to mark the start of my EiC rotation. I'm reviewing all open issues and noting what I see.

I see one review is due about now and the second one on Feb 21st. It all looks good to me. I'll step back. Thanks!
