### 2025-02-03 1c6b69d
insert wait between all ollama steps in update workflow
