### 2025-02-07 cf198bd
update 'update_cache' docs to fix #103 [ci skip]
