+++ man/print.pkgmatch.Rd
@@ -31,6 +31,7 @@ Other utils:
 \code{\link{head.pkgmatch}()},
 \code{\link{pkgmatch_browse}()},
 \code{\link{pkgmatch_load_data}()},
+\code{\link{pkgmatch_update_cache}()},
 \code{\link{text_is_code}()}
 }
 \concept{utils}
