+++ man/head.pkgmatch.Rd
@@ -31,6 +31,7 @@ head (p) # Shows first 5 rows of full `data.frame` object
 Other utils: 
 \code{\link{pkgmatch_browse}()},
 \code{\link{pkgmatch_load_data}()},
+\code{\link{pkgmatch_update_cache}()},
 \code{\link{print.pkgmatch}()},
 \code{\link{text_is_code}()}
 }
