+++ man/text_is_code.Rd
@@ -29,6 +29,7 @@ Other utils:
 \code{\link{head.pkgmatch}()},
 \code{\link{pkgmatch_browse}()},
 \code{\link{pkgmatch_load_data}()},
+\code{\link{pkgmatch_update_cache}()},
 \code{\link{print.pkgmatch}()}
 }
 \concept{utils}
