### 2025-02-11 bcbc3ce
set 'verbose_limit' as pkg option rather than hard-coded
