### 2025-02-11 c4215ce
also test error on undefined verbose_limit option
