### 2025-02-12 #671 @Selbosh
@mpadge Having some issues getting `ollama` to cooperate on Ubuntu; unfortunately this also causes the package to give a cryptic error message.

```r
pkgmatch_similar_pkgs(input)
```
```
Error in names(out) <- nms :
  'names' attribute [1] must be the same length as the vector [0]
```
Found some documentation [here](https://pypi.org/project/ollamac/) so will try to figure it out and get the review done soon.
