### 2025-02-12 #671 @Selbosh
I think I found one half of the the problem: `set_ollama_url` does not modify the system environment variables, so the system call to `ollama list` within `ollama_check()` will fail depending on how you have set up your shell environment variables (e.g. if you `export OLLAMA_HOST=localhost:8888` but then open R in another session, then the exported variable is lost).

In addition, `ollama_check()` fails and gives the error above if `ollama` is running, but without any models downloaded (it's trying to iterate over a list of length zero, I guess?)

So the current sequence, after setting `ollama serve` running on `localhost:8888` in another shell and manually pulling the two embeddings in the vignette, is:

```r
Sys.getenv('OLLAMA_URL')
## [1] ""
library(pkgmatch)
ollama_check()
## Error: could not connect to ollama app, is it running?
## Error in out[[1]] : subscript out of bounds
## In addition: Warning message:
## In system("ollama list", intern = TRUE) :
##  running command 'ollama list' had status 1
set_ollama_url('127.0.0.1:8888')
ollama_check()
## Error: could not connect to ollama app, is it running?
## Error in out[[1]] : subscript out of bounds
## In addition: Warning message:
## In system("ollama list", intern = TRUE) :
##  running command 'ollama list' had status 1
Sys.setenv(OLLAMA_HOST = '127.0.0.1:8888')
ollama_check()
## [1] TRUE
```

But then I still do not get much farther, because
```r
pkgmatch_similar_pkgs(input)
## Error in `httr2::req_perform()`:
## ! Failed to perform HTTP request.
## Caused by error in `curl::curl_fetch_memory()`:
## ! Couldn't connect to server [localhost]: Failed to connect to localhost port 11434 after 1 ms: Connection refused
## Run `rlang::last_trace()` to see where the error occurred.
```
which implies that there is least one other place that `11434` is hard-coded. A quick search found

https://github.com/ropensci-review-tools/pkgmatch/blob/21c1e138535f86de6501ec8549f7c8a6f6ac4bf9/R/embeddings.R#L240

which unfortunately means I can't proceed on Ubuntu without fixing this bug first.
