### 2025-02-12 ropensci-review-tools/pkgmatch#106 @Selbosh
ollama host url is hardcoded
The following line is problematic because it does not respect the system environment variables, which means the package does not work at all if you are unable to run Ollama on port 11434, as appears to be the case [on Ubuntu](https://github.com/ollama/ollama/issues/707).

https://github.com/ropensci-review-tools/pkgmatch/blob/21c1e138535f86de6501ec8549f7c8a6f6ac4bf9/R/embeddings.R#L240

Suggest changing to

```r
u <- file.path(getOption("pkgmatch.ollama.url"), "api/embeddings")
```

or similar
