### 2025-02-12 #671 @Selbosh
Hello @mpadge and @MargaretSiple-NOAA! Thank you giving me the opportunity to review the package. It is my first package review, and I am still learning the process, so I hope that my feedback is useful. As mentioned above, I had a couple of issues getting started, but have found some fixes that should with any luck make it easier for others who encounter similar problems.

Please let me know if I can clarify anything. Here are my comments:

---

In the `README`, the statement of need should be more prominent; it's not obvious who the target audience of the package is or why the package is needed. It would be great if you could explain early on, clearly, what problem `pkgmatch` solves, maybe giving examples of use cases from the start.

I think that you could improve the documentation by making it more self-contained and making it easier for the user to find information quickly. Some information that should be in the `README` is only in a vignette, and some information that should be at the start of one vignette is instead left as a link to another vignette. The dependence on LLMs is also somewhat buried: this could be a make-or-break feature for many users, but it's only mentioned at the bottom of the Installation section. I can imagine it could be a selling point that `pkgmatch` exploits LLMs, but it could also be a deal-breaker for some users either due to technical limitations or ethical concerns.

The package relies on `ollama` but it's not explained in the README what this is or why the package should depend on it, or whether you can use some functions without it. There is a function called `ollama_check()` that ostensibly checks for and installs `ollama`: add this to the `README`, as it seems like the first thing you would want to run after loading the package! "Procedures for setting that up are described in a separate vignette" is not super helpful without a direct link to or giving the name of said vignette! Ideally I'd put much of this information in the README, because it could be a decider on whether somebody can/will install the package in the first place. Given we are talking about local LLMs here, not everybody will know what their computational demands are, so it would be helpful to give some general indication of system requirements in the 'Before you begin' vignette and maybe in the README.

Generally speaking, I think the documentation assumes a bit too much knowledge on the part of the user, by not motivating the task clearly enough, not explaining what some functions do or what some terms mean. There are one or two functions documented in terms of what they are _not_ rather than what they _are_, which can be a bit confusing (e.g. `pkgmatch_bm25_fn_calls()`). Some function descriptions (e.g. `pkgmatch_bm25()`) simply comprise an external Web link, which is not ideal for offline use, nor for users who want to know what a function does without having to click around. For the same functions, the current Titles in the documentation are too long or specific and would be better off as Descriptions, and the current Descriptions as Details or Notes.

The purpose of some of the vignettes is unclear. For example, "Good-enough practices for language model packages": who is this for, exactly? Why does it need to be a vignette? Why do "Data caching and updating" or "Why local LMs?" need their own vignettes, instead of being included as sections in a larger general guide? This information is surely useful, but it could be merged into a single vignette and/or the `README`, maybe with a dedicated vignette for developers, to better signpost the user to the information they need (or don't need).

I was able to install the package (v0.4.3.018) first try on WSL2 (Ubuntu on Windows 11) and didn't face any issues with building vignettes or `R CMD CHECK`. All tests passed successfully.

Not really sure where to start, I decided to run the example in the `README`:
```r
pkgmatch_similar_pkgs(input)
```

This threw an error:
```
Error in names(out) <- nms :
  'names' attribute [1] must be the same length as the vector [0]
```

I then remembered that I hadn't started the `ollama` server, so I go to the command line and run `ollama serve`, to get
```
Error: listen tcp 127.0.0.1:11434: bind: address already in use
```
This might be quite esoteric but could be worth mentioning in the vignette. A bit of detective work suggests that the error message above is in fact coming from `ollama_check()`, which when run throws exactly the same error. This left me a bit stuck, due to a couple of issues in the error handling, and due to [a bug in the `embeddings.R` file](https://github.com/ropensci-review-tools/pkgmatch/issues/106). I made a local patch to fix it on my end and was able to proceed from here.

Once I got the package working, the benefits of it seem clear! I can see this being immensely useful to package developers and advanced users who are looking for existing functionality, either for specialist tasks or trying not to reinvent the wheel while developing their own packages. However, as described above, the documentation comes across as being aimed at people who are already familiar with the package, rather than to new users who are trying to get it to work for the first time.

Another interface issue I note is that with most function calls, there is a slight delay and we see that data is being downloaded in the background:
```r
pkgmatch_similar_pkgs (".", corpus = "cran")
##  [100%] Downloaded 231213455 bytes...
##  [100%] Downloaded 36834164 bytes...
##  [100%] Downloaded 3948804 bytes...
##  [100%] Downloaded 13016426 bytes...
```
but what data is being downloaded, why and where to? Is it the corpus? Is it the embeddings? Maybe we should ask the user if they'd like to use up their bandwidth and disk space on this, or reassure them that it's going to a temporary directory and/or can easily be cleaned out at the end of the session? Which functionality will work offline, if any? This could be made a bit clearer on the first use.

The `browse = TRUE` function threw a further error for me, possibly related to using VS Code.
```r
pkgmatch_similar_fns (input, browse = TRUE)
# VSCode WebView only supports showing local http content.
# Opening in external browser...
# Browsing https://docs.ropensci.org/charlatan/reference/SequenceProvider
# Error in vapply(urls, utils::browseURL, integer(1L)) :
#   values must be length 1,
#  but FUN(X[[1]]) result is length 0
```
However, the help page successfully opened in the browser, so I'm not sure what the issue is here. I was able to run other examples successfully locally, though nearly all of them are 'Don't run' because of the external dependencies; I wonder if there is the possibility to embed some toy embeddings and `system.file`s in the package?

#### Other comments

In the examples and `README`, you have spaces between function names and the brackets, e.g. `fun ()` instead of `fun()`, which is not recommended in [most style guides](https://devguide.ropensci.org/pkg_building.html#code-style).

Overall I think this is a potentially great package with real potential, but the main thing holding it back is selling to the user what it's for and giving an easy way to get started. There are also one or two bugs or edge cases that thwarted me personally, but might not come up for other users, and so this might have coloured my impressions of the package. To do this, I'd suggest filling in the gaps and maybe focus on having a vignette with an extended use case that shows off the package's capabilities.

In the long run, I'd suggest clearly marking out which functionality is dependent on which external tools, as then a new user might be able to explore some features even if they have issues with others or do not want to set them up yet.
