### 2025-02-13 8fdd628
use ollama url from option to fix #106
