### 2025-02-13 8007403
need to pass explicit ollama.url envvar in tests for #106
