### 2025-02-13 3759f9d
use local test envvar for ollama.url #106
