### 2025-02-13 72e4e48
return fixed ollama_url in all test envs for #106
