### 2025-02-13 0d6cf59
update one further test condition on gh workflows
