### 2025-02-14 #671 @mpadge
Thank you very much @Selbosh for your review. You've really immediately picked up the big weakness of package design thus far: It was pulled together for a very specific use case, with documentation mostly hobbled together around that narrow vision, rather than me having thought nearly as much as I should have about general usage. In short: I agree with all of your concerns, and will address them hopefully within the coming fortnight. You should see several new issues in the repo linked immediately below here.
