### 2025-02-14 ropensci-review-tools/pkgmatch#110 @mpadge
Improve function-level docs
From @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> There are one or two functions documented in terms of what they are not rather than what they are, which can be a bit confusing (e.g. pkgmatch_bm25_fn_calls()). Some function descriptions (e.g. pkgmatch_bm25()) simply comprise an external Web link, which is not ideal for offline use, nor for users who want to know what a function does without having to click around. For the same functions, the current Titles in the documentation are too long or specific and would be better off as Descriptions, and the current Descriptions as Details or Notes.
