### 2025-02-14 ropensci-review-tools/pkgmatch#111 @mpadge
Restrcuture vignettes
From @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> The purpose of some of the vignettes is unclear. For example, "Good-enough practices for language model packages": who is this for, exactly? Why does it need to be a vignette? Why do "Data caching and updating" or "Why local LMs?" need their own vignettes, instead of being included as sections in a larger general guide? This information is surely useful, but it could be merged into a single vignette and/or the README, maybe with a dedicated vignette for developers, to better signpost the user to the information they need (or don't need).
