### 2025-02-14 ropensci-review-tools/pkgsimil#112 @mpadge
Improve error messages
From @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> Not really sure where to start, I decided to run the example in the README:
>
> ``` r
> pkgmatch_similar_pkgs(input)
> ```
>
> This threw an error:
> ``` r
> Error in names(out) <- nms :
>   'names' attribute [1] must be the same length as the vector [0]
> ```

Those kinds of errors should be intercepted earlier, with more instructive error text.
