### 2025-02-14 ropensci-review-tools/pkgmatch#113 @mpadge
Add note about ollama "address already in use"
This also from @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> I then remembered that I hadn't started the ollama server, so I go to the command line and run ollama serve, to get
>
> ```
> Error: listen tcp 127.0.0.1:11434: bind: address already in use
> ```

That error is common, and discussed in many ollama issues, yet can generally be ignored, and things work regardless. That should be explained.
