### 2025-02-14 ropensci-review-tools/pkgmatch#114 @mpadge
Improve download information
From @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> Another interface issue I note is that with most function calls, there is a slight delay and we see that data is being downloaded in the background:
>
> ``` r
> pkgmatch_similar_pkgs (".", corpus = "cran")
> ##  [100%] Downloaded 231213455 bytes...
> ##  [100%] Downloaded 36834164 bytes...
> ##  [100%] Downloaded 3948804 bytes...
> ##  [100%] Downloaded 13016426 bytes...
> ```
>
> but what data is being downloaded, why and where to? Is it the corpus? Is it the embeddings? Maybe we should ask the user if they'd like to use up their bandwidth and disk space on this, or reassure them that it's going to a temporary directory and/or can easily be cleaned out at the end of the session? Which functionality will work offline, if any? This could be made a bit clearer on the first use.
