### 2025-02-14 ropensci-review-tools/pkgmatch#115 @mpadge
Allow some examples to actually run
From @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> I was able to run other examples successfully locally, though nearly all of them are 'Don't run' because of the external dependencies; I wonder if there is the possibility to embed some toy embeddings and system.files in the package?

Fake data are already used in the mock test suites, so this should indeed be possible.
