### 2025-02-14 ropensci-review-tools/pkgsimil#116 @mpadge
Clarify which fns depend on external tools (ollama)
From @selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> In the long run, I'd suggest clearly marking out which functionality is dependent on which external tools, as then a new user might be able to explore some features even if they have issues with others or do not want to set them up yet.
