### 2025-02-24 #671 @agricolamz
---

I was avoiding reading the review by @Selbosh, and now I see that some of my comments are the same.

## Major comments

- I am not happy with the side effects of the package. Here is my first run of the `pkgmatch_similar_pkgs()` function:

```
> library(pkgmatch)
> input <- "
+ Packages for analysing evolutionary trees, with a particular focus
+ on visualising inter-relationships among distinct trees.
+ "
> pkgmatch_similar_pkgs(input)
 [100%] Downloaded 3497647 bytes...
 [100%] Downloaded 1076837 bytes...
Would you like to download it now (y/n) ? n
Would you like to download it now (y/n) ? n
Error in `httr2::req_perform()`:
! HTTP 404 Not Found.
Run `rlang::last_trace()` to see where the error occurred.
1: ollama model [jina/jina-embeddings-v2-base-en] is not installed.
2: ollama model [ordis/jina-embeddings-v2-base-code] is not installed.
```

First, it seems that the package quietly downloads something, and I don’t know what it is or where it is stored (now I know that it is in `/home/agricolamz/.cache/R`). Next, the function asked me whether “I would like to download it now?”. But I see that the function already downloaded something... Why did the package suddenly become so kind on the third and the fourth download?

After I installed the `ollama` models `jina/jina-embeddings-v2-base-en` (274 MB) and `ordis/jina-embeddings-v2-base-code` (323 MB) and played around, I decided to run the function using the `corpus = 'cran'` argument and got even bigger downloads:

```
[100%] Downloaded 232067039 bytes...
[100%] Downloaded 37056488 bytes...
```

I don't think that 273 Mb will cause any damage, but I expect that most users will disagree that silence implies consent.

- Another problem that I have a feeling that package does not work (the way I expect).

First I decided to check some popular packages

$ pkgmatch_similar_pkgs(corpus = "cran",
                      "Package that works with dates and times in the tidy format")

[1] "iso8601"         "rebus.datetimes" "datetime"        "tibbletime"
[5] "datefixR"
```

`lubridate` is number 15. I decided to check, whether the result is LLM motivated, so I ran 6 possible proportions of `lm_proportion`:

As you can see from the plots, my expectations were not met. That was striking for me, since I based my prompts on the documentation phrasing. It is possible that my function call meets this 8,096 token limit, so the model took the latest part of the package, omitting the very first entries. But this is just a hypothesis. I must admit: the suggested packages sometimes are really close to the prompt query. If you want to fix this, I would suggest merging information from LLM and BM25 with information about the number of downloads or any other measure of the package's popularity.

I also test the `pkgmatch_similar_pkgs()` function with the lingtypology tar.gz and the result was suprising for me:

As you see, the package correctly identifies the code, but for some reason, it struggles with the text.

- I disagree with your strategy of update of the pre-generated embeddings (see "Data caching and updating" vignette). I think that all updates should be done within the package itself, otherwise it is a high risk to get non-reproducible (even non-repeatable) result.

## Smaller comments/suggestions

- As you see, I have got an error on the first run. So I would propose to add a possibility to kindly refuse downloading something (without throwing an error).
- It looks like you are using two specific ollama models in the package; however, if I have an ollama on my computer, it is highly probable that I already downloaded some models (e.g. llama), and the embeddings from them could be useful. So I’d propose to make it possible to use `ollamar::embeddings()` and call some models that already present on the computer.
- With all due respect to the ROpenSci community, I think that corpus = "cran" should be the default option.
- The `browse = TRUE` argument seems to be not working in RStudio. It opens the link for the first package and throws an error:

```
Error in vapply(urls, utils::browseURL, integer(1L)) :
  values must be length 1,
 but FUN(X[[1]]) result is length 0
```

- It is not a problem at all, rather a question. I see through all documentation a space between the function name and the bracket. Is it a new code style that I am not aware of?

- Another side effect that I found is that the output of the `pkgmatch_similar_pkgs()` function is a dataframe and not a vector as I expected from the function's behavior. As I realized, there is a special print method (it is documented in the documentation). I would propose to have an additional argument that makes the function return a vector or the whole sorted dataframe. I am suggesting this because even the datafarame output is of a class pkgmatch that for some reason makes it impossible to merge multiple instances of the function output (e.g. gathered with the different lm_proportion value).

- I find the "Good-enough practices for language model packages" article from the documentation unnecessary. The contents of the "Why local language models (LMs)?" may can be devided between "README.md" and "Get started" page.

- Why is there no `lm_proportion` argument in the `pkgmatch_similar_fns()` function?

- You forgot the ending bracket in `fs::path(rappdirs::user_cache_dir(), "R", "pkgmatch"` in the "Data caching and updating" vignette.
