### 2025-03-18 ropensci-review-tools/pkgmatch#118 @mpadge
Vignette with extended use case
From @Selbosh's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2654565738):

> Overall I think this is a good idea for a package with real potential, but the main thing holding it back is selling to the user what it's for and giving an easy way to get started. To do this, I'd suggest filling in the gaps and maybe focus on having a vignette with an extended use case that shows off the package's capabilities. There are also one or two bugs or edge cases that thwarted me personally, but might not come up for other users, and so this might have coloured my impressions of the package.
