### 2025-03-18 ropensci-review-tools/pkgmatch#119 @mpadge
Improve strategy for updating local data
From @agricolamz's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2677331428):

> I disagree with your strategy of update of the pre-generated embeddings (see "Data caching and updating" vignette). I think that all updates should be done within the package itself, otherwise it is a high risk to get non-reproducible (even non-repeatable) result.

That's a really important point, and the strategy definitely should be changed. @agricolamz: I think it's still useful or important to make updating as easy as possible, at least to enable new packages to be potentially included in output. What would you think of this strategy:

- Implement an option on package load, `pkgmatch.update.interval`, with a default equal to an environment variable if set, so people can override in their own `.Renviron` settings, but otherwise has default of 30.
- When downloads are triggered in an interactive session, begin with information about update interval, and inform people how they can change it.
- Document that whole updating strategy somehwere (like a separate vignette).

Package would then largely function as it does now, but would be more verbose about it, and would enable users to switch off automatic downloading in a fairly standard and straightforward way.
