### 2025-03-18 ropensci-review-tools/pkgmatch#120 @mpadge
Make `corpus = "cran"` the default option?
From @agricolamz's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2677331428):

> With all due respect to the ROpenSci community, I think that corpus = "cran" should be the default option.

That's a reasonable point. The reason "ropensci" was chosen is actually because it will deliver results faster, especially for the (presumed majority of) people who will be running ollama on CPU. For example, I get results back from the "ropensci" corpus in 0.1-0.2 seconds, whereas "cran" takes 5-10 seconds, or generally around 50 times slower.

My thoughts were, and still are, that having a default behaviour that may lead to many users waiting 10 seconds or more for a function to return would be a strong disincentive to further usage. @agricolamz: Given this consideration, what would you suggest here?
