### 2025-03-18 ropensci-review-tools/pkgmatch#121 @mpadge
Improve object returned from similar_pkgs fn
From @agricolamz's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2677331428):

> Another side effect that I found is that the output of the pkgmatch_similar_pkgs() function is a dataframe and not a vector as I expected from the function's behavior. As I realized, there is a special print method (it is documented in the documentation). I would propose to have an additional argument that makes the function return a vector or the whole sorted dataframe. I am suggesting this because even the datafarame output is of a class pkgmatch that for some reason makes it impossible to merge multiple instances of the function output (e.g. gathered with the different lm_proportion value).
