### 2025-03-18 ropensci-review-tools/pkgmatch#122 @mpadge
lm_proportion arg in similar_fns
From @agricolamz's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2677331428):

> Why is there no lm_proportion argument in the pkgmatch_similar_fns() function?
