### 2025-03-18 89e5006
update similar_fns docs to close #122
