+++ R/similar-fns.R
@@ -1,7 +1,9 @@
 #' Identify R functions best matching a given input string
 #'
-#' @description Function matching is only available for
-#' functions from the corpus of rOpenSci packages.
+#' @description Function matching is only available for functions from the
+#' corpus of rOpenSci packages. Function matching is also based on LM output
+#' only, and unlike package matching does not combine LM output with BM25
+#' word-frequency matching.
 #'
 #' @inheritParams pkgmatch_similar_pkgs
 #' @param input A text string.
