### 2025-03-18 c63f599
add noRd to get_verbose_limit fn
