### 2025-03-18 ropensci-review-tools/pkgmatch#125 @mpadge
Output of LM models different to expectations
From @agricolamz's [review](https://github.com/ropensci/software-review/issues/671#issuecomment-2677331428), calling `similar_pkgs()` with following inputs often fails to produce the expected results:

- "Package that works with dates and times in tidy format"; expected answer: `lubridate`
- "Package for data manipulation in the tidy format", expected answer: `dplyr`
- "Package for fitting Bayesian generalized multivariate (non-)linear multilevel models using Stan", expected answer: `brms`, `rstan`
- "Package for creating interactive maps using Leaflet library", expected answer: 'leaflet'
- "Package for creating interactive tables with DataTables Javascript library", expected answer: `DT`
- "Package for generation model-based estimates and visualization them in ggplot", expected answer: `ggeffects`
