### 2025-03-18 ropensci-review-tools/pkgmatch#122 @mpadge
That's because that function uses LM output only. The corpus is way too big to calculate reference BM25 values. Function docs have now been updated to explain this:

https://github.com/ropensci-review-tools/pkgmatch/blob/3a0918c465efb9f00705a4f5f8450a7abc319278/R/similar-fns.R#L3-L6

(Although note that that then devotes more time to explaining what the function does *not* do than on what it actually does; and so should likely be updated as part of #110)
