### 2025-03-18 ropensci-review-tools/pkgmatch#120 @agricolamz
I'd suggest not using the default option at all, and make users decide. If they didn't fill an argument, then you can put a message asking them to choose or cancel.
