### 2025-03-18 fb9a429
improve 'ollama_is_running' fn for #114
