### 2025-03-18 c09355e
call 'ollama_check()' on pkg load
