### 2025-03-18 767b50e
set update frequency by envvar if present for #114
