### 2025-03-19 2ecbaa8
use cli::keypress for corpus input for #120
