### 2025-03-19 1c0d0e2
set more corpus params to default NULL for #114
And add 'check_corpus_param' fn
