### 2025-03-19 3b29e5e
add a lot of download verbosity messaging for #114
