### 2025-03-20 8f772d1
fix test for #114 by adding explicit 'corpus' arg
