### 2025-03-20 27f886f
modify 'check_corpus' for #114 to fail if no keypress support
gha windows machines do not support keypress, so fail without that.
