### 2025-03-20 edc9cdf
use packageStartupMessage not cli in .onLoad
