+++ R/zzz.R
@@ -24,7 +24,7 @@
         error = function (e) e
     )
     if (inherits (chk, "error")) {
-        cli::cli_alert_warning (chk$message)
+        packageStartupMessage (chk$message)
     }
 
 
