### 2025-03-21 439f2f4
begin readme rewrite for #109
