### 2025-03-21 #671 @mpadge
@MargaretSiple-NOAA @Selbosh @agricolamz I've started updating the package in response to the really helpful reviews. You can see several issues listed above. I hope to finish them within the next couple of weeks, and will then write a full response here. Thank you all 👍 😄
