### 2025-03-21 b740375
re-order vignettes in pkgdown menu for #111
