### 2025-03-21 c5ff26d
add note about ollama download sizes in vignette for #114
