+++ R/bm25.R
@@ -43,6 +43,14 @@ pkgmatch_bm25 <- function (input, txt = NULL,
 
 pkgmatch_bm25_internal <- function (input, txt, idfs, corpus) {
 
+    fname <- get_cache_file_name (
+        what = "idfs",
+        corpus = corpus,
+        fns = FALSE,
+        raw = FALSE
+    )
+    send_dl_message (fname)
+
     if (is.null (txt)) {
         if (is.null (idfs)) {
             idfs <- pkgmatch_load_data ("idfs", corpus = corpus, fns = FALSE)
@@ -136,6 +144,12 @@ pkgmatch_bm25_fn_calls <- function (path, corpus = NULL) {
 }
 
 pkgmatch_bm25_fn_calls_internal <- function (path, corpus) { # nolint
+
+    fnames <- c (
+        get_cache_file_name (what = "calls", corpus, fns = FALSE, raw = FALSE),
+        get_cache_file_name (what = "calls", corpus, fns = FALSE, raw = TRUE)
+    )
+    send_dl_message (fnames)
 
     tokens_idf <-
         pkgmatch_load_data (what = "calls", corpus = corpus, raw = FALSE)
