+++ R/bm25.R
@@ -145,11 +145,14 @@ pkgmatch_bm25_fn_calls <- function (path, corpus = NULL) {
 
 pkgmatch_bm25_fn_calls_internal <- function (path, corpus) { # nolint
 
-    fnames <- c (
-        get_cache_file_name (what = "calls", corpus, fns = FALSE, raw = FALSE),
-        get_cache_file_name (what = "calls", corpus, fns = FALSE, raw = TRUE)
-    )
-    send_dl_message (fnames)
+    calling_fn <- as.character (sys.call (sys.parent (n = 2L)) [[1]])
+    if (calling_fn != "pkgmatch_similar_pkgs") {
+        fnames <- c (
+            get_cache_file_name (what = "calls", corpus, fns = FALSE, raw = FALSE),
+            get_cache_file_name (what = "calls", corpus, fns = FALSE, raw = TRUE)
+        )
+        send_dl_message (fnames)
+    }
 
     tokens_idf <-
         pkgmatch_load_data (what = "calls", corpus = corpus, raw = FALSE)
