### 2025-03-25 cf00551
packageStartupMessage in .onAttach not .onLoad
