### 2025-03-25 05df840
improve send_dl_message in test environments for #114
