### 2025-03-25 89aae5b
fake 'list_remote_files' in test envs
To get file sizes for dl messages in #114
