### 2025-03-25 bf93d8e
test mocked result of 'list_remote_files' for #114
