### 2025-03-25 dbbcf20
fix test of 'list_remote_files' for #114
