+++ tests/testthat/test-data-update.R
@@ -262,9 +262,12 @@ test_that ("list remote files", {
 
     expect_s3_class (flist, "data.frame")
     expect_equal (ncol (flist), 6L)
-    expect_named (flist, c ("file_name", "size", "timestamp", "tag", "owner", "repo"))
+    expect_named (
+        flist,
+        c ("file_name", "size", "timestamp", "tag", "owner", "repo")
+    )
     expect_true (all (flist$owner == "ropensci-review-tools"))
     expect_true (all (flist$repo == "pkgmatch"))
-    expect_true (all (flist$timestamp == "2025-01-01"))
+    expect_true (all (flist$timestamp == as.POSIXct ("2025-01-01T00:00:00")))
     expect_true (all (grepl ("\\.Rds", flist$file_name)))
 })
