### 2025-03-25 ropensci-review-tools/pkgmatch#114 @mpadge
Done via the above commits. Initial message if no cached data exist is:
https://github.com/ropensci-review-tools/pkgmatch/blob/3df185f0b65a91a7f2dcf9e96032ae4b7128b19c/R/cache.R#L247-L251

Each function which requires data now starts by assembling list of data files needed. Those are then passed to the new `send_dl_message()` function:
https://github.com/ropensci-review-tools/pkgmatch/blob/3df185f0b65a91a7f2dcf9e96032ae4b7128b19c/R/cache.R#L238

That grabs data from the [remote file list](https://github.com/ropensci-review-tools/pkgmatch/releases/tag/v0.4.0), and displays on screen the number of files needed and total download size. Each download then issues a message about what is being downloaded. Thanks @Selbosh for the inpetus here!
