### 2025-03-25 280ac10
extend data updating vignette for #119
