### 2025-03-25 ropensci-review-tools/pkgmatch#119 @mpadge
All done and documented in https://docs.ropensci.org/pkgmatch/articles/C_data-caching-and-updating.html#local-caching-and-updating-for-users
