### 2025-03-25 892c94f
update similar-pkgs docs for #121
