### 2025-03-25 63c6483
update bm25 fn docs for #110
