### 2025-03-25 414c10e
update bm25_fn_calls docs for #110
