+++ man/pkgmatch_bm25.Rd
@@ -30,13 +30,13 @@ A \code{data.frame} of package names and 'BM25' measures against text
 from whole packages both with and without function descriptions.
 }
 \description{
-BM25 values match single inputs to document corpora by weighting terms by
-their inverse frequencies, so that relatively rare words contribute more to
-match scores than common words. For each input, the BM25 value is the sum of
-relative frequencies of each term in the input multiplied by the Inverse
-Document Frequency (IDF) of that term in the entire corpus. See the
-Wikipedia page at \url{https://en.wikipedia.org/wiki/Okapi_BM25} for further
-details.
+BM25 values match single inputs to document corpora by
+weighting terms by their inverse frequencies, so that relatively rare words
+contribute more to match scores than common words. For each input, the BM25
+value is the sum of relative frequencies of each term in the input
+multiplied by the Inverse Document Frequency (IDF) of that term in the
+entire corpus. See the Wikipedia page at
+\url{https://en.wikipedia.org/wiki/Okapi_BM25} for further details.
 }
 \examples{
 \dontrun{
