### 2025-03-26 7e0145f
update fn docs for load_data and update_cache for #110
