+++ man/pkgmatch_load_data.Rd
@@ -2,9 +2,7 @@
 % Please edit documentation in R/cache.R
 \name{pkgmatch_load_data}
 \alias{pkgmatch_load_data}
-\title{Load embeddings generated by the \link{pkgmatch_embeddings_from_pkgs}
-function, either for all rOpenSci packages or, if \code{fns = TRUE},  all
-individual functions within those packages.}
+\title{Load 'pkgmatch' data for specified corpus.}
 \usage{
 pkgmatch_load_data(
   what = "embeddings",
@@ -14,15 +12,8 @@ pkgmatch_load_data(
 )
 }
 \arguments{
-\item{what}{One of:
-\itemize{
-\item "embeddings" to load pre-generated embeddings;
-\item "idfs" to load pre-generated Inverse Document Frequency weightings;
-\item "functions" to load pre-generated frequency tables for text
-descriptions of function calls; or
-\item "calls" to load pre-generated frequency tables for actual function
-calls.
-}}
+\item{what}{One of the four data types described above: "embeddings",
+"idfs", "functions", or "calls".}
 
 \item{corpus}{Must be specified as one of "ropensci" or "cran". If
 \code{embeddings} or \code{idfs} parameters are not specified, they will be
@@ -31,9 +22,9 @@ function will then return the most similar package from the specified
 corpus. Note that calculations will \code{corpus = "cran"} will generally take
 longer, because the corpus is much larger.}
 
-\item{fns}{If \code{FALSE} (default), load embeddings for all rOpenSci packages;
-otherwise load (considerably larger dataset of) embeddings for all
-individual functions.}
+\item{fns}{If \code{FALSE} (default), load embeddings for all packages; otherwise
+load (considerably larger dataset of) embeddings for all individual
+functions.}
 
 \item{raw}{Only has effect of \code{what = "calls"}, in which case default of
 \code{FALSE} loads single Inverse Document Frequency table to entire corpus;
@@ -41,12 +32,22 @@ otherwise if \code{TRUE}, loads raw function call counts for each package in
 corpus.}
 }
 \value{
-The loaded \code{data.frame}.
+The loaded data.
 }
 \description{
-Load embeddings generated by the \link{pkgmatch_embeddings_from_pkgs}
-function, either for all rOpenSci packages or, if \code{fns = TRUE},  all
-individual functions within those packages.
+Load pre-computed data for a specified corpus. Data types are:
+\itemize{
+\item "embeddings" for language model embeddings;
+\item "idfs" for Inverse Document Frequency weightings;
+\item "functions" for frequency tables for text descriptions of function
+calls; or
+\item "calls" for frequency tables for actual function calls.
+}
+
+This function is called within the main
+\link{pkgmatch_similar_pkgs} and \link{pkgmatch_similar_fns} functions
+to load required data there, and should not generally need to be explicitly
+called.
 }
 \examples{
 \dontrun{
