### 2025-03-26 bdfe3d4
update fn docs for 'update_data' for #110
