### 2025-03-26 c6ed458
update embeddings fn docs for #110
