### 2025-03-26 7fce8bb
update ollama fn docs for #110
