+++ R/ollama.R
@@ -144,7 +144,19 @@ ollama_is_running <- function () {
     return (chk)
 }
 
-#' Check that ollama is installed with required models, and download if not.
+#' @title Check ollama installation
+#'
+#' @description Performs the following checks:
+#' \itemize{
+#' \item Check that ollama is installed
+#' \item Check that ollama is running
+#' \item Check that ollama has the required models, and download if not
+#' }
+#'
+#' The required models are the Jina AI embeddings:
+#' \url{https://ollama.com/jina/jina-embeddings-v2-base-en} for text
+#' embeddings, and \url{https://ollama.com/ordis/jina-embeddings-v2-base-code}
+#' for code embeddings.
 #'
 #' Note that the URL of a locally-running ollama instance is presumed by
 #' default to be "127.0.0.1:11434". Other values can be set using the
@@ -153,7 +165,7 @@ ollama_is_running <- function () {
 #' @param sudo Set to `TRUE` if ollama is running in docker with sudo
 #' privileges.
 #' @return TRUE if everything works okay, otherwise the function will error
-#' before returning.
+#' before returning, and issue an informative error message.
 #'
 #' @examples
 #' \dontrun{
