### 2025-03-26 d259547
update print method fn docs for #110
