### 2025-03-26 2dba00f
update 'similar-fns' fn docs for #110
