### 2025-03-26 afbab0a
update tree-sitter fn docs for #110
