### 2025-03-26 ropensci-review-tools/pkgmatch#110 @mpadge
Docs for every single function updated via #134. Thanks  @Selbosh for noting that.
