### 2025-03-26 a4f5a71
add inst/extdata/embeddings-pkg for #115
To allow examples to run using pre-generated embeddings
