### 2025-03-26 76119af
finish example data generation for embeddings, bm25 for ##115
