### 2025-03-26 745bad2
add idfs_fn_calls to example data generation for #115
