### 2025-03-26 1d6c58f
add 'ex_fn_calls' to generate example fn call data for #115
