### 2025-03-26 e625935
fix example data for #115
