### 2025-03-26 712181e
get fn_calls fake example data from one 1 pkg not all
For #115, where previous version applied treesitter to all installed pkgs,
which took way too long, and results can easily be faked by only
generating from 1 pkg and copying across
