+++ R/example-data.R
@@ -136,25 +136,21 @@ ex_idfs_fn_calls <- function (pkg_nms, fname) {
 
 ex_fn_calls <- function (pkg_nms, fname) {
     ip <- data.frame (installed.packages ())
-    n_pkgs <- min (nrow (ip), length (pkg_nms))
-    index <- sample (nrow (ip), replace = FALSE, size = n_pkgs)
-    ip <- ip [index, ]
 
-    fn_calls <- lapply (ip$Package, function (p) {
+    tags <- NULL
+    while (is.null (tags)) {
+        i <- sample (nrow (ip), size = 1L)
+        pkg <- ip$Package [i]
         tags <- tryCatch (
-            pkgmatch_treesitter_fn_tags (p),
+            pkgmatch_treesitter_fn_tags (pkg),
             error = function (e) NULL
         )
-        res <- NULL
-        if (!is.null (tags)) {
-            fns <- tags$name [which (!grepl (paste0 ("^", p), tags$name))]
-            res <- table (fns)
-        }
-        return (res)
-    })
-    index <- which (!vapply (fn_calls, is.null, logical (1L)))
-    fn_calls <- fn_calls [index]
-    names (fn_calls) <- pkg_nms [index]
+    }
+    fns <- tags$name [which (!grepl (paste0 ("^", pkg), tags$name))]
+    fn_tbl <- table (fns)
+
+    fn_calls <- lapply (pkg_nms, function (i) fn_tbl)
+    names (fn_calls) <- pkg_nms
 
     saveRDS (fn_calls, fname)
     return (fname)
