### 2025-03-26 8d55252
add family tag and missing Rd file for example data fn ##115
