+++ man/print.pkgmatch.Rd
@@ -32,6 +32,7 @@ head (p) # Shows first 5 rows of full `data.frame` object
 }
 \seealso{
 Other utils: 
+\code{\link{generate_pkgmatch_example_data}()},
 \code{\link{head.pkgmatch}()},
 \code{\link{pkgmatch_browse}()},
 \code{\link{pkgmatch_load_data}()},
