### 2025-03-26 c514eeb
need to control example data cache by option not envvar #115
