### 2025-03-26 9edf773
replace cli_alert_info with cli_inform
