### 2025-03-26 9132ce7
need to hard-code 'expected_embedding_length' for ex data #115
