### 2025-03-26 6cde86b
finally switch on example for similar_pkgs #115
