### 2025-03-26 e6655a4
include pkg version info in example data #115
