### 2025-03-26 a2fabca
switch on full example for similar pkgs #115
