### 2025-03-26 8576ba2
extdata as Rds does not work; convert to json for #115
