### 2025-03-26 dffafd9
suppress pkg startup messages in treesitter calls for #115
