### 2025-03-26 39d4a59
require jsonlite namespace quietly for reading ex data #115
