### 2025-03-26 485d84a
restructure example with test data #115
