### 2025-03-26 890ee89
add additional ex data for fn embeddings #115
And activate another example
