### 2025-03-26 f450c41
use internal ex data to run bm25 example #115
