+++ man/pkgmatch_bm25.Rd
@@ -39,13 +39,14 @@ entire corpus. See the Wikipedia page at
 \url{https://en.wikipedia.org/wiki/Okapi_BM25} for further details.
 }
 \examples{
-\dontrun{
-input <- "Download open spatial data from NASA"
-bm25 <- pkgmatch_bm25 (input)
-# Or pre-load document-frequency weightings:
-idfs <- pkgmatch_load_data ("idfs", fns = FALSE)
-bm25 <- pkgmatch_bm25 (input, idfs = idfs)
-}
+# The following function simulates remote data in temporary directory, to
+# enable package usage without downloading. Do not run for normal usage.
+generate_pkgmatch_example_data ()
+
+pkgmatch_bm25 (input = "curl", corpus = "cran")
+# Or pre-load document-frequency weightings and pass those:
+idfs <- pkgmatch_load_data ("idfs", corpus = "cran", fns = FALSE)
+pkgmatch_bm25 (input, corpus = "cran", idfs = idfs)
 }
 \seealso{
 Other bm25: 
