### 2025-03-26 ba60eee
add 'test-example-data.R' for #115
