### 2025-03-26 39820dc
fix example for bm25 #115
