### 2025-03-26 ropensci-review-tools/pkgmatch#115 @mpadge
Thanks @Selbosh, PR #135 has added two internal data files, and a host of routines to simulate some fake data in `tempdir()`. That then allows the main examples to actually be run. The only one that still can't be run is `pkgmatch_similar_fns()`, but I think that's fine, and would require additional fake data.
