### 2025-03-27 13dd81c
fix idf extraction in bm25
similar_pkgs_from_text joined similarities to bm25, but previous version
removed vers + .tar.gz from package names, so join failed, leaving all
bm25 values as NA -> 0
