### 2025-03-27 c2a19ba
improve grep for 'input_mentions_fns' for #125
