### 2025-03-27 ropensci-review-tools/pkgmatch#125 @mpadge
## Preliminary note

This is a very important part of what I was hoping to come out of the review process, that somebody would dig deeply into the actual results that the package produces. My disclaimer at the outset is that a part of my motivation for developing the package is to make the processes of LM recommendation systems more transparent. And those process are very generally and very often messy, noisy, and unexpected. A heap of the effort of organizations developing recommendation engines goes into post-processing the initial results. `pkgmatch` implements very little of that, and so exposes users to the messy world within these systems. In that respect, my hope is as much that the package will be educational, as that it will be actually directly useful.

## Initial explorations

Final ranks are based on combinations of input factors, each of which is internally ranked. The following function uses some internal code to return the ranks of the "target" package within each individual component.

``` r
library (pkgmatch)
pm_ranks <- function (input, target) {
    corpus <- "cran"
    embeddings <- pkgmatch_load_data (what = "embeddings", corpus = corpus)
    idfs <- pkgmatch_load_data (what = "idfs", corpus = corpus)
    index <- which (!duplicated (names (idfs$token_lists$with_fns)))
    idfs$token_lists$with_fns <- idfs$token_lists$with_fns [index]
    index <- which (!duplicated (names (idfs$token_lists$wo_fns)))
    idfs$token_lists$wo_fns <- idfs$token_lists$wo_fns [index]
    input_is_code <- text_is_code (input)
    res <- pkgmatch:::similar_pkgs_from_text (
        input = input,
        embeddings = embeddings,
        idfs = idfs,
        corpus = corpus,
        input_is_code = input_is_code
    )
    res <- pkgmatch:::make_cran_version_column (res)

    cols <- grep ("\\_fns$", names (res), value = TRUE)
    ranks <- vapply (cols, function (i) {
        decr <- !grepl ("^bm25", i)
        index <- order (res [[i]], decreasing = decr)
        which (res$package [index] == target)
    }, integer (1L))
    return (ranks)
}
```

### lubridate

``` r
target <- "lubridate"
input <- "Package that works with dates and times in tidy format"
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>             27             33            695           3425
# Actual title of lubridate package:
input <- "Make dealing with dates a little easier"
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>            189             78              2             31
```
The actual title matches really well with BM25, but pretty poorly with LM similarities.
``` r
# Actual description of lubridate package:
input <- paste0 (
    "Functions to work with date-times and time-spans: fast and ",
    "user friendly parsing of date-time data, extraction and updating of ",
    "components of a date-time (years, months, days, hours, minutes, and ",
    "seconds), algebraic manipulation on date-time and time-span objects. ",
    "The 'lubridate' package has a consistent and memorable syntax that ",
    "makes working with dates easy and fun."
)
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>             69              1           8026           3679
```
In contrast to the title, the full description matches terribly with BM25, but excellently with LM similarity. Why? There's no real way of knowing, other than that the input here is much longer, so provides more context for the LM.
``` r
# This will analyse full installed package, although not in same way as actual
# code, which includes additional processing when input is identified as a
# package.
input <- "lubridate"
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>           3169              2           3148           3159
input <- "lubridate package" # text string
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>           1271              7            243            371
```
And once again, the LM similarity measures are pretty good there, whereas the BM25 values are not so good at all.

Finally, the following function extracts all words from the full help page for the package, which is then used as input:
``` r
# Full package description from lubridate:
get_pkg_desc <- function (target) {
    rd <- tools::Rd_db (target) [[paste0 (target, "-package.Rd")]]
    rd <- gsub ("[[:punct:]]|\\n", "", as.character (rd))
    rd <- gsub ("^\\s+", "", rd)
    rd <- rd [which (nzchar (rd))]
    rm <- c ("code", "link", "section", "url", "item")
    input <- paste0 (unique (rd [which (!rd %in% rm)]), collapse = " ")
    nwords <- length (strsplit (input, "\\s*") [[1]])
    cli::cli_inform ("Number of words in {target} description = {nwords}")
    return (input)
}
input <- get_pkg_desc (target)
#> Number of words in lubridate description = 5753
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>           3935            169           3730           9006
```
And that performs far worse that either the title or the description.

### dplyr

Note at the outset that the textual descriptions for the `dplyr` package are comparatively very brief, and also very generic. There's really very little specific terminology to match on to this package, so results will likely be poor.

``` r
target <- "dplyr"
input <- "Package for data manipulation in the tidy format"
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>            292            148          22228          21982
# Actual title of dplyr package:
input <- "A grammar of data manipulation"
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>           1960            729          22051          22415
# Actual description of dplyr package:
input <- paste0 (
    "A fast, consistent tool for working with data frame like ",
    "objects, both in memory and out of memory."
)
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>           3601           2423          22398          22423
```

And all of those results are next to useless for identifying the target package, but that is expected for reasons identified at the outset. The following are notably better, but still not particularly good at actually identifying the desired target package.

``` r
input <- "dplyr" # full package
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>             54             11          22371          22420
input <- "dplyr package" # simple text
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>             62            121          22351          22401

# Full package description from dplyr:
input <- get_pkg_desc (target)
#> Number of words in dplyr description = 643
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>             88             72          22081          22405
```

But looking at the actual matches in the results shown by @agricolamz shows something like this:

``` r
input <- "Package for data manipulation in the tidy format"
pkgmatch_similar_pkgs (input, corpus = "cran", n = 10L)
#>  [1] "tidybins"       "tidyedgar"      "flatr"          "fastverse"
#>  [5] "tidygraph"      "tidybayes"      "tidydatatutor"  "tidyfst"
#>  [9] "TidyConsultant" "tidyverse"
# Actual title of dplyr package:
input <- "A grammar of data manipulation"
pkgmatch_similar_pkgs (input, corpus = "cran", n = 10L)
#>  [1] "table.express" "cheese"        "cleandata"     "smarter"
#>  [5] "xegaBNF"       "SSBtools"      "funprog"       "comperes"
#>  [9] "hdd"           "gdata"
# Actual description of dplyr package:
input <- paste0 (
    "A fast, consistent tool for working with data frame like ",
    "objects, both in memory and out of memory."
)
pkgmatch_similar_pkgs (input, corpus = "cran", n = 10L)
#>  [1] "objectremover"   "rtape"           "fastmatch"       "fastrerandomize"
#>  [5] "R.cache"         "fastdigest"      "tidytable"       "dplyr"
#>  [9] "cheapr"          "chillR"
input <- "dplyr" # full package
pkgmatch_similar_pkgs (input, corpus = "cran", n = 10L)
#> $text
#>  [1] "A3"            "fplyr"         "AalenJohansen" "tidytable"
#>  [5] "AATtools"      "stickyr"       "cubelyr"       "ABACUS"
#>  [9] "daewr"         "abasequence"
#>
#> $code
#>  [1] "dplyr"         "A3"            "RAthena"       "AalenJohansen"
#>  [5] "AATtools"      "noctua"        "duckdb"        "ABACUS"
#>  [9] "abclass"       "pool"
input <- "dplyr package" # simple text
pkgmatch_similar_pkgs (input, corpus = "cran", n = 10L)
#>  [1] "nplyr"           "tidier"          "duckplyr"        "dtplyr"
#>  [5] "reasonabletools" "MetricsWeighted" "tidyselect"      "tidyEdSurvey"
#>  [9] "dfidx"           "dibble"

# Full package description from dplyr:
input <- get_pkg_desc (target = "dplyr")
#> Number of words in dplyr description = 643
pkgmatch_similar_pkgs (input, corpus = "cran", n = 10L)
#>  [1] "splithalf"   "midfieldr"   "normfluodbf" "tidier"      "dplyr"
#>  [6] "duckplyr"    "nplyr"       "dverse"      "datasailr"   "tidyformula"
```

And `dplyr` at least appears in some of those top 10 results, while `lubridate` never does.

----

### Preliminary Conclusion

Aspects of the preceding matches for `lubridate` were quite good, while individual component ranks for `dplyr` were comparably worse. And yet these are the final ranks from direct output of `pkgmatch_similar_pkgs()`:

Package | Input | rank
--- | --- | ---
lubridate | pkg-title | 179
lubridate | pkg-description | 48
lubridate | full-description | 817
lubridate | pkg-code | 6
lubridate | "lubridate package" | 20
dplyr | pkg-title | 17
dplyr | pkg-description | 8
dplyr | full-description | 5
dplyr | pkg-code | 1311
dplyr | "dplyr package" | 22

This then just demonstrates how difficult the general task of recommendation is. The individual results which performed worse (dplyr) still generated superior final results, while results with individually accurate components (lubridate) failed to generate overall useful results. The general approach here, of naively combining as a variety of results from different aspects of the input using a single algorithm is also generally reflective of approaches used through the recommendation-engine literature as well as in practice. Tweaking combinations of individual results to fit some subset of expected outcomes is never done, for very good reasons (overfitting, among others).

Obviously the general issues here will need to be addressed. I think the best approach might be within some kind of additional vignette or the like, via #118

### **Effects** of word frequencies on results?

One interesting aspect nevertheless remains that only one match for `dplyr` is outside the top 20, and yet most matches for lubridate are well outside that (in spite of superior scores on some individual components). Why? The match for the full description is terrible, and the results above show that all individual components are equally terrible:

``` r
pm_ranks (input, target = target)
#> simil_with_fns   simil_wo_fns  bm25_with_fns    bm25_wo_fns
#>           3935            169           3730           9006
```
The only one that is half-way okay is the LM similarity measure without function descriptions. But the BM25 values provide useful insight there: They directly reflect how unique the complete package description is in terms of word frequencies across the entire corpus constructed from all CRAN packages. The lubridate package simply fails to be textually unique or distinctive. This is maybe because so many packages mention dates and manipulation and the kinds of verbs and nouns ubiquitous throughout that package? These are IDFs across the corpus for the main words in titles and descriptions of both lubridate and dplyr:

``` r
input <- c ("date", "data", "tidy", "manipulation", "format", "grammar")
idfs <- pkgmatch_load_data (what = "idfs", corpus = "cran")
idfs$idfs$wo_fns |>
    dplyr::filter (token %in% input) |>
    dplyr::arrange (dplyr::desc (idf))
#>          token       idf
#> 1      grammar 5.3064902
#> 2 manipulation 4.0328072
#> 3         tidy 3.9555674
#> 4         date 3.3142638
#> 5       format 2.5620403
#> 6         data 0.7006476
```

And that seems to indicate that `dplyr` just does a better job of using infrequent words in its text than lubridate. (And that prompts a final thought: This raises another use-case for this package, as an aid for package developers to boost the prominence of their packages in LM results...)
