### 2025-05-05 #671 @MargaretSiple-NOAA
Thank YOU @maurolepore for keeping us all on track!

@mpadge , FYI, I am going into be at sea for field work May 23-July 11 this year so if you are still working on it for a while longer, we may have a period of inactivity during which we could put it on hold or [some other option that I don't know yet].
