### 2025-05-06 #671 @agricolamz
As a field linguist myself, I wish you good luck!
