### 2025-05-06 #671 @mpadge
Thanks both. In that case I'll endeavour to get my responses back before you head off. I'll aim to do that quickly enough to finish this before May 23rd. If not, we'll maybe consider [some other option that I don't know yet] 😄 ... but hopefully not 🚀
