### 2025-05-20 f2f3018
add verbose output to update_cache for #114
