### 2025-05-20 981f357
add lots to new vignette for #125
