+++ vignettes/E_why-are-the-results-not-what-i-expect.Rmd
@@ -20,3 +20,191 @@ code, or even entire R packages. "pkgmatch" finds the most closely matching
 packages using a combination of Language Models (LMs, or equivalently, "LLMs"
 for large language models), and traditional [token-frequency
 algorithms](https://en.wikipedia.org/wiki/Okapi_BM25).
+
+Start by loading the package:
+
+```{r library}
+library (pkgmatch)
+```
+
+This vignette will consider the following single example prompt:
+
+```{r}
+input <- "Package that works with dates and times in tidy format"
+```
+
+The expected result from that prompt is [the `lubridate`
+package](https://lubridate.tidyverse.org/). By default, the object returned
+from [the `pkmgatch_similar_pkgs()`
+function()](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
+prints the top five matching packages:
+
+```{r sim-pkgs1, eval = FALSE}
+pkgmatch_similar_pkgs (input, corpus = "cran")
+```
+```{r sim-pkgs1-out, echo = FALSE}
+c ("datefixR", "fpp3", "rebus.datetimes", "sweep", "datetime")
+```
+
+And while expected results may not necessarily always be in the first position,
+the absence of [`lubridate`](https://lubridate.tidyverse.org/) from the top
+five matches is indeed unexpected. This vignette dives in to the details of why
+this might happen.
+
+### Obtain package source
+
+The following code downloads tarballs of source code for the five packages
+listed above, plus `lubridate`.
+
+```{r dl-pkgs}
+pkgs <- c (
+    "datefixR", "fpp3", "rebus.datetimes", "sweep", "datetime", "lubridate"
+)
+path <- utils::download.packages (pkgs, destdir = fs::path_temp ())
+chk <- lapply (path [, 2], function (p) {
+    utils::untar (p, exdir = fs::path_temp (), tar = "internal")
+})
+pkg_paths <- fs::path (fs::path_temp (), pkgs)
+stopifnot (all (fs::dir_exists (pkg_paths)))
+```
+
+### Generate embeddings from package source code
+
+Embeddings used in {pkgmatch} are generated from [the
+`pkgmatch_embeddings_from_pkgs()`
+function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_embeddings_from_pkgs.html).
+This function accepts one main input parameter specifying paths to one or more
+local directories containing source code.
+
+```{r embeddings, eval = FALSE}
+emb <- pkgmatch_embeddings_from_pkgs (pkg_paths)
+str (emb)
+```
+
+```{r embeddings-output, echo = FALSE}
+text_with_fns <- array (runif (768 * 6, -1, 1), dim = c (768, 6))
+text_wo_fns <- array (runif (768 * 6, -1, 1), dim = c (768, 6))
+code <- array (runif (768 * 6, -1, 1), dim = c (768, 6))
+colnames (text_with_fns) <- colnames (text_wo_fns) <- colnames (code) <- pkgs
+emb <- list (
+    text_with_fns = text_with_fns,
+    text_wo_fns = text_wo_fns,
+    code = code
+)
+str (emb)
+```
+
+
+### Modifying embeddings through chunking
+
+Results of `pkgmatch` functions can be improved through generating several
+embeddings for each component of each package. LLMs commonly achieve this
+through "chunking", which refers to taking different "chunks" of an input that
+generally extends beyond the admissible "context window" of the model. The
+models used here have context windows of 8096 tokens, which is often enough to
[…]
