### 2025-05-20 #671 @mpadge
@MargaretSiple-NOAA Update: I'm not quite going to get my responses done by May 23rd. If I may, please let me know what you think of this as a plan:

- I get my responses done and posted here by end of this month of May
- I'll ping both @Selbosh and @agricolamz for their thoughts, feedback, and judgement
- We back-and-forth during month of June while you're away
- Hopefully by mid-July when you return, everything will be ready to approve.

Let me know whether that sounds good, and David and George, hopefully you'll both have a small amount of time in June to have a look over my changes. Thanks all!
