### 2025-05-21 515aa73
rewrite vignette intro for #125
