### 2025-05-21 3d97046
extend first intro sec of #125 vignette
