### 2025-05-21 de15b1d
rewrite 2nd half of vignette for #125
