+++ vignettes/E_why-are-the-results-not-what-i-expect.Rmd
@@ -214,20 +214,24 @@ package in terms of similarities without function documentation text. Thus,
 regardless of how LM and BM25 values are combined, `lubridate` is never
 included in the top five matches.
 
+While BM25 values are derived from empirical word frequencies, and so are
+effectively fixed, the LM similarity metrics are derived from embeddings, and
+as stated at the output, these are inherently approximate. The next section
+describes how {pkgmatch} might be refined to improve the accuracy of these
+approximate representations.
 
 ---
+
 ## Modifying embeddings through chunking
 
-
-Results of `pkgmatch` functions can be improved through generating several
-embeddings for each component of each package. LLMs commonly achieve this
-through "chunking", which refers to taking different "chunks" of an input that
-generally extends beyond the admissible "context window" of the model. The
-models used here have context windows of 8096 tokens, which is often enough to
-enter the entire code or text of a package. Chunking effects can nevertheless
-be achieved through permuting components of the input in different orders.
-The embedding for "This plus that. Then those." will be different from that for
-"Then those. This plus that.":
+One common strategy to improve the accuracy of LM embeddings is "chunking",
+which refers to taking different "chunks" of an input that generally extends
+beyond the admissible "context window" of the model. The models used here have
+context windows of 8096 tokens, which is often enough to enter the entire code
+or text of a package. Chunking effects can nevertheless be achieved through
+permuting components of the input in different orders. The embedding for "This
+plus that. Then those." will differ to that for "Then those. This plus
+that.":
 
 ```{r ex-embeddings, eval = FALSE}
 head (get_embeddings ("This plus that. Then those."))
@@ -246,50 +250,46 @@ matrix (
 )
 ```
 
-And those two sets of embeddings are quite similar, and yet not identical. The
-following code modifies the procedure used within the `get_embeddings()`
-function, to create differently permuted chunks of the package input.
+Those two sets of embeddings are quite similar, and yet not identical. Chunking
+is generally used to generate several embeddings for any given input, with
+similarity metrics averaged over these several embeddings then providing a more
+accurate approximation to the underlying textual similarities. In other words,
+chunking reduces the noise inherent in approximate representation through
+embeddings.
+
+We will now demonstrate how chunking can be applied within the general
+"pkgmatch" workflow shown above. The following code modifies the procedure used
+within the `get_embeddings()` function, to create differently permuted chunks
+of the package input. (This requires calling two non-exported functions using
+the "three-dots" notation, `:::`, in contrast to the standard "two dots" for
+exported functions.)
 
 ```{r}
 txt_with_fns <- vapply (pkg_paths, function (p) pkgmatch:::get_pkg_text (p), character (1L))
 txt_wo_fns <- pkgmatch:::rm_fns_from_pkg_txt (txt_with_fns)
-code <- vapply (pkg_paths, function (p) pkgmatch:::get_pkg_code (p), character (1L))
 ```
 
-This code then takes advantage of the fact that the `get_pkg_text()` function
-inserts markdown-formatted section headers. These are broken up and randomly
-rearranged to form new chunks. For the full data, text from function
-definitions is nevertheless kept in place after all other package text.
[…]
