### 2025-05-21 67eaf62
add initial vignette #125 disclaimer about pre-calculation
