### 2025-05-21 cd61c9a
first full draft of vignette #125
