### 2025-05-21 213b36d
add missing similarity output df in vignette #125 [ci skip]
