### 2025-05-21 af4fcf7
add final section to complete vignette for #125 [ci skip]
