### 2025-05-21 ropensci-review-tools/pkgmatch#125 @mpadge
Those commits suffice to close this, by adding a new ["_Why are results not what I expect?_" vignette](https://docs.ropensci.org/pkgmatch/articles/E_why-are-the-results-not-what-i-expect.html).
