### 2025-05-21 50ec4be
rm good-enough-practices vignette
