### 2025-06-06 f338c97
start extended use-case vignette for #118
