### 2025-06-10 4b1e32e
cross-link vignettes in E
