### 2025-06-10 3fe7236
add LM section to vignette #118
