### 2025-06-10 26d34d4
more extended use case vignette #118
