### 2025-06-10 c6dcd59
start similar_fns section of extended use vignette #118
