### 2025-06-11 6b839f2
edit use-case vignette #118 thus far
