### 2025-06-12 5b10daf
rename sim_fns col 'function' -> 'pkg_fn'
Because 'function' is reserved and causes problems as a name
