### 2025-06-12 1b2bd4b
update man entry of simil-fns
