### 2025-06-12 ec19498
finish simil-fns sec of vignette #118
