### 2025-06-12 8e46530
start whole pkg search section in vignette #118
