### 2025-06-12 0bdac77
restructure 'get_pkg_readme' to general 'pkg_rmd' file for #147
