### 2025-06-12 e1c7946
include vignettes in pkg text for #147
