### 2025-06-12 d48fddb
fix cosine similarity calls in similar_pkgs_from_pkg for #147
