### 2025-06-13 bd85237
start implementation of embedding chunking for #147
