+++ R/embeddings.R
@@ -34,6 +34,12 @@ convert_paths_to_pkgs <- function (packages) {
 #'
 #' @param packages A vector of either names of installed packages, or local
 #' paths to directories containing R packages.
+#' @param n_chunks Number of randomly permuted chunks of input text to use to
+#' generate average embeddings. Values should generally be > 1, because the
+#' text of many packages exceeds the context window for the language models,
+#' and so permutations ensure that all text is captured in resultant
+#' embeddings. Note, however, that computation times scale linearly with this
+#' value.
 #' @param functions_only If `TRUE`, calculate embeddings for function
 #' descriptions only. This is intended to generate a separate set of embeddings
 #' which can then be used to match plain-text queries of functions, rather than
@@ -55,8 +61,14 @@ convert_paths_to_pkgs <- function (packages) {
 #' names (emb_pkg)
 #' colnames (emb_pkg$text_with_fns) # "curl"
 pkgmatch_embeddings_from_pkgs <- function (packages = NULL,
+                                           n_chunks = 5L,
                                            functions_only = FALSE) {
 
+    checkmate::assert_integer (n_chunks, len = 1L)
+    if (identical (Sys.getenv ("PKGMATCH_TESTS"), "true")) {
+        n_chunks <- 1L
+    }
+    chunk_seq <- seq_len (n_chunks)
     checkmate::assert_logical (functions_only, len = 1L)
     checkmate::assert_character (packages)
     if (all (grepl ("\\.tar\\.gz$", packages))) {
@@ -76,9 +88,13 @@ pkgmatch_embeddings_from_pkgs <- function (packages = NULL,
     if (!opt_is_quiet () && length (packages) > get_verbose_limit ()) {
         cli::cli_inform ("Extracting package text ...")
         txt_with_fns <-
-            pbapply::pblapply (pkgs_full, function (p) get_pkg_text (p))
+            pbapply::pblapply (pkgs_full, function (p) {
+                lapply (chunk_seq, function (i) get_pkg_text_internal (p))
+            })
     } else {
-        txt_with_fns <- lapply (pkgs_full, function (p) get_pkg_text (p))
+        txt_with_fns <- lapply (pkgs_full, function (p) {
+            lapply (chunk_seq, function (i) get_pkg_text_internal (p))
+        })
     }
 
     txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
@@ -199,18 +215,41 @@ pkgmatch_embeddings_from_text <- function (input = NULL) {
 rm_fns_from_pkg_txt <- function (txt) {
 
     lapply (txt, function (i) {
-        i_vec <- strsplit (i, "\\n") [[1]]
-        index <- grep ("^\\s*##\\s+Functions", i_vec)
-        if (length (index) > 0L) {
-            index <- seq (max (index), length (i_vec))
-            i_vec <- i_vec [-(index)]
+        is_list <- is.list (i)
+        if (!is_list) {
+            i <- list (i)
         }
-        paste0 (i_vec, collapse = "\\n")
+        res <- lapply (i, function (j) {
+            j_vec <- strsplit (j, "\\n") [[1]]
+            index <- grep ("^\\s*##\\s+Functions", j_vec)
+            if (length (index) > 0L) {
+                index <- seq (max (index), length (j_vec))
+                j_vec <- j_vec [-(index)]
+            }
+            paste0 (j_vec, collapse = "\\n")
+        })
+        if (!is_list) {
+            res <- unlist (res)
+        }
+        return (res)
     })
 }
 
 get_all_fn_descs <- function (txt) {
-    fn_txt <- lapply (txt, function (i) {
+
+    # txt is a list of (packages, chunks). First reduce down to the chunk with
[…]
