### 2025-06-13 4e2d18c
update fn docs for #147 with 'n_chunks' param
