### 2025-06-13 530fd8e
namespace fn calls
