### 2025-06-13 1ac9bc9
fix code for embeddings_from_pkg example #147
