### 2025-06-13 f1d4fec
improve extraction of fn defs for code embeddings
Get straight from parsing input files
