### 2025-06-13 a268b43
also chunk code inputs to embeddings for #147
