### 2025-06-13 8fd6931
update mock test results with chunked code embeddings for #147
