### 2025-06-13 c286e93
test verbose output of chunked code for #147
