### 2025-06-13 402e3a1
unambiguous names for test object results for #147
