### 2025-06-13 f13dfff
more temp test fixes
