### 2025-06-13 16f5e64
fix tests for #147 with withr::with_options
