### 2025-06-13 8191ed9
restructure testthat/withr lines to fix tests for #147
