### 2025-06-13 0809e7d
another try to fix tests for #147
capture_message fails, but expect_snapshot works, so it's a testthat issue, not here
