### 2025-06-13 547edff
set r-lib message verbosity for tests #147
