### 2025-06-13 4d703ad
reduce 'opt_is_quiet' to rm github runner condition #147
