### 2025-06-13 dfaa7ec
suppress verbose output on gh except test-coverage job #147
