+++ R/zzz.R
@@ -47,6 +47,16 @@ get_verbose_limit <- function () {
     getOption ("pkgmatch.verbose_limit")
 }
 
+# Includes line to suppress all output on GHA, via default envvars listed in
+# https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/store-information-in-variables#default-environment-variables # nolint
+# Generating terminal output on GHA generates segfault errors like
+# https://github.com/ropensci-review-tools/pkgmatch/actions/runs/12848889636/job/35826799668
+opt_is_quiet <- function () {
+    getOption ("rlib_message_verbosity", "notset") == "quiet" ||
+        (Sys.getenv ("GITHUB_ACTIONS") == "true" &&
+            Sys.getenv ("GITHUB_JOB") != "test-coverage")
+}
+
 #' @title Get the URL for local ollama API
 #'
 #' @description Return the URL of the specified ollama API. Default is
