### 2025-06-13 9ac8ac4
only expect snapshot of verbose embeddings output on test jobs #147
