### 2025-06-13 773b958
rm ambiguous 'Requires' statement from DESC
