### 2025-06-13 337508d
don't snapshot verbose embeddings output #148
Because that's only generated on GHA test-coverage job, so can't be locally generated anymore.
The conditions are all explicitly covered in the tests anyway, so there's no loss
