+++ tests/testthat/_snaps/get-embeddings.md
@@ -9,18 +9,3 @@
       Generating text embeddings [2 / 2] ...
       Generating code embeddings ...
 
----
-
-    Code
-      withr::with_options(list(pkgmatch.verbose_limit = 0L), {
-        emb <- httptest2::with_mock_dir("emb_raw", {
-          pkgmatch_embeddings_from_pkgs(packages)
-        })
-      })
-    Message
-      Extracting package text ...
-      Generating text embeddings [1 / 2] ...
-      Generating text embeddings [2 / 2] ...
-      Extracting package code ...
-      Generating code embeddings ...
-
