### 2025-06-15 dd19f98
fix extraction of embeddings from code
