### 2025-06-15 3d8c509
improve embeddings generate when no input #147
