### 2025-06-15 bf2de46
ensure release-data packages are dirs only
