### 2025-06-15 a622e68
update release-data-script to alll progressive re-starts [ci skip]
