+++ data-raw/release-data-script.R
@@ -8,208 +8,229 @@ path <- "/<path>/<to>/<ropensci>/<repos>"
 packages <- fs::dir_ls (path, type = "directory")
 
 # ----------------- EMBEDDINGS FOR ROPENSCI -----------------
-embeddings <- pkgmatch_embeddings_from_pkgs (packages)
-
-saveRDS (embeddings, "embeddings-ropensci.Rds")
-embeddings_fns <-
-    pkgmatch_embeddings_from_pkgs (packages, functions_only = TRUE)
-saveRDS (embeddings_fns, "embeddings-fns.Rds")
+f <- "embeddings-ropensci.Rds"
+if (!fs::file_exists (f)) {
+    embeddings <- pkgmatch_embeddings_from_pkgs (packages)
+    saveRDS (embeddings, f)
+}
+f <- "embeddings-fns.Rds"
+if (!fs::file_exists (f)) {
+    embeddings_fns <-
+        pkgmatch_embeddings_from_pkgs (packages, functions_only = TRUE)
+    saveRDS (embeddings_fns, f)
+}
 
 # -------------------- BM25 FOR ROPENSCI --------------------
-num_cores <- parallel::detectCores () - 2L
-cl <- parallel::makeCluster (num_cores)
-txt_with_fns <- pbapply::pblapply (
-    packages,
-    function (p) pkgmatch:::get_pkg_text (p),
-    cl = cl
-)
-parallel::stopCluster (cl)
+f <- c ("bm25-ropensci.Rds", "bm25-ropensci-fns.Rds")
+if (!all (fs::file_exists (f))) {
+    num_cores <- parallel::detectCores () - 2L
+    cl <- parallel::makeCluster (num_cores)
+    txt_with_fns <- pbapply::pblapply (
+        packages,
+        function (p) pkgmatch:::get_pkg_text (p),
+        cl = cl
+    )
+    parallel::stopCluster (cl)
 
-txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
-idfs <- list (
-    with_fns = bm25_idf (txt_with_fns),
-    wo_fns = bm25_idf (txt_wo_fns)
-)
-token_lists <- list (
-    with_fns = bm25_tokens_list (txt_with_fns),
-    wo_fns = bm25_tokens_list (txt_wo_fns)
-)
-bm25_data <- list (idfs = idfs, token_lists = token_lists)
-saveRDS (bm25_data, "bm25-ropensci.Rds")
+    txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
+    idfs <- list (
+        with_fns = bm25_idf (txt_with_fns),
+        wo_fns = bm25_idf (txt_wo_fns)
+    )
+    token_lists <- list (
+        with_fns = bm25_tokens_list (txt_with_fns),
+        wo_fns = bm25_tokens_list (txt_wo_fns)
+    )
+    bm25_data <- list (idfs = idfs, token_lists = token_lists)
+    saveRDS (bm25_data, f [1])
 
-txt_fns <- get_all_fn_descs (txt_with_fns)
-fns_idfs <- bm25_idf (txt_fns$desc)
-fns_lists <- bm25_tokens_list (txt_fns$desc)
-index <- which (vapply (fns_lists, nrow, integer (1L)) > 0L)
-fns_lists <- fns_lists [index]
-names (fns_lists) <- txt_fns$fn [index]
-bm25_data <- list (idfs = fns_idfs, token_lists = fns_lists)
-saveRDS (bm25_data, "bm25-ropensci-fns.Rds")
+    txt_fns <- get_all_fn_descs (txt_with_fns)
[…]
