### 2025-06-15 5e8c736
add cli headers to release-data-script for #147 [ci skip]
