### 2025-06-15 37a708c
emit all headers in release-data-script for #147 [ci skip]
