### 2025-06-15 5e15cf8
fs is regexp not regexpr [ci skip]
