### 2025-06-15 3ec08fd
catch error in eval for pkg-code #147
