### 2025-06-16 82066d1
fix function code parsing
Have to tryCatch on every single call, otherwise it fails
