### 2025-06-16 dac0391
catch all errors in text and code extraction loops
