### 2025-06-17 4957d17
randomly order rd files in each embedding chunk for #147
