### 2025-06-17 cb260ff
update vignettes for #147
