+++ vignettes/E_why-are-the-results-not-what-i-expect.Rmd
@@ -51,7 +51,7 @@ as is, results may differ slightly from those shown here.
 
 To start, we need to load the package:
 
-```{r library}
+```{r library, eval = FALSE}
 library (pkgmatch)
 ```
 
@@ -70,7 +70,7 @@ prints the top five matching packages:
 pkgmatch_similar_pkgs (input, corpus = "cran")
 ```
 ```{r sim-pkgs1-out, echo = FALSE}
-c ("datefixR", "fpp3", "rebus.datetimes", "sweep", "datetime")
+c ("iso8601", "fpp3", "tibbletime", "cleaner", "grates")
 ```
 
 And while expected results may not necessarily always be in the first position,
@@ -88,7 +88,7 @@ packages listed above, plus `lubridate`.
 
 ```{r dl-pkgs}
 pkgs <- c (
-    "datefixR", "fpp3", "rebus.datetimes", "sweep", "datetime", "lubridate"
+    "iso8601", "fpp3", "tibbletime", "cleaner", "grates", "lubridate"
 )
 path <- utils::download.packages (
     pkgs,
@@ -157,13 +157,12 @@ print (similarities)
 ```
 
 ```{r similarities-output, echo = FALSE}
-p <- c ("rebus.datetimes", "datetime", "datefixR", "lubridate", "sweep", "fpp3")
-wi <- c (0.8304852, 0.8202274, 0.8194707, 0.7905832, 0.7714603, 0.7071283)
-wo <- c (0.8186134, 0.8351010, 0.8186624, 0.7940536, 0.7984862, 0.8093774)
+p <- c ("grates", "fpp3", "iso8601", "tibbletime", "lubridate", "cleaner")
+wi <- c (0.0123306, -0.0031622, -0.0129680, -0.0201389, -0.0311186, -0.0342489)
+wo <- c (-0.0106373, 0.0051597, -0.0519957, 0.0264501, 0.0112935, -0.0004984)
 similarities <- data.frame (package = p, simil_with_fns = wi, simil_wo_fns = wo)
 print (similarities)
 ```
-
 
 That generates two vectors of similarities, with values for each package
 generated from embeddings calculated both with and without text from function
@@ -188,9 +187,9 @@ print (similarities)
 ```
 ```{r idf-bm25-output, echo = FALSE}
 similarities$bm25_with_fns <-
-    c (12.7782429, 12.6079534, 13.5863247, 0.2140736, 15.6956661, 12.2372817)
+    c (13.90114, 14.62466, 14.55312, 15.85009, 13.49620, 17.64303)
 similarities$bm25_wo_fns <-
-    c (13.1704148, 12.6606973, 16.3750351, 0.6645225, 15.0961380, 14.8849026)
+    c (17.37301, 17.43921, 18.50910, 17.26252, 16.67262, 20.56125)
 print (similarities)
 ```
 
@@ -206,9 +205,8 @@ pkgmatch_rerank (similarities)
 ```
 ```{r rank-output-output, echo = FALSE}
 data.frame (
-    package = c (
-        "datefixR", "datetime", "sweep", "rebus.datetimes", "fpp3",
-        "lubridate"
+    packages = c (
+        "cleaner", "tibbletime", "fpp3", "iso8601", "lubridate", "grates"
     ),
     rank = 1:6
 )
@@ -320,12 +318,10 @@ print (similarities)
 ```{r similar-with-chunks-output, echo = FALSE}
 data.frame (
     package = c (
-        "datetime", "rebus.datetimes", "datefixR", "lubridate",
-        "sweep", "fpp3"
+        "iso8601", "grates", "tibbletime", "fpp3", "cleaner", "lubridate"
     ),
     similarity = c (
-        0.8379673, 0.8235914, 0.8203636, 0.7936824, 0.7849108,
-        0.7724631
[…]
