### 2025-06-17 9b29e63
version bump -> 0.5.0 for #147
