+++ vignettes/F_extended-use-case.Rmd
@@ -219,7 +219,7 @@ Entire packages can also be used as input to pkgmatch functions. The simplest
 way to do this is to submit the name of an installed package, like this:
 
 ```{r whole-pkg-search, eval = FALSE}
-pkgs <- pkgmatch_similar_pkgs ("crul", corpus = "cran")
+pkgs <- pkgmatch_similar_pkgs ("osmdata", corpus = "cran")
 ```
 
 As explained in [the _How does pkgmatch work?_
@@ -238,7 +238,14 @@ pkgs
 ```
 ```{r whole-pkg-search-output, echo = FALSE}
 list (
-    text = c ("crul", "A3", "RCurl", "AmpGram", "AalenJohansen"),
-    code = c ("abclass", "A3", "webutils", "AalenJohansen", "tgver")
+    text = c ("factormodel", "crul", "healthequal", "httr2", "crumble"),
+    code = c ("shinyKGode", "crul", "sphereML", "irtawsi", "httr2")
 )
 ```
+
+As expected, the `crul` package appears in the top results for matching both
+text and code (although not in first place, for reasons described in [the
+_Why are the results not what I expect?_
+vignette](https://docs.ropensci.org/pkgmatch/articles/E_why-are-the-results-not-what-i-expect.html)).
+Moreover, the only two packages matched on both code and text are `crul`
+itself, and the highly-similar [`httr2` package](https://httr2.r-lib.org/).
