### 2025-06-18 9f1448a
rewrite initial pkgmatch section of use-case vignette #118
