+++ vignettes/F_extended-use-case.Rmd
@@ -130,27 +130,40 @@ pkgmatch::pkgmatch_similar_pkgs (text, corpus = "cran")
 c ("RWsearch", "ore", "sos", "gghilbertstrings", "rjsoncons")
 ```
 
-Two of those five results are directly related, while others provide the
-described functionality in specific rather than general domains. (To see
-details, assign the result of this `pkgmatch_similar_pkg()` query to a value
-and pass that to
-[`pkgmatch_browse()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_browse.html)
-to open the resultant package pages in your default web browser.) This
-illustrates a contrast between the overly-general results of search engines or
-language models, and pkgmatch results which are highly specific because they
-are restricted to R packages only.
+Of those top five matches:
 
-In this case, we expect no match for reasons described above. Attempting to
-generate a match using search engines or language models requires extensive
-post-filtering of generally irrelevant results before reducing the potential
-pool of results to zero, in order to discover that there is no package which
-does what we want. Attempting to generate a result using pkgmatch very quickly
-reveals that matching packages do indeed exist, yet all are restricted to
-specific domains. Within those domains, the results are nevertheless useful,
-and so more likely to lead to productive next steps through forcing us to
-reduce expectations of a general package, to consider the potential of perhaps
-achieving what we wanted through using one of the domain-specific ones returned
-by pkgmatch.
+- The ['RWsearch'
+package](https://cran.r-project.org/web/packages/RWsearch/index.html) is
+directly related.
+- The ['ore' package](https://github.com/jonclayden/ore) is a regular
+expression interface, and so seems mis-matched, yet nevertheless provides a
+host of search-related functions, including many function names which include
+"search".
+- The ['sos' package](https://github.com/sbgraves237/sos) is also directly
+related.
+
+The other two seem unrelated, but:
+
+- The ['gghilbertstrings' package](https://github.com/Sumidu/gghilbertstrings)
+seems entirely unrelated, but it does present an [example based on
+"search-engine-results"](https://github.com/Sumidu/gghilbertstrings), and so
+maybe appears for that reason?
+- The ['rjsoncons' package](https://github.com/mtmorgan/rjsoncons) is also not
+clearly related to our search term, but like [the 'ore'
+package](https://github.com/jonclayden/ore) also provides a wealth of functions
+related to querying and extraction.
+
+These results illustrate a contrast between the overly-general results of
+search engines or language models, and pkgmatch results which are highly
+specific because they are restricted to R packages only. Although we expected
+no precise match for reasons described above, pkgmatch nevertheless revealed
+the presence of two very clearly related packages,
+['RWsearch'](https://cran.r-project.org/web/packages/RWsearch/index.html) and
+['sos'](https://github.com/sbgraves237/sos), neither of which were returned in
+results from either [search
+engines](https://duckduckgo.com/?q=R%20package%20to%20return%20web%20search%20engine%20results%20into%20R%20as%20strings%20or%20URLs%20r%20programming),
+or [large](https://www.phind.com/search/cmbqhybqt0000206imod5zpqo) language
+[models](https://www.perplexity.ai/search/r-package-to-return-web-search-9mTr_z_eT2iqkKutaU3YYA).
 
 ## Searches based on function code
 
