### 2025-06-18 3b43fe8
finish use-case vignette to close #118
