### 2025-06-18 ropensci-review-tools/pkgmatch#116 @mpadge
Plan is now to enable most functionality to work without ollama, through reverting to BM25 matches only, and issue an accompanying warning.
