### 2025-06-18 6687ed6
allow similar_pkgs_from_text to run without ollama for #116
