### 2025-06-18 ff8da7c
allow similar_pkgs_from_pkg to run without ollama for #116
