### 2025-06-18 9b8e1af
update man entry with latest pkg release
