### 2025-06-18 eb735ee
suppress warning in example code for #116
