### 2025-06-18 7a4d40a
detail locally-dependent token limit in vignette #125
