### 2025-06-21 #671 @MargaretSiple-NOAA
Hi @mpadge -- I am at sea currently and I realize I haven't kept up on my check-ins. Shall we put this on hold until July 11, when I return from this survey? That will give you some additional time to get through these very thorough and helpful comments from both reviewers.
