### 2025-06-22 9147643
add 'local_cran_mirror' param to main data_update fn
