### 2025-06-22 75b3bb6
rm redundant requireNamespace(piggyback) calls
