### 2025-06-22 0f3c350
pass common 'flist' data to update_cran/ros fns
Remove previous duplication within each separate fn
