### 2025-06-22 ba5606f
fix 'local_mirror_path' varname in update_data
