### 2025-06-22 eba1bcb
add 'local_mirror_path' param to update-ropensci fn
