### 2025-06-22 60a7b2c
improve logic for uploading updated files
