### 2025-06-22 1a47e38
fix update-cran return value when no updates
