### 2025-06-23 372466d
update pre-commit-config
