### 2025-06-23 2fa8b52
fix wrong var name in update fn [ci skip]
