### 2025-06-23 8f3fb24
switch off failing test on macos
