### 2025-06-23 #671 @mpadge
Thanks @MargaretSiple-NOAA, yes, I think that's a good idea. Updates in the meantime:

- I failed to keep my [previous promise](https://github.com/ropensci/software-review/issues/671#issuecomment-2894480752) to get responses back in time, for which I can only apologise
- i have nevertheless been implementing many improvements, most notably through addressing the concerns of @agricolamz about results not matching expectations.
- Documentation and vignettes have been improved and greatly extended.

I'm confident I'll be able to post something here before July 11. Thank you all for your patience.
