+++ README.Rmd
@@ -66,13 +66,17 @@ The package can then loaded for use with
 library (pkgmatch)
 ```
 
-The package takes input either from a text description or local path to an R
-package, and finds similar packages based on both Language Model (LM)
-embeddings, and more traditional text and code matching algorithms. The LM
-embeddings require a locally-running instance of [ollama](https://ollama.com),
-as described in a separate vignette.
+### ollama installation
+
+The package works with language model (LM) embeddings generated by a 
+locally-running instance of [ollama](https://ollama.com). This needs to be installed and configured
+by following the procedures described in a separate vignette.
 
 ## Using the `pkgmatch` package
+
+The 'pkgmatch' package takes input either from a text description or local path to an R
+package, and finds similar packages based on both Language Model (LM)
+embeddings, and more traditional text and code matching algorithms. 
 
 The package has two main functions:
 
