+++ README.md
@@ -57,13 +57,19 @@ The package can then loaded for use with
 library (pkgmatch)
 ```
 
-The package takes input either from a text description or local path to
-an R package, and finds similar packages based on both Language Model
-(LM) embeddings, and more traditional text and code matching algorithms.
-The LM embeddings require a locally-running instance of
-[ollama](https://ollama.com), as described in a separate vignette.
+### ollama installation
+
+The package works with language model (LM) embeddings generated by a
+locally-running instance of [ollama](https://ollama.com). This needs to
+be installed and configured by following the procedures described in a
+separate vignette.
 
 ## Using the `pkgmatch` package
+
+The ’pkgmatch' package takes input either from a text description or
+local path to an R package, and finds similar packages based on both
+Language Model (LM) embeddings, and more traditional text and code
+matching algorithms.
 
 The package has two main functions:
 
@@ -127,11 +133,12 @@ pkgmatch_similar_pkgs (path, corpus = "cran")
 ```
 
     ## $text
-    ## [1] "luca"          "httr2"         "httptest2"     "scatterplot3d"
-    ## [5] "wesanderson"  
+    ## [1] "luca"             "httr2"            "httptest2"        "scatterplot3d"   
+    ## [5] "KnapsackSampling"
     ## 
     ## $code
-    ## [1] "prenoms"      "httr2"        "paperplanes"  "httptest2"    "text2sdgData"
+    ## [1] "ggplot2movies" "httr2"         "prenoms"       "httptest2"    
+    ## [5] "crul"
 
 The `input` parameter can also be a local path to a full source code
 repository.
@@ -172,9 +179,7 @@ corresponding to those best-matching functions.
 ## Contributors
 
 <!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
-
 <!-- prettier-ignore-start -->
-
 <!-- markdownlint-disable -->
 
 All contributions to this project are gratefully acknowledged using the
@@ -185,49 +190,34 @@ specification. Contributions of any kind are welcome!
 ### Code
 
 <table>
-
 <tr>
-
 <td align="center">
-
 <a href="https://github.com/mpadge">
 <img src="https://avatars.githubusercontent.com/u/6697851?v=4" width="100px;" alt=""/>
 </a><br>
 <a href="https://github.com/ropensci-review-tools/pkgmatch/commits?author=mpadge">mpadge</a>
 </td>
-
 <td align="center">
-
 <a href="https://github.com/Bisaloo">
 <img src="https://avatars.githubusercontent.com/u/10783929?v=4" width="100px;" alt=""/>
 </a><br>
 <a href="https://github.com/ropensci-review-tools/pkgmatch/commits?author=Bisaloo">Bisaloo</a>
 </td>
-
 </tr>
-
 </table>
 
 ### Issues
 
 <table>
-
 <tr>
-
 <td align="center">
-
 <a href="https://github.com/MargaretSiple-NOAA">
 <img src="https://avatars.githubusercontent.com/u/73858992?u=7ea549d423535a74d69a75ff6303af35496290fb&v=4" width="100px;" alt=""/>
 </a><br>
 <a href="https://github.com/ropensci-review-tools/pkgmatch/issues?q=is%3Aissue+author%3AMargaretSiple-NOAA">MargaretSiple-NOAA</a>
 </td>
-
 </tr>
-
 </table>
-
 <!-- markdownlint-enable -->
-
 <!-- prettier-ignore-end -->
-
 <!-- ALL-CONTRIBUTORS-LIST:END -->
