+++ README.Rmd
@@ -19,18 +19,20 @@ Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repost
 
 # pkgmatch
 
-A tool to help find R packages, by matching packages either to a text
-description, or to any given package. Can find matching packages either from
-rOpenSci's [suite of packages](https://ropensci.org/packages), or from all
-packages currently on [CRAN](https://cran.r-project.org).
+A tool that uses language models to help find R packages, by matching packages
+either to a text description, or to any given package. Can find matching
+packages either from rOpenSci's [suite of
+packages](https://ropensci.org/packages), or from all packages currently on
+[CRAN](https://cran.r-project.org).
 
 ## Installation
 
 This package relies on a locally-running instance of
 [ollama](https://ollama.com). Procedures for setting that up are described in a
 [separate vignette](https://docs.ropensci.org/pkgmatch/articles/ollama.html)
-(`vignette("ollama", package = "pkgmatch")`). ollama needs to be installed
-before this package can be used.
+(`vignette("ollama", package = "pkgmatch")`). Although some functionality of
+this package may be used without ollama, the main functions require ollama to
+be installed.
 
 Once ollama is running, the easiest way to install this package is via the
 [associated
@@ -66,23 +68,18 @@ The package can then loaded for use with
 library (pkgmatch)
 ```
 
-### ollama installation
-
-The package works with language model (LM) embeddings generated by a 
-locally-running instance of [ollama](https://ollama.com). This needs to be installed and configured
-by following the procedures described in a separate vignette.
-
 ## Using the `pkgmatch` package
 
-The 'pkgmatch' package takes input either from a text description or local path to an R
-package, and finds similar packages based on both Language Model (LM)
+The 'pkgmatch' package takes input either from a text description or local path
+to an R package, and finds matching packages based on both Language Model (LM)
 embeddings, and more traditional text and code matching algorithms. 
 
 The package has two main functions:
 
 - [`pkgmatch_similar_pkgs()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
 to find similar rOpenSci or CRAN packages based input as either a local path to
-an entire package, or as a single descriptive text string; and
+an entire package, the name of an installed package, or as a single descriptive
+text string; and
 - [`pkgmatch_similar_fns()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_fns.html)
 to find similar functions from rOpenSci packages based on descriptive text
 input. (Not available for functions from CRAN packages.)
@@ -102,16 +99,18 @@ The corpus parameter must be specified as one of "ropensci" or "cran". The CRAN
 corpus is much larger than the rOpenSci corpus, and matching for `corpus =
 "cran"` will generally take notably longer.
 
-Websites of packages returned by the
-[`pkgmatch_similar_pkgs()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
-function can be automatically opened, either by passing `browse = TRUE`, or by
-storing the value of a function as an object and passing that to the
-[`pkgmatch_browse()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_browse.html)
-function.
+Websites of packages returned by [the `pkgmatch_similar_pkgs()`
+function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
+can be automatically opened, either by passing `browse = TRUE`, or by storing
+the return value of [the `pkgmatch_similar_pkgs()`
+function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
+as an object and passing that to [the `pkgmatch_browse()`
+function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_browse.html).
[…]
