### 2025-06-24 f87194e
skip failing no-ollama test #116 on windows
