### 2025-06-24 b110586
update pkgmatch vignette
