### 2025-07-02 27776c7
document return value to update cache fn [ci skip]
