### 2025-07-08 eb39715
improve test in readme, vignette for #117
