### 2025-07-08 ropensci-review-tools/pkgmatch#109 @mpadge
I think this has been covered in #117, and a few other improvements along the way.
