### 2025-07-08 2af19dd
restructure vignettes for #111
