+++ vignettes/F_why-are-the-results-not-what-i-expect.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set (
 ```
 
 This vignette presumes that you've read [the "_How does pkgmatch work?_"
-vignette](https://docs.ropensci.org/pkgmatch/articles/B_how-does-it-work.html),
+vignette](https://docs.ropensci.org/pkgmatch/articles/C_how-does-it-work.html),
 which initially explains that,
 
 > The "pkgmatch" package package finds packages, as well as individual
