### 2025-07-09 ropensci-review-tools/pkgmatch#112 @mpadge
I think this has been satisfactorily addressed, primarily through all the commits linked in #114
