### 2025-07-09 #671 @mpadge
I am deeply indebted to both @Selbosh and @agricolamz for your extremely helpful reviews. With due apology for my response having taken so long, please note that this was largely because the package has been very extensively rewritten, restructured, and improved in response to your feedback.

Please note also that the following responses were largely written as "mere" technical responses, and so are largely devoid of expressions of gratitude or appreciation. Rest assured: all comments were enormously appreciated, and changes made in response have really helped to improve the package enormously, for which I am very grateful. Thank you both! I'll post separate responses here, and for ease of subsequent discussion prefix each one with unique identifiers.

# Responses to @Selbosh

## Documentation

#### DS01

> In the README, the statement of need should be more prominent; it's not obvious who the target audience of the package is or why the package is needed. It would be great if you could explain early on, clearly, what problem pkgmatch solves, maybe giving examples of use cases from the start.

Addressed via https://github.com/ropensci-review-tools/pkgmatch/issues/109 and https://github.com/ropensci-review-tools/pkgmatch/issues/117

README now starts with explicit description of LLM usage, and ["Get started" vignette](https://docs.ropensci.org/pkgmatch/articles/pkgmatch.html) starts with a longer description of what the package does and why. That vignette also proceeds straight to a [_What does the package do?_ section](https://docs.ropensci.org/pkgmatch/articles/pkgmatch.html#what-does-the-package-do).

---

#### DS02

> I think that you could improve the documentation by making it more self-contained and making it easier for the user to find information quickly. Some information that should be in the README is only in a vignette, and some information that should be at the start of one vignette is instead left as a link to another vignette.

Addressed mainly via https://github.com/ropensci-review-tools/pkgmatch/issues/111, which extensively restructured the vignettes. The most important improvement is the list at the [end of the REAMDE](https://docs.ropensci.org/pkgmatch/index.html#package-vignettes) with brief descriptions of the contents and purposes of all vignettes. As explained [in that issue](https://github.com/ropensci-review-tools/pkgmatch/issues/111#issuecomment-3051817461),

---

#### DS03

> All individual vignettes have been retained (and an extra one added via #118), because I hope this should enable people to more readily find what they need, rather than having to parse sections of longer vignettes

> The dependence on LLMs is also somewhat buried: this could be a make-or-break feature for many users, but it's only mentioned at the bottom of the Installation section. I can imagine it could be a selling point that pkgmatch exploits LLMs, but it could also be a deal-breaker for some users either due to technical limitations or ethical concerns.

That was a really important point, and has now been addressed by revising initial text of [main README](https://docs.ropensci.org/pkgmatch/index.html) to state that `pkgmatch` is,

> A tool that uses languages models to help find R packages.

All text changes are linked in [issue#117](https://github.com/ropensci-review-tools/pkgmatch/issues/117#issuecomment-3048200447). But most importantly, [issue#116](https://github.com/ropensci-review-tools/pkgmatch/issues/116) enables almost all functions to work without ollama, and issue warnings that results may be less accurate that those with ollama. The previous absolute need for ollama is thus effectively circumvented, which hopefully avoids the "deal-breaker" status.

---

#### DS04

> The package relies on ollama but it's not explained in the README what this is or why the package should depend on it, or whether you can use some functions without it. There is a function called ollama_check() that ostensibly checks for and installs ollama: add this to the README, as it seems like the first thing you would want to run after loading the package! "Procedures for setting that up are described in a separate vignette" is not super helpful without a direct link to or giving the name of said vignette! Ideally I'd put much of this information in the README, because it could be a decider on whether somebody can/will install the package in the first place. Given we are talking about local LLMs here, not everybody will know what their computational demands are, so it would be helpful to give some general indication of system requirements in the 'Before you begin' vignette and maybe in the README.

All done via above-linked https://github.com/ropensci-review-tools/pkgmatch/issues/116 and https://github.com/ropensci-review-tools/pkgmatch/issues/117. Initial bit of README now starts,

> This package relies on a locally-running instance of [ollama](https://ollama.com). Procedures for setting that up are described in a [separate vignette](https://docs.ropensci.org/pkgmatch/articles/B_ollama.html) (`vignette("ollama", package = "pkgmatch")`). Although some functionality of this package may be used without ollama, the main functions require ollama to be installed.

That vignette also now includes [this statement](https://docs.ropensci.org/pkgmatch/articles/B_ollama.html):

> Note that the [ollama](https://ollama.com) software is around 1GB, and its usage in `pkgmatch` requires downloading two additional language models of around 300MB each. Installation will thus use around 2GB of disk space (and will generally take quite some time). [ollama](https://ollama.com) will automatically run on GPUs if detected, but can also run (in somewhat slower mode) on CPUs.

---

#### DS05

> Generally speaking, I think the documentation assumes a bit too much knowledge on the part of the user, by not motivating the task clearly enough, not explaining what some functions do or what some terms mean. There are one or two functions documented in terms of what they are not rather than what they are, which can be a bit confusing (e.g. pkgmatch_bm25_fn_calls()). Some function descriptions (e.g. pkgmatch_bm25()) simply comprise an external Web link, which is not ideal for offline use, nor for users who want to know what a function does without having to click around. For the same functions, the current Titles in the documentation are too long or specific and would be better off as Descriptions, and the current Descriptions as Details or Notes.

All function documentation extensively revised and improved in response to this very valid criticism, mostly in https://github.com/ropensci-review-tools/pkgmatch/issues/110. See [example for `pkgmatch_bm25()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_bm25.html), which now starts with an extended description of how this function actually works.

---

#### DS06

> The purpose of some of the vignettes is unclear. For example, "Good-enough practices for language model packages": who is this for, exactly? Why does it need to be a vignette? Why do "Data caching and updating" or "Why local LMs?" need their own vignettes, instead of being included as sections in a larger general guide? This information is surely useful, but it could be merged into a single vignette and/or the README, maybe with a dedicated vignette for developers, to better signpost the user to the information they need (or don't need).

That previous "Good-enough practices" vignette has been completely removed. While I agree with the general concern that multiple vignettes may be initially more confusing that fewer, longer vignettes, I have nevertheless retained all individual ones, but added much clearer signposting via https://github.com/ropensci-review-tools/pkgmatch/issues/111. As stated above, I think that retaining multiple vignettes will help people more directly find relevant information, rather than having to wade through more extended ones. That then relies on them being sufficiently signposted, for which I hope that [the explanations in the README](https://docs.ropensci.org/pkgmatch/index.html#package-vignettes) do a good job.

----

## Functionality

#### DS07

> I was able to install the package (v0.4.3.018) first try on WSL2 (Ubuntu on Windows 11) and didn't face any issues with building vignettes or R CMD CHECK. All tests passed successfully.

> Not really sure where to start, I decided to run the example in the README:

> pkgmatch_similar_pkgs(input)

> This threw an error:

> Error in names(out) <- nms :
>   'names' attribute [1] must be the same length as the vector [0]

> I then remembered that I hadn't started the ollama server, so I go to the command line and run ollama serve, to get

> Error: listen tcp 127.0.0.1:11434: bind: address already in use

> This might be quite esoteric but could be worth mentioning in the vignette. A bit of detective work suggests that the error message above is in fact coming from ollama_check(), which when run throws exactly the same error. This left me a bit stuck, due to a couple of issues in the error handling, and due to a bug in the embeddings.R file. I made a local patch to fix it on my end and was able to proceed from here.

I think all of those concerns have been addressed via https://github.com/ropensci-review-tools/pkgmatch/issues/112, https://github.com/ropensci-review-tools/pkgmatch/issues/113, and other issues cross-linked within those. There should generally be far fewer errors, especially as https://github.com/ropensci-review-tools/pkgmatch/issues/116 allows the package to work without ollama running at all.

---

#### DS08

> Once I got the package working, the benefits of it seem clear! I can see this being immensely useful to package developers and advanced users who are looking for existing functionality, either for specialist tasks or trying not to reinvent the wheel while developing their own packages. However, as described above, the documentation comes across as being aimed at people who are already familiar with the package, rather than to new users who are trying to get it to work for the first time.

Thank you for saying that! There's a whole new vignette with [_Example applications_](https://docs.ropensci.org/pkgmatch/articles/A_extended-use-case.html) which hopefully makes the general case even clearer and more convincing. (But see *DS13* below for remaining concerns I have with that vignette.)

---

#### DS09

> Another interface issue I note is that with most function calls, there is a slight delay and we see that data is being downloaded in the background:

> pkgmatch_similar_pkgs (".", corpus = "cran")
> [100%] Downloaded 231213455 bytes...
> [100%] Downloaded 36834164 bytes...
> [100%] Downloaded 3948804 bytes...
> [100%] Downloaded 13016426 bytes...

> but what data is being downloaded, why and where to? Is it the corpus? Is it the embeddings? Maybe we should ask the user if they'd like to use up their bandwidth and disk space on this, or reassure them that it's going to a temporary directory and/or can easily be cleaned out at the end of the session? Which functionality will work offline, if any? This could be made a bit clearer on the first use.

That has all been addressed via https://github.com/ropensci-review-tools/pkgmatch/issues/114, which means:

- An initial message is given saying how many datasets need to be downloaded
- Downloads then give name of the download, as well as count (`1 / 10`, for example)
- Information is given about _where_ these are cached, and also that this cache directory may be safely removed at any time.

---

#### DS10

> The browse = TRUE function threw a further error for me, possibly related to using VS Code.

> pkgmatch_similar_fns (input, browse = TRUE)
> VSCode WebView only supports showing local http content.
> Opening in external browser...
> Browsing https://docs.ropensci.org/charlatan/reference/SequenceProvider
> Error in vapply(urls, utils::browseURL, integer(1L)) :
>   values must be length 1,
>  but FUN(X[[1]]) result is length 0

I think that should have all been fixed. Linking functions to package URLs is a bit of a dark art, because `.Rd` files actually specify a "topic" which may or may not match on to actual function names. I expect issues may still arise when trying to open pages from `pkgmatch_similar_fns()`, but am confident it's at a generally useable state from which issues can be invidiually resolved if an when they arise.

---

#### DS11

> However, the help page successfully opened in the browser, so I'm not sure what the issue is here. I was able to run other examples successfully locally, though nearly all of them are 'Don't run' because of the external dependencies; I wonder if there is the possibility to embed some toy embeddings and system.files in the package?

That was a great idea, and has been done exactly as suggested with the help of [a `generate_pkgmatch_example_data()` function](https://docs.ropensci.org/pkgmatch/reference/generate_pkgmatch_example_data.html). See active example in use at https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html#ref-examples.

---

#### DS12

> In the examples and README, you have spaces between function names and the brackets, e.g. fun () instead of fun(), which is not recommended in most style guides.

I know, and apologise for any cognitive challenge in visually parsing my code style. It's called [the `spaceout` style](https://github.com/ropensci-review-tools/spaceout), and explained [in 'CONTRIBUTING.md'](https://github.com/ropensci-review-tools/pkgmatch/blob/main/CONTRIBUTING.md#code-style).

---

#### DS13

> Overall I think this is a good idea for a package with real potential, but the main thing holding it back is selling to the user what it's for and giving an easy way to get started. To do this, I'd suggest filling in the gaps and maybe focus on having a vignette with an extended use case that shows off the package's capabilities.

That's really useful feedback, thank you. The ["extended use case" vignette](https://docs.ropensci.org/pkgmatch/articles/A_extended-use-case.html) has been written. I still worry that I'm a bit too technically close to everything here to clearly see the proverbial forest, so would appreciate any specific further feedback you might be able to provide in order to better address your concerns here. I belive reasonably firmly that I have satisfactorily addressed all review concerns with the singular exception of this one, which still worries me a bit. Any and all further help will be greatly appreciated!

---

#### DS14

> In the long run, I'd suggest clearly marking out which functionality is dependent on which external tools, as then a new user might be able to explore some features even if they have issues with others or do not want to set them up yet.

I trust this has been addressed via https://github.com/ropensci-review-tools/pkgmatch/issues/116, which allows (almost) everything to work without external tools (ollama).
