### 2025-07-18 #671 @agricolamz
Hi, I didn't have much time to react yet, but I wanted to note that it make sense to add some motivation for using your package comparing it to the existing tools:

1) https://r-packages.io/
2) LLMs themselfs
