### 2025-07-21 45bbf9e
fix names of bm25 data, and rm duplicates to fix #182
