### 2025-07-21 dd46276
update .hooks/descr
