### 2025-07-24 18ff248
call untar with explicit connections so they can be closed
