### 2025-07-24 e4dd8be
clean up connections after untar for #187
