### 2025-07-25 ea0ab40
rm semi-duplicated 'tarball_to_path' fn; use only 'extract_tarball'
which now extracts straight into temp dir only, not 'tarballs' sub-dir
