### 2025-07-25 0ae112d
test updated tarball extraction fns
And update test-similar to skip if ollama running
