### 2025-07-25 3af43ae
fix tarball extraction and update release-data-script
