+++ data-raw/release-data-script.R
@@ -208,19 +208,11 @@ if (!all (fs::file_exists (f))) {
     calls <- pbapply::pblapply (packages, function (f) {
         res <- NULL
         ntries <- 0L
-        td <- fs::path (fs::path_temp (), "tarballs")
         while (is.null (res)) {
-            # i <- match (f, packages)
-            # cli::cli_alert_info ("[{i}] {f} [{ntries}]")
             res <- tryCatch (
                 pkgmatch::pkgmatch_treesitter_fn_tags (f),
                 error = function (e) NULL
             )
-            if (is.null (res)) {
-                if (fs::dir_exists (td)) {
-                    fs::dir_delete (td)
-                }
-            }
             ntries <- ntries + 1L
             if (ntries > 5L) {
                 break
@@ -228,11 +220,6 @@ if (!all (fs::file_exists (f))) {
         }
         if (is.null (res)) {
             res <- data.frame (name = character (0L))
-            # clean temp dir to restart:
-            td <- fs::path (fs::path_temp (), "tarballs")
-            if (fs::dir_exists (td)) {
-                fs::dir_delete (td)
-            }
         }
         # Need to explicitly close any extra connections:
         cons <- showConnections (all = TRUE)
