### 2025-07-25 3da4a68
close all potentiall hanging connections on untar to fix #187
