+++ data-raw/release-data-script.R
@@ -201,7 +201,8 @@ if (!fs::file_exists (f)) {
 cli::cli_h1 ("CRAN function calls")
 f <- c ("fn-calls-cran.Rds", "idfs-fn-calls-cran.Rds")
 if (!all (fs::file_exists (f))) {
-    cli::cli_inform ("Extract function calls from all CRAN packages ...")
+    npkgs <- format (length (packages), big.mark = ",")
+    cli::cli_inform ("Extract function calls from {npkgs} CRAN packages ...")
     num_cores <- parallel::detectCores () - 1L
     cl <- parallel::makeCluster (num_cores)
 
@@ -221,7 +222,7 @@ if (!all (fs::file_exists (f))) {
         if (is.null (res)) {
             res <- data.frame (name = character (0L))
         }
-        # Need to explicitly close any extra connections:
+        # Need to explicitly close any extra connections (#187):
         cons <- showConnections (all = TRUE)
         index <- which (cons [, "isopen"] == "closed")
         for (i in index) {
