### 2025-07-25 f4ac4d6
skip test-treesitter on \!test_all; add missing fn exs
