### 2025-07-25 5b3bc73
skip all treesitters tests if \!test_all
B/c on pkgcheck it fails to find the namespace file
