+++ tests/testthat/test-treesitter.R
@@ -1,6 +1,8 @@
 test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") ||
     identical (Sys.getenv ("GITHUB_JOB"), "test-coverage"))
 
+skip_if (!test_all)
+
 test_that ("tree-sitter", {
 
     withr::local_envvar (list ("PKGMATCH_TESTS" = "true"))
@@ -22,7 +24,6 @@ test_that ("tree-sitter", {
 })
 
 
-skip_if (!test_all)
 test_that ("tree-sitter installed package", {
     pkg <- "rappdirs"
     tags <- pkgmatch_treesitter_fn_tags (pkg)
