### 2025-07-27 20313e1
pass explicit 'exdir' param to 'extract_tarball' fn
