### 2025-07-27 d7ed4a0
update release-data-script for #187
