### 2025-07-27 3441b23
rm unused connection clean-up lines
