### 2025-07-29 cfd1d6f
more updates to update-cran.R for #187
