### 2025-07-29 64dae49
update all ollama fns to use 'sys' calls for #187
