+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: pkgmatch
 Title:  Find R Packages Matching Either Descriptions or Other R Packages
-Version: 0.5.0.049
+Version: 0.5.0.050
 Authors@R: c(
     person("Mark", "Padgham", , "mark.padgham@email.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2172-5265")),
@@ -24,7 +24,9 @@ Imports:
     methods,
     pbapply,
     Rcpp,
+    readr,
     rvest,
+    sys,
     tibble,
     tidyr,
     tokenizers,
