### 2025-07-29 5cf9a5e
close 1 more connection for #187
