### 2025-07-29 fd534e3
finally located source of dangling connections and fixed for #187
