### 2025-07-29 d385f21
fix desc parsing in 'get_local_pkg_deps' for #187
