### 2025-07-29 05697f5
fix data-update-cran fn for #187
