+++ R/data-update-cran.R
@@ -52,7 +52,8 @@ pkgmatch_update_cran <- function (flist, local_mirror_path = NULL) {
     })
     paths <- unlist (paths)
 
-    cli::cli_inform ("Analysing {npkgs} packages...")
+    npkgs_dl <- length (paths)
+    cli::cli_inform ("Analysing {npkgs_dl} packages...")
     op_is_quiet <- opt_is_quiet ()
     op <- getOption ("rlib_message_verbosity")
     options ("rlib_message_verbosity" = "quiet")
@@ -65,18 +66,8 @@ pkgmatch_update_cran <- function (flist, local_mirror_path = NULL) {
         )
 
         if (!op_is_quiet) {
-            pkgmatch_update_progress_message (p, 1, npkgs, pt0, op_is_quiet)
+            pkgmatch_update_progress_message (p, 1, npkgs_dl, pt0, op_is_quiet)
         }
-
-        # This is necessary because `utils::untar()` can create hanging
-        # connections (see #187):
-        cons <- showConnections (all = TRUE)
-        index <- which (cons [, "isopen"] == "closed")
-        for (i in index) {
-            con <- getConnection (row.names (cons) [i])
-            tryCatch (close (con), error = function (e) NULL)
-        }
-
 
         return (dat)
     })
