### 2025-07-29 fef0c5c
copy full 'url_exists' fn from pkgcheck for #187
