### 2025-07-29 7d5ce37
revert ollama fns to 'system' not 'sys::' for #187
