### 2025-07-29 f9a4517
no need to close curl_download connection for #187
