### 2025-07-29 34b9c3f
update release-data-script for #187
