### 2025-07-29 e31d5ca
minor tweaks for #187
