### 2025-08-15 #671 @MargaretSiple-NOAA
Hi @mpadge, thank you for the revisions based on these great comments. I think the package has benefitted greatly from both reviews.

I have a couple of editing questions/comments while we await approvals:
- Can you use that 'submit response' command to the ropensci-review-bot so it officially logs both your responses to reviewers? I've seen folks skip this part but I want to be thorough.
- I really like your vignette **Example applications**. This seems like a good place for @agricolamz 's suggested packages/tools that might do comparable searching tasks (r-packages dot io, searches with broader multipurpose LLMs).
- I have found some clarity/usability edits that I think I would recommend to the vignettes (some are copy edit-adjacent but still impact the interpretation). I'm happy to handle those (or not) however you'd like.

@Selbosh and @agricolamz -- please look over @mpadge's changes and if you are satisfied with the revisions, post your approvals (or further remarks).
