### 2025-08-21 fbd3b63
update pre-commit-config
