+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@
 # R specific hooks: https://github.com/lorenzwalthert/precommit
 repos:
 -   repo: https://github.com/lorenzwalthert/precommit
-    rev: v0.4.3.9012
+    rev: v0.4.3.9013
     hooks:
     -   id: style-files
         args: [--style_pkg=spaceout, --style_fun=spaceout_style, --include_roxygen_examples=FALSE]    
@@ -51,7 +51,7 @@ repos:
     -   id: deps-in-desc
     # -   id: pkgdown
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
+    rev: v6.0.0
     hooks: 
     -   id: check-added-large-files
         args: ['--maxkb=200']
