### 2025-08-21 a9848f1
catch httr2 requests to cran which return empty body
