+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@
 # R specific hooks: https://github.com/lorenzwalthert/precommit
 repos:
 -   repo: https://github.com/lorenzwalthert/precommit
-    rev: v0.4.3.9013
+    rev: v0.4.3.9014
     hooks:
     -   id: style-files
         args: [--style_pkg=spaceout, --style_fun=spaceout_style, --include_roxygen_examples=FALSE]    
