### 2025-08-26 3fb0311
update pre-commit-config, gha workflows
