### 2025-09-01 0b4a179
update update workflow to twice per day [ci skip]
Too many recent fails from > 100 new pkgs per day; this should help
