+++ .github/workflows/update.yaml
@@ -3,7 +3,7 @@ on:
   #   branches:
   #     - main
   schedule:
-    - cron: 1 0 * * *
+    - cron: 1 0,12 * * *
 
 name: Update pkgmatch data
 
