+++ vignettes/C_how-does-it-work.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set (
 )
 ```
 
-The "pkgmatch" package package finds packages, as well as individual functions,
+The "pkgmatch" package finds packages, as well as individual functions,
 which best match a given input. Inputs can be text descriptions, sections of
 code, or even entire R packages. "pkgmatch" finds the most closely matching
 packages using a combination of Language Models (LMs, or equivalently, "LLMs"
@@ -47,7 +47,7 @@ this may lead to truncation of code inputs for very large packages, but that
 should generally be very unlikely.
 
 Text input of packages includes text from the package description (in the
-"DESCRIPTION" file), followed by the "README" file (where present), text of all
+"DESCRIPTION" file), followed by text from the "README" file (where present), text of all
 vignettes, and text of all function descriptions, and parameters of each
 function. All code is removed from these text sources, including removal of all
 code chunks in `.Rmd` documents.
@@ -86,9 +86,8 @@ similarities between inputs and all packages within these corpora.
 
 ### LM similarities
 
-"pkgmatch" access LMs through a local [ollama server](https://ollama.com), as
-described in a [separate
-vignette](https://docs.ropensci.org/pkgmatch/articles/B_ollama.html).
+"pkgmatch" accesses LMs through a local [ollama server](https://ollama.com), as
+described in the [ollama installation vignette](https://docs.ropensci.org/pkgmatch/articles/B_ollama.html).
 LMs are exclusively used here to generate "embeddings", which are numeric
 vectors of a fixed length representing a (reduced) version of the
 representation of the underlying text data within the multi-dimensional tensor
@@ -166,3 +165,8 @@ ranks submitted to the RRF algorithm. A default value of `k = 60` is used at
 all times, as recommended by the original paper, which also suggests very low
 sensitivity of results to variations in this value. Where inputs are entire
 packages, final rankings are generated separately for code and text.
+
+
+## References
+
+Cormack, G. V., Clarke, C. L. A. & Buettcher, S. Reciprocal rank fusion outperforms condorcet and individual rank learning methods. in Proceedings of the 32nd international ACM SIGIR conference on Research and development in information retrieval 758–759 (ACM, Boston MA USA, 2009). doi:10.1145/1571941.1572114.
