### 2025-09-08 bcc3eb4
minor copy edits to dx
