### 2025-10-27 e7017fa
fix :bug: in re-ranking fn
