### 2025-10-27 e3baadd
retain matches with fn text data in similar_pkgs fn
