### 2025-10-27 2ad391f
add mock test data for embeddings-fns
