### 2025-10-27 05ab10e
update ollama api call structure for #200
