### 2025-10-27 9ab0f3f
update all test mocks with new ollama api calls for #200
