### 2025-10-27 ddf699b
rm temp tarball exdir in data-update-cran
