### 2025-10-27 eb8daa3
revert tests that were switched off
