+++ tests/testthat/test-get-embeddings.R
@@ -1,4 +1,5 @@
-is_test_job <- identical (Sys.getenv ("GITHUB_JOB"), "test-coverage")
+is_test_job <- (identical (Sys.getenv ("GITHUB_JOB"), "test-coverage") ||
+    identical (Sys.getenv ("MPADGE_LOCAL"), "true"))
 
 expect_embeddings_matrix <- function (x) {
     expect_type (x, "double")
@@ -61,13 +62,12 @@ test_that ("raw embeddings", {
     )
     expect_identical (emb0, emb)
     if (is_test_job) {
-        # These tests fail on GH runners only, but I can't reproduce:
-        # expect_false (identical (msgs0, msgs))
-        # expect_length (msgs, 5L)
-        # expect_length (grep ("Extracting", msgs), 2L)
-        # expect_length (grep ("Generating", msgs), 3L)
-        # expect_length (grep ("text", msgs), 3L)
-        # expect_length (grep ("code", msgs), 2L)
+        expect_false (identical (msgs0, msgs))
+        expect_length (msgs, 5L)
+        expect_length (grep ("Extracting", msgs), 2L)
+        expect_length (grep ("Generating", msgs), 3L)
+        expect_length (grep ("text", msgs), 3L)
+        expect_length (grep ("code", msgs), 2L)
     }
 
     expect_type (emb, "list")
