+++ tests/testthat/test-pkg-data.R
@@ -1,3 +1,6 @@
+is_test_job <- (identical (Sys.getenv ("GITHUB_JOB"), "test-coverage") ||
+    identical (Sys.getenv ("MPADGE_LOCAL"), "true"))
+
 test_that ("get pkg local text", {
     path <- pkgmatch_test_skeleton ()
     expect_true (dir.exists (path))
@@ -20,7 +23,7 @@ test_that ("get pkg local text", {
     expect_true (nchar (code) < 1000)
 
     # ---- test utils fns -----
-    skip_if_not (identical (Sys.getenv ("GITHUB_JOB"), "test-coverage"))
+    skip_if_not (is_test_job)
     # Identified as code because of markdown
     expect_true (text_is_code (txt))
     expect_false (text_is_code ("This is not code"))
