+++ inst/httptest2/redact.R
@@ -1,0 +1,11 @@
+function (resp) {
+
+    resp <- httptest2::gsub_response (
+        resp,
+        "https://search.r-project.org/CRAN/refmans/testthat/html",
+        "search.r-project/",
+        fixed = TRUE
+    )
+
+    return (resp)
+}
