### 2025-10-28 ropensci-review-tools/pkgmatch#203 @mpadge
Fix model downloading messages
From @Selbosh's [review comment](https://github.com/ropensci/software-review/issues/671#issuecomment-3292032506)
