### 2025-10-28 a86304f
ensure rlib messgage verbosity set for downloading ollama models #203
