### 2025-10-28 fe4e2b0
fix parsing of ollama list output for #203
