### 2025-10-28 4a57a1f
improve initial model dl messages for #203
