### 2025-10-28 2930736
fix 'dl_jina_model' fn for #203
