+++ R/ollama.R
@@ -136,7 +136,7 @@ ollama_dl_jina_model <- function (what = "base") {
     if (ollama_has_jina_model (what)) {
         return (TRUE)
     }
-    out <- system (paste ("ollama pull", jina_model (what), intern = FALSE))
+    out <- system (paste ("ollama pull", jina_model (what)), intern = FALSE)
     return (out == 0)
 }
 
