### 2025-10-28 ea7abb5
fix 'interactive()' check for messages #203
