### 2025-10-28 20984d8
return direct result of 'system' call from 'dl_jina_model' for #203
