### 2025-10-28 53a1d0a
improve 'ollama_models' for #203
