### 2025-10-29 ropensci-review-tools/pkgmatch#203 @mpadge
Thanks @Selbosh, that was indeed a mess which I hadn't sufficiently debugged. All fixed now, and both messages and responses to all possible combinations of answers should be clear and sensible.
