### 2025-10-29 ropensci-review-tools/pkgsimil#210 @mpadge
Fix piggyback dependency
From @Selbosh's [review comment](https://github.com/ropensci/software-review/issues/671#issuecomment-3292032506):

> This package \["piggyback"\] appears in Suggests but not in Imports. Not a big deal but could confuse some users and maybe it should be in Imports or this error could be caught within the package?
