+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: pkgmatch
-Title:  Find R Packages Matching Either Descriptions or Other R Packages
-Version: 0.5.0.084
+Title: Find R Packages Matching Either Descriptions or Other R Packages
+Version: 0.5.0.085
 Authors@R: c(
     person("Mark", "Padgham", , "mark.padgham@email.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2172-5265")),
@@ -50,6 +50,6 @@ NeedsCompilation: yes
 Encoding: UTF-8
 Language: en-GB
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 Config/testthat/edition: 3
 VignetteBuilder: knitr
