+++ R/data-update.R
@@ -38,8 +38,6 @@ pkgmatch_update_data <- function (upload = TRUE,
                                   local_cran_mirror = NULL,
                                   local_ropensci_mirror = NULL) {
 
-    requireNamespace ("piggyback", quietly = TRUE)
-
     if (is.null (local_cran_mirror) && is.null (local_ropensci_mirror)) {
         results_path <-
             fs::dir_create (fs::path (fs::path_temp (), "pkgmatch-results"))
