### 2025-10-29 #671 @mpadge
@Selbosh Thanks again for your helpful comments.

- The difficulties with initial downloading of ollama models were something I hadn't really tested or debugged at all, so your insights there helped a lot. All fixed in https://github.com/ropensci-review-tools/pkgmatch/issues/203.
-  The `piggyback` dependency has been moved from _Suggests_ to _Imports_, as you suggested
- Finally, the information when downloading pre-computed embedding data has been updated and made more informative and verbose. It now looks like this:

```
This function requires data to be downloaded.
Data will be downloaded to `~/.cache/R/pkgmatch`.
This directory may be safely deleted at any time.
See the pkgmatch 'Data caching and updating' vignette for details.
Data for the ropensci corpus comprised 2 files totalling around 6.4MB.
Do you want to proceed (y/n)?
```
