### 2025-10-30 cd6c642
need to hard-code 8k context window for both models
