### 2025-10-30 d441e7f
update mock test results with new ollama calls with hard-coded 'num_ctx = 8k'
