### 2025-10-31 d3c49df
revert to default 'n_chunks = 1' for #212
