### 2025-10-31 45c1620
eval fns in null env
