### 2025-10-31 f2c9a72
don't 'eval' functions in parsing pkg code
