### 2025-11-01 5fad824
fix bug in cran-update code
