### 2025-11-01 0693ee4
update more mock test embeddings
