### 2025-11-05 bf97e5d
call 'send_dl_message' from update cache fn
To provide full information on dl location and how to manage cache
