### 2025-11-05 03f2329
update dockerfile and workflow for #221
