### 2025-11-06 e28bf92
fix docker workflow syntax [docker build]
