+++ .github/workflows/docker.yaml
@@ -16,7 +16,7 @@ jobs:
     if: |
       github.event_name == 'schedule' ||
       (github.event_name != 'schedule' &&
-      contains(github.event.head_commit.message, 'docker build')
+      contains(github.event.head_commit.message, 'docker build'))
 
     env:
       GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
