+++ .github/workflows/docker.yaml
@@ -30,7 +30,6 @@ jobs:
         uses: docker/setup-buildx-action@v3
 
       - name: Login to GitHub Container Registry
-        if: ${{ github.event_name != 'pull_request' }}
         uses: docker/login-action@v3
         with:
           registry: ghcr.io
@@ -42,5 +41,7 @@ jobs:
         with:
           context: .
           push: ${{ github.event_name != 'pull_request' }}
+          secrets: |
+              "GITHUB_PAT=${{ secrets.GITHUB_TOKEN }}"
           tags: |
             ghcr.io/ropensci-review-tools/pkgmatch:latest
