### 2025-11-06 7a63406
do not dl data in dockerfile for #221
