+++ Dockerfile
@@ -5,7 +5,7 @@ RUN apt-get update && apt-get install -y curl
 
 RUN --mount=type=secret,id=GITHUB_PAT,env=GITHUB_PAT installGithub.r \
     ropensci-review-tools/pkgmatch
-RUN Rscript -e 'pkgmatch::pkgmatch_update_cache()'
+# RUN Rscript -e 'pkgmatch::pkgmatch_update_cache()'
 
 RUN curl -fsSL https://ollama.com/install.sh | sh
 
