### 2025-11-06 65b95b4
separate data-raw/start.sh file for Docker build #221
The main dockerfile now just installs ollama but nothing else;
the script is the entrypoint which downloads models and calls
the update function
