### 2025-11-06 0df3ab0
rm start.sh; dl ollama models in container; docker build #221
