+++ Dockerfile
@@ -25,9 +25,11 @@ RUN --mount=type=secret,id=GITHUB_PAT,env=GITHUB_PAT installGithub.r \
 
 RUN curl -fsSL https://ollama.com/install.sh | sh
 
-COPY data-raw/start.sh /start.sh
-RUN chmod +x /start.sh
+RUN nohup bash -c "ollama serve &" \
+    && sleep 5 \
+    && ollama pull jina/jina-embeddings-v2-base-en \
+    && ollama pull unclemusclez/jina-embeddings-v2-base-code
 
 EXPOSE 11434
 
-ENTRYPOINT ["/start.sh"]
+ENTRYPOINT ["Rscript", "-e", "pkgmatch::pkgmatch_update_data()"]
