+++ .github/workflows/docker.yaml
@@ -23,26 +23,26 @@ jobs:
 
     steps:
 
-      - name: Checkout
-        uses: actions/checkout@v4
+    - name: Checkout
+      uses: actions/checkout@v4
 
-      - name: Buildx
-        uses: docker/setup-buildx-action@v3
+    - name: Buildx
+      uses: docker/setup-buildx-action@v3
 
-      - name: Login to GitHub Container Registry
-        uses: docker/login-action@v3
-        with:
-          registry: ghcr.io
-          username: ${{ github.repository_owner }}
-          password: ${{ secrets.GITHUB_TOKEN }}
+    - name: Login to GitHub Container Registry
+      uses: docker/login-action@v3
+      with:
+        registry: ghcr.io
+        username: ${{ github.repository_owner }}
+        password: ${{ secrets.GITHUB_TOKEN }}
 
-      - name: Build and push
-        uses: docker/build-push-action@v6
-        with:
-          context: .
-          push: ${{ github.event_name != 'pull_request' }}
-          secrets: |
-              "GITHUB_PAT=${{ secrets.GITHUB_TOKEN }}"
-          tags: |
-            mpadge/pkgcheck:latest
-            ghcr.io/ropensci-review-tools/pkgmatch:latest
+    - name: Build and push
+      uses: docker/build-push-action@v6
+      with:
+        context: .
+        push: ${{ github.event_name != 'pull_request' }}
+        secrets: |
+            "GITHUB_PAT=${{ secrets.GITHUB_TOKEN }}"
+        tags: |
+          mpadge/pkgcheck:latest
+          ghcr.io/ropensci-review-tools/pkgmatch:latest
