### 2025-11-07 d3d326b
add image for #221 to docker hub too [docker build]
