+++ .github/workflows/docker.yaml
@@ -29,6 +29,13 @@ jobs:
     - name: Buildx
       uses: docker/setup-buildx-action@v3
 
+    - name: Login to Dockerhub
+      if: ${{ github.event_name != 'pull_request' }}
+      uses: docker/login-action@v3
+      with:
+        username: ${{ secrets.DOCKER_USERNAME }}
+        password: ${{ secrets.DOCKER_PASSWORD }}
+
     - name: Login to GitHub Container Registry
       uses: docker/login-action@v3
       with:
