### 2025-11-07 3e82227
add missing 'requireNamespace' for hms
