### 2025-11-07 fac25ec
docker build with explicit suggests pkgs in dockerfile
Because these are otherwise not installed by installGithub.r r2u fn
