### 2025-11-07 #671 @mcsiple
Hi all -- I am writing from my academic GitHub here so I can follow this thread. FYI I am still keeping tabs on this review despite the shutdown.
