### 2025-11-11 3edf00c
add bioc corpus to cache fns for #238
