### 2025-11-11 ee44376
extend 'check_corpus_param' to include bioc #238, and update man entry
