+++ man/pkgmatch_similar_pkgs.Rd
@@ -19,10 +19,10 @@ pkgmatch_similar_pkgs(
 \item{input}{Either a text string, a path to local source code of an R
 package, or the name of any installed R package.}
 
-\item{corpus}{Must be specified as one of "ropensci" or "cran". If
-\code{embeddings} or \code{idfs} parameters are not specified, they will be
-automatically downloaded for the corpus specified by this parameter. The
-function will then return the most similar package from the specified
+\item{corpus}{Must be specified as one of "ropensci", "cran", or "bioc" (for
+BioConductor). If \code{embeddings} or \code{idfs} parameters are not specified, they
+will be automatically downloaded for the corpus specified by this parameter.
+The function will then return the most similar package from the specified
 corpus. Note that calculations will \code{corpus = "cran"} will generally take
 longer, because the corpus is much larger.}
 
