### 2025-11-11 73e0869
update more corpus docs and params to include bioc #238
