+++ man/pkgmatch_update_data.Rd
@@ -33,6 +33,10 @@ by any unauthorized users will error and have no effect unless run with
 sub-directory "pkgmatch-results" of R's current temporary directory. This
 updating may take a very long time!
 
+The function does not update the BioConductor data. Because those are fixed
+to specific BioConductor releases, they are only updated manually with the
+\link{pkgmatch_generate_bioc} function.
+
 Note that this function is categorically different from
 \link{pkgmatch_update_cache}. This function updates the internal data used
 by the \code{pkgmatch} package, and should only ever be run by package
