### 2025-11-11 62ae91d
can't link to internal 'generate_bioc' fn in docs
