### 2025-11-11 8b42635
improve messages from 'similar_pkgs' fn when cached data are up to date
