### 2025-11-11 29d91bb
fix structure of cached bm25-bioc for #238
