### 2025-11-11 e8da250
mention bioc corpus param in readme
