+++ R/RcppExports.R
@@ -1,6 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-rcpp_bm25 <- function (idfs, tokensList, these_tokens, ntoks_avg) {
-    .Call (`_pkgmatch_rcpp_bm25`, idfs, tokensList, these_tokens, ntoks_avg)
+rcpp_bm25 <- function(idfs, tokensList, these_tokens, ntoks_avg) {
+    .Call(`_pkgmatch_rcpp_bm25`, idfs, tokensList, these_tokens, ntoks_avg)
 }
