### 2025-11-25 84822a8
copy edits to vignette A
