### 2025-12-16 ea5fc6d
minor copy edits to vignette C
