+++ vignettes/C_how-does-it-work.Rmd
@@ -14,13 +14,13 @@ knitr::opts_chunk$set (
 )
 ```
 
-The "pkgmatch" package finds packages, as well as individual functions,
+The `pkgmatch` package finds packages, as well as individual functions,
 which best match a given input. Inputs can be text descriptions, sections of
-code, or even entire R packages. "pkgmatch" finds the most closely matching
+code, or even entire R packages. `pkgmatch` finds the most closely matching
 packages using a combination of Language Models (LMs, or equivalently, "LLMs"
 for large language models), and traditional [token-frequency
 algorithms](https://en.wikipedia.org/wiki/Okapi_BM25). This document describes
-how the "pkgmatch" package works.
+how the `pkgmatch` package works.
 
 ## Summary
 
@@ -48,7 +48,7 @@ should generally be very unlikely.
 
 Text input of packages includes text from the package description (in the
 "DESCRIPTION" file), followed by text from the "README" file (where present), text of all
-vignettes, and text of all function descriptions, and parameters of each
+vignettes, text of all function descriptions, and parameters of each
 function. All code is removed from these text sources, including removal of all
 code chunks in `.Rmd` documents.
 
@@ -86,7 +86,7 @@ similarities between inputs and all packages within these corpora.
 
 ### LM similarities
 
-"pkgmatch" accesses LMs through a local [ollama server](https://ollama.com), as
+`pkgmatch` accesses LMs through a local [ollama server](https://ollama.com), as
 described in the [ollama installation vignette](https://docs.ropensci.org/pkgmatch/articles/B_ollama.html).
 LMs are exclusively used here to generate "embeddings", which are numeric
 vectors of a fixed length representing a (reduced) version of the
@@ -102,14 +102,14 @@ in LMs, because similarities in *orientation* of two (or more) embedding
 vectors are generally more informative than similarities in magnitude.
 
 For a given input and corpus, the output of this LM component specifies a
-cosine similarities for each package, along with a corresponding rank, where
+cosine similarity for each package, along with a corresponding rank, where
 the first-ranked package is the most similar.
 
 ### Token-frequency similarities
 
 Similarities from LMs alone are often not reliable, and better results are
 often generated through combining LM results with equivalent outputs from
-alternative algorithms. The "pkgmatch" package also passes all input to an
+alternative algorithms. The `pkgmatch` package also passes all input to an
 internal version of the ["Best Match 25" (BM25)
 algorithm](https://en.wikipedia.org/wiki/Okapi_BM25), an algorithm that is
 commonly used in conjunction with LM outputs. Like LM cosine similarities,
@@ -128,7 +128,7 @@ time on, always available for immediate use in all subsequent calculations.
 #### Tokens for text input
 
 For text input, the BM25 algorithm requires inputs to be converted to "tokens",
-for which this package relies on [rOpenSci's tokenizer
+for which this package relies on [rOpenSci's `tokenizers`
 package](https://docs.ropensci.org/tokenizers/). The BM25 values then assess
 similarities through adding similarities for all tokens shared between two
 documents, weighted by the inverse of token frequencies across all packages
@@ -141,7 +141,7 @@ common tokens.
 For code input, tokens for BM25 input are taken as the names of all functions
 called, pre-pended with namespaces, such as `base::print()`. Function call are
 identified from the "treesitter" algorithm, using functionality of the
-["treesitter" R package](https://davisvaughan.github.io/r-tree-sitter/). This
+[`treesitter` R package](https://davisvaughan.github.io/r-tree-sitter/). This
 package is currently restricted to [the R language
 only](https://github.com/r-lib/tree-sitter-r), and thus code input in this
 package considers code from `/R` sub-directories only, and excludes any source
