### 2025-12-18 7edcacc
use r-univ api instead of roregistry to close #247 [ci skip]
