### 2025-12-29 6e748df
uniform formatting for packages mentioned in vignettes
