### 2025-12-29 edac191
copy edits to vignettes A and E
