+++ vignettes/A_extended-use-case.Rmd
@@ -157,8 +157,8 @@ search engines or language models, and `pkgmatch` results which are highly
 specific because they are restricted to R packages only. Although we expected
 no precise match for reasons described above, `pkgmatch` nevertheless revealed
 the presence of two very clearly related packages,
-['RWsearch'](https://cran.r-project.org/package=RWsearch) and
-['sos'](https://github.com/sbgraves237/sos), neither of which were returned in
+[`RWsearch`](https://cran.r-project.org/package=RWsearch) and
+[`sos`](https://github.com/sbgraves237/sos), neither of which were returned in
 results from either [search
 engines](https://duckduckgo.com/?q=R%20package%20to%20return%20web%20search%20engine%20results%20into%20R%20as%20strings%20or%20URLs%20r%20programming),
 or large language models ([Phind](https://www.phind.com/search/cmbqhybqt0000206imod5zpqo), [Perplexity](https://www.perplexity.ai/search/r-package-to-return-web-search-9mTr_z_eT2iqkKutaU3YYA)).
