### 2025-12-29 60e7ffa
copy edits to vignette F
