+++ README.md
@@ -2,6 +2,7 @@
 
 [![R build
 status](https://github.com/ropensci-review-tools/pkgmatch/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci-review-tools/pkgmatch/actions?query=workflow%3AR-CMD-check)
+[![](https://badges.ropensci.org/671_status.svg)](https://github.com/ropensci/software-review/issues/671)
 [![codecov](https://codecov.io/gh/ropensci-review-tools/pkgmatch/branch/main/graph/badge.svg)](https://app.codecov.io/gh/ropensci-review-tools/pkgmatch)
 [![Project Status:
 Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -27,8 +28,8 @@ functions require ollama to be installed.
 
 Once ollama is running, the easiest way to install this package is via
 the [associated
-`r-universe`](https://ropensci-review-tools.r-universe.dev/).
-As shown there, simply enable the universe with
+`r-universe`](https://ropensci-review-tools.r-universe.dev/). As shown
+there, simply enable the universe with
 
 ``` r
 options (repos = c (
@@ -74,14 +75,16 @@ matching algorithms.
 The package has two main functions:
 
 - [`pkgmatch_similar_pkgs()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
-to find similar [rOpenSci](https://ropensci.org/packages/),
-[CRAN](https://cran.r-project.org), or
-[BioConductor](https://bioconductor.org/packages/release/BiocViews.html#___Software)
-packages based on input as either a local path to an entire package, the name
-of an installed package, or as a single descriptive text string; and
+  to find similar [rOpenSci](https://ropensci.org/packages/),
+  [CRAN](https://cran.r-project.org), or
+  [BioConductor](https://bioconductor.org/packages/release/BiocViews.html#___Software)
+  packages based on input as either a local path to an entire package,
+  the name of an installed package, or as a single descriptive text
+  string; and
 - [`pkgmatch_similar_fns()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_fns.html)
-to find similar functions from rOpenSci or BioConductor packages based on
-descriptive text input. (Not available for functions from CRAN packages.)
+  to find similar functions from rOpenSci or BioConductor packages based
+  on descriptive text input. (Not available for functions from CRAN
+  packages.)
 
 The following code demonstrates how these functions work, first matching
 general text strings packages from rOpenSci:
@@ -94,12 +97,14 @@ on visualising inter-relationships among distinct trees.
 pkgmatch_similar_pkgs (input, corpus = "ropensci")
 ```
 
-    ## [1] "phylogram"    "phruta"       "rotl"         "taxa"         "lingtypology"
+    ## [1] "unifir"          "waywiser"        "medrxivr"        "tidyqpcr"       
+    ## [5] "googleLanguageR"
 
-The corpus parameter must be specified as one of "ropensci", "cran", or "bioc"
-(for [BioConductor](https://bioconductor.org); all case-insensitive). The CRAN
-corpus is much larger than the rOpenSci or BioConductor corpora, and matching
-for `corpus = "cran"` will generally take notably longer.
+The corpus parameter must be specified as one of “ropensci”, “cran”, or
+“bioc” (for [BioConductor](https://bioconductor.org); all
[…]
