+++ README.Rmd
@@ -61,6 +61,8 @@ and running one of the following lines:
 ```{r remotes, eval = FALSE}
 remotes::install_github ("ropensci-review-tools/pkgmatch")
 pak::pkg_install ("ropensci-review-tools/pkgmatch")
+remotes::install_git ("https://codeberg.org/ropensci-review-tools/pkgmatch")
+remotes::install_git ("https://codefloe.com/ropensci-review-tools/pkgmatch")
 ```
 
 The package can then loaded for use with
