### 2026-03-04 40a527b
rm ollama from similar_pkgs for #254
