+++ R/similar-pkgs.R
@@ -6,13 +6,9 @@
 #' specify either [rOpenSci's package suite](https://ropensci.org/packages), or
 #' [CRAN](https://cran.r-project.org).
 #'
-#' Ranks are obtained from scores derived from:
-#' \itemize{
-#' \item Cosine similarities between Language Model (LM) embeddings for the
-#' `input`, and corresponding embeddings for the specified corpus. \item ["Best
-#' Match 25" (BM25)](https://en.wikipedia.org/wiki/Okapi_BM25) scores based on
-#' document token frequencies.
-#' }
+#' Ranks are obtained from scores derived from ["Best Match 25"
+#' (BM25)](https://en.wikipedia.org/wiki/Okapi_BM25) scores based on document
+#' token frequencies.
 #'
 #' For text input, ranks are generally obtained for packages both including and
 #' excluding function descriptions as part of the package text, giving two sets
@@ -33,27 +29,17 @@
 #' @param input Either a text string, a path to local source code of an R
 #' package, or the name of any installed R package.
 #' @param corpus Must be specified as one of "ropensci", "cran", or "bioc" (for
-#' BioConductor). If `embeddings` or `idfs` parameters are not specified, they
-#' will be automatically downloaded for the corpus specified by this parameter.
-#' The function will then return the most similar package from the specified
+#' BioConductor). If `idfs` parameter is not specified, data will be
+#' automatically downloaded for the corpus specified by this parameter. The
+#' function will then return the most similar package from the specified
 #' corpus. Note that calculations will `corpus = "cran"` will generally take
 #' longer, because the corpus is much larger.
-#' @param embeddings Large Language Model embeddings for a suite of packages,
-#' generated from \link{pkgmatch_embeddings_from_pkgs}. If not provided,
-#' pre-generated embeddings will be downloaded and stored in a local cache
-#' directory.
 #' @param idfs Inverse Document Frequency tables for a suite of packages,
 #' generated from \link{pkgmatch_bm25}. If not provided, pre-generated IDF
 #' tables will be downloaded and stored in a local cache directory.
 #' @param input_is_code A binary flag indicating whether `input` is code or
 #' plain text. Ignored if `input` is path to a local package; otherwise can be
 #' used to force appropriate interpretation of input type.
-#' @param lm_proportion A value between 0 and 1 to control the relative
-#' contributions of results from Language Models ("LMs") versus results from
-#' traditional token-frequency models. Final rankings are generated by
-#' combining these two kinds of results, so that `lm_proportion = 0` will
-#' return results from token frequency analyses only, while `lm_proportion =
-#' 1` will return results from LMs only.
 #' @param n When the result of this function is printed to screen, the top `n`
 #' packages will be displayed.
 #' @param browse If `TRUE`, automatically open webpages of the top `n` matches
@@ -72,10 +58,10 @@
 #' To see all data, use `as.data.frame()`. See the example below for how to
 #' manipulate these objects.
 #'
-#' @note The first time this function is run without passing either
-#' `embeddings` or `idfs`, required values will be automatically downloaded and
-#' stored in a locally persistent cache directory. Especially for the "cran"
-#' corpus, this downloading may take quite some time.
+#' @note The first time this function is run without passing `idfs`, required
+#' values will be automatically downloaded and stored in a locally persistent
[…]
