+++ man/pkgmatch_similar_pkgs.Rd
@@ -7,10 +7,8 @@
 pkgmatch_similar_pkgs(
   input,
   corpus = NULL,
-  embeddings = NULL,
   idfs = NULL,
   input_is_code = text_is_code(input),
-  lm_proportion = 0.5,
   n = 5L,
   browse = FALSE
 )
@@ -20,16 +18,11 @@ pkgmatch_similar_pkgs(
 package, or the name of any installed R package.}
 
 \item{corpus}{Must be specified as one of "ropensci", "cran", or "bioc" (for
-BioConductor). If \code{embeddings} or \code{idfs} parameters are not specified, they
-will be automatically downloaded for the corpus specified by this parameter.
-The function will then return the most similar package from the specified
+BioConductor). If \code{idfs} parameter is not specified, data will be
+automatically downloaded for the corpus specified by this parameter. The
+function will then return the most similar package from the specified
 corpus. Note that calculations will \code{corpus = "cran"} will generally take
 longer, because the corpus is much larger.}
-
-\item{embeddings}{Large Language Model embeddings for a suite of packages,
-generated from \link{pkgmatch_embeddings_from_pkgs}. If not provided,
-pre-generated embeddings will be downloaded and stored in a local cache
-directory.}
 
 \item{idfs}{Inverse Document Frequency tables for a suite of packages,
 generated from \link{pkgmatch_bm25}. If not provided, pre-generated IDF
@@ -38,12 +31,6 @@ tables will be downloaded and stored in a local cache directory.}
 \item{input_is_code}{A binary flag indicating whether \code{input} is code or
 plain text. Ignored if \code{input} is path to a local package; otherwise can be
 used to force appropriate interpretation of input type.}
-
-\item{lm_proportion}{A value between 0 and 1 to control the relative
-contributions of results from Language Models ("LMs") versus results from
-traditional token-frequency models. Final rankings are generated by
-combining these two kinds of results, so that \code{lm_proportion = 0} will
-return results from token frequency analyses only, while \code{lm_proportion = 1} will return results from LMs only.}
 
 \item{n}{When the result of this function is printed to screen, the top \code{n}
 packages will be displayed.}
@@ -72,12 +59,8 @@ corpus in terms of how well they match that input. The "corpus" argument can
 specify either \href{https://ropensci.org/packages}{rOpenSci's package suite}, or
 \href{https://cran.r-project.org}{CRAN}.
 
-Ranks are obtained from scores derived from:
-\itemize{
-\item Cosine similarities between Language Model (LM) embeddings for the
-\code{input}, and corresponding embeddings for the specified corpus. \item \href{https://en.wikipedia.org/wiki/Okapi_BM25}{"Best Match 25" (BM25)} scores based on
-document token frequencies.
-}
+Ranks are obtained from scores derived from \href{https://en.wikipedia.org/wiki/Okapi_BM25}{"Best Match 25" (BM25)} scores based on document
+token frequencies.
 
 For text input, ranks are generally obtained for packages both including and
 excluding function descriptions as part of the package text, giving two sets
@@ -95,10 +78,10 @@ means the effect of changing \code{lm_proportion} can easily be examined by
 simply repeating calls to this function.
 }
 \note{
-The first time this function is run without passing either
-\code{embeddings} or \code{idfs}, required values will be automatically downloaded and
-stored in a locally persistent cache directory. Especially for the "cran"
-corpus, this downloading may take quite some time.
+The first time this function is run without passing \code{idfs}, required
+values will be automatically downloaded and stored in a locally persistent
[…]
