+++ man/pkgmatch_similar_fns.Rd
@@ -9,11 +9,6 @@ pkgmatch_similar_fns(input, embeddings = NULL, n = 5L, browse = FALSE)
 \arguments{
 \item{input}{A text string.}
 
-\item{embeddings}{Large Language Model embeddings for a suite of packages,
-generated from \link{pkgmatch_embeddings_from_pkgs}. If not provided,
-pre-generated embeddings will be downloaded and stored in a local cache
-directory.}
-
 \item{n}{When the result of this function is printed to screen, the top \code{n}
 packages will be displayed.}
 

[…]
