### 2026-03-04 5a6aa02
rm ollama from similar_pkgs_from_text for #254
