### 2026-03-04 9afa4c9
entirely rm 'similar-fns.R' for #254
Similar fn matching was entirely llm-based, so now completely removed.
