### 2026-03-04 f1f503d
rm simmilarity-metrics.R for #254
The main similarity_embeddings and cosine_similrity functions
