### 2026-03-04 6927a2a
rm embeddings entirely from cache fns for #254
