+++ R/cache.R
@@ -2,7 +2,6 @@
 #'
 #' @description Load pre-computed data for a specified corpus. Data types are:
 #' \itemize{
-#' \item "embeddings" for language model embeddings;
 #' \item "idfs" for Inverse Document Frequency weightings;
 #' \item "functions" for frequency tables for text descriptions of function
 #' calls; or
@@ -15,10 +14,10 @@
 #' called.
 #'
 #' @inheritParams pkgmatch_similar_pkgs
-#' @param what One of the four data types described above: "embeddings",
-#' "idfs", "functions", or "calls".
-#' @param fns If `FALSE` (default), load embeddings for all packages; otherwise
-#' load (considerably larger dataset of) embeddings for all individual
+#' @param what One of the three data types described above: "idfs",
+#' "functions", or "calls".
+#' @param fns If `FALSE` (default), load data for all packages; otherwise
+#' load (considerably larger dataset of) data for all individual
 #' functions.
 #' @param raw Only has effect of `what = "calls"`, in which case default of
 #' `FALSE` loads single Inverse Document Frequency table to entire corpus;
@@ -31,12 +30,10 @@
 #'
 #' @examples
 #' \dontrun{
-#' embeddings <- pkgmatch_load_data ("embeddings")
-#' embeddings_fns <- pkgmatch_load_data ("embeddings", fns = TRUE)
 #' idfs <- pkgmatch_load_data ("idfs")
 #' idfs_fns <- pkgmatch_load_data ("idfs", fns = TRUE)
 #' }
-pkgmatch_load_data <- function (what = "embeddings",
+pkgmatch_load_data <- function (what = "idfs",
                                 corpus = "ropensci",
                                 fns = FALSE,
                                 raw = FALSE) {
@@ -92,7 +89,7 @@ pkgmatch_cache_update_interval <- function () {
 #' }
 pkgmatch_update_cache <- function () {
 
-    what <- c ("embeddings", "idfs", "functions", "calls")
+    what <- c ("idfs", "functions", "calls")
     corpus <- c ("ropensci", "cran", "bioc")
     fns <- c (FALSE, TRUE)
     raw <- c (FALSE, TRUE)
@@ -167,9 +164,6 @@ m_list_remote_files <- function () {
             "bm25-cran.Rds",
             "bm25-ropensci-fns.Rds",
             "bm25-ropensci.Rds",
-            "embeddings-cran.Rds",
-            "embeddings-fns.Rds",
-            "embeddings-ropensci.Rds",
             "fn-calls-cran.Rds",
             "fn-calls-ropensci.Rds",
             "idfs-fn-calls-cran.Rds",
@@ -191,14 +185,11 @@ list_remote_files <- memoise::memoise (m_list_remote_files)
 get_cache_file_name <- function (what, corpus, fns, raw) {
 
     corpus <- match.arg (tolower (corpus), c ("ropensci", "cran", "bioc"))
-    what <- match.arg (what, c ("embeddings", "idfs", "functions", "calls"))
+    what <- match.arg (what, c ("idfs", "functions", "calls"))
 
     if (corpus == "ropensci") {
 
         fname <- switch (what,
-            "embeddings" = ifelse (
-                fns, "embeddings-fns.Rds", "embeddings-ropensci.Rds"
-            ),
             "idfs" = ifelse (fns, "bm25-ropensci-fns.Rds", "bm25-ropensci.Rds"),
             "functions" = "fn-calls-ropensci.Rds",
             "calls" = ifelse (
@@ -209,7 +200,6 @@ get_cache_file_name <- function (what, corpus, fns, raw) {
     } else if (corpus == "cran") {
 
[…]
