+++ NAMESPACE
@@ -11,7 +11,6 @@ export(pkgmatch_browse)
 export(pkgmatch_embeddings_from_pkgs)
 export(pkgmatch_embeddings_from_text)
 export(pkgmatch_load_data)
-export(pkgmatch_similar_fns)
 export(pkgmatch_similar_pkgs)
 export(pkgmatch_treesitter_fn_tags)
 export(pkgmatch_update_cache)
