### 2026-03-04 ab64f0a
rm embeddings from all 'data-update-' fns for #254
