### 2026-03-04 ad00b0f
rm all embedding-related fns from 'embeddings.R' for #254
That leaves a few generic functions which will have to be moved elsewhere
