+++ R/embeddings.R
@@ -21,238 +21,6 @@ convert_paths_to_pkgs <- function (packages) {
     return (packages)
 }
 
-#' @title Return raw embeddings from package text and function definitions.
-#'
-#' @description This function accepts a vector of either names of installed
-#' packages, or paths to local source code directories, and calculates language
-#' model (LM) embeddings for both text descriptions within the package
-#' (documentation, including of functions), and for the entire code base.
-#' Embeddings may also be calculating separately for all function descriptions.
-#'
-#' The embeddings are currently retrieved from a local 'ollama' server
-#' (\url{https://ollama.com}) running Jina AI embeddings
-#' (\url{https://ollama.com/jina/jina-embeddings-v2-base-en} for text, and
-#' \url{https://ollama.com/unclemusclez/jina-embeddings-v2-base-code} for code).
-#'
-#' @note Although it is technically much faster to perform the extraction of
-#' text and code in parallel, doing so generates unpredictable errors in
-#' extracting tarballs, which frequently cause the whole process to crash. The
-#' only way to safely ensure that all tarballs are successfully extracted and
-#' code parsed it to run this single-threaded.
-#'
-#' @param packages A vector of either names of installed packages, or local
-#' paths to directories containing R packages.
-#' @param n_chunks Number of randomly permuted chunks of input text to use to
-#' generate average embeddings. Issue #212 demonstrates that > 95% of all CRAN packages
-#' have <= 8192 tokens, and so fit within the context windows of the embeddings
-#' models used here. Chunking is therefore generally not necessary, and so this
-#' parameter defaults to a value of 1. Also note that computation
-#' times scale linearly with this value.
-#' @param functions_only If `TRUE`, calculate embeddings for function
-#' descriptions only. This is intended to generate a separate set of embeddings
-#' which can then be used to match plain-text queries of functions, rather than
-#' entire packages.
-#' @return If `!functions_only`, a list of two matrices of embeddings: one for
-#' the text descriptions of the specified packages, including individual
-#' descriptions of all package functions, and one for the entire code base. For
-#' `functions_only`, a single matrix of embeddings for all function
-#' descriptions.
-#'
-#' @family embeddings
-#' @export
-#'
-#' @examples
-#' packages <- "curl"
-#' emb_fns <- pkgmatch_embeddings_from_pkgs (packages, functions_only = TRUE)
-#' colnames (emb_fns) # All functions the package
-#' emb_pkg <- pkgmatch_embeddings_from_pkgs (packages, functions_only = FALSE)
-#' names (emb_pkg)
-#' colnames (emb_pkg$text_with_fns) # "curl"
-pkgmatch_embeddings_from_pkgs <- function (packages = NULL,
-                                           n_chunks = 1L,
-                                           functions_only = FALSE) {
-
-    checkmate::assert_integer (n_chunks, len = 1L)
-    if (identical (Sys.getenv ("PKGMATCH_TESTS"), "true")) {
-        n_chunks <- 1L
-    }
-    chunk_seq <- seq_len (n_chunks)
-    checkmate::assert_logical (functions_only, len = 1L)
-    checkmate::assert_character (packages)
-    if (all (grepl ("\\.tar\\.gz$", packages))) {
-        checkmate::assert_file_exists (packages)
-    } else {
-        if (!all (pkg_is_installed (packages))) {
-            checkmate::assert_directory_exists (packages)
-        }
-    }
-
-    pkgs_full <- packages
-    packages <- convert_paths_to_pkgs (pkgs_full)
[…]
