### 2026-03-04 f289c27
rm R/ollama for #254
