+++ R/ollama.R
@@ -1,266 +1,0 @@
-# Functions to check ollama status
-
-is_windows <- function () {
-    grepl ("windows", Sys.info () ["sysname"], ignore.case = TRUE)
-}
-
-is_docker_sudo <- function () {
-    if (is_windows ()) {
-        return (FALSE)
-    }
-    cmd <- "ps aux | grep -v grep | grep dockerd"
-    suppressWarnings (
-        out <- tryCatch (
-            system (cmd, intern = TRUE, ignore.stderr = TRUE),
-            error = function (e) NULL
-        )
-    )
-    chk <- any (grepl ("root\\s", out))
-    if (chk) {
-        # Retain 'TRUE' only if current user in not in "docker" group:
-        cmd <- "grep /etc/group -e 'docker'"
-        suppressWarnings (
-            out <- tryCatch (
-                system (cmd, intern = TRUE, ignore.stderr = TRUE),
-                error = function (e) NULL
-            )
-        )
-        if (!is.null (out)) {
-            user_docker <- gsub ("^.*\\:", "", out)
-            suppressWarnings (
-                user_current <- tryCatch (
-                    system ("echo $USER", intern = TRUE, ignore.stderr = TRUE),
-                    error = function (e) NULL
-                )
-            )
-            if (user_current == user_docker) {
-                chk <- FALSE
-            }
-        }
-    }
-    return (chk)
-}
-
-has_ollama <- function (sudo = is_docker_sudo ()) {
-    has_ollama_local () || has_ollama_docker (sudo = sudo)
-}
-
-has_ollama_local <- function () {
-    lib_name <- "ollama"
-    cmd <- ifelse (is_windows (), "where", "which")
-    result <- system (
-        paste (cmd, lib_name),
-        ignore.stdout = TRUE,
-        ignore.stderr = TRUE
-    )
-
-    return (result == 0)
-}
-
-has_ollama_docker <- function (sudo = is_docker_sudo ()) {
-    cmd <- "docker ps -a"
-    if (sudo) {
-        cmd <- paste ("sudo", cmd)
-    }
-    suppressWarnings (
-        out <- tryCatch (
-            system (cmd, intern = TRUE, ignore.stderr = TRUE),
-            error = function (e) NULL
-        )
-    )
-    chk <- !is.null (out)
-    if (chk) {
-        chk <- any (grepl ("ollama-models", out))
-    }
-    return (chk)
-}
-
-ollama_models <- function () {
-    stopifnot (ollama_is_running ())
-
-    out <- system ("ollama list", intern = TRUE)
-    nms <- strsplit (out [1], "(\\s|\\t)+") [[1]]
-
-    if (length (out) == 1L) { # no models installed
-        out <- data.frame (array (dim = c (0L, length (nms))))
-        names (out) <- nms
-        out$version <- logical (0L)
-        return (out)
-    }
-
-    # results of 'ollama list' are left-aligned as:
-    # - NAME: Name of model, generally as "org/model:latest", and always without spaces
-    # - ID: hash, always without spaces
-    # - SIZE: "[0-9]+ MB"
-    # - MODIFIED: "[0-9]+ hours/days/weeks ago"
-    pos_start <- vapply (
-        nms,
-        function (n) regexpr (n, out [1]),
-        integer (1L),
-        USE.NAMES = FALSE
-    )
-    pos_end <- pos_start [-1] - 1L
-
-    out <- out [-1]
-
-    out <- lapply (out, function (i) {
-        pos <- cbind (pos_start, c (pos_end, nchar (i)))
-        unname (apply (pos, 1, function (j) {
-            gsub ("\\s+$", "", substring (i, j [1], j [2]))
[…]
