### 2026-03-04 529ce8d
update docs without ollama fns for #254
