+++ R/zzz.R
@@ -4,7 +4,6 @@
     op <- options ()
 
     op.pkgmatch <- list ( # nolint
-        pkgmatch.ollama.url = Sys.getenv ("OLLAMA_HOST", "127.0.0.1:11434"),
         pkgmatch.verbose_limit = 50L
     )
 
@@ -20,17 +19,6 @@
     }
 
     invisible ()
-}
-
-.onAttach <- function (libname, pkgname) {
-
-    chk <- tryCatch (
-        ollama_check (),
-        error = function (e) e
-    )
-    if (inherits (chk, "error")) {
-        packageStartupMessage (chk$message)
-    }
 }
 # nocov end
 
@@ -55,75 +43,4 @@ opt_is_quiet <- function () {
     getOption ("rlib_message_verbosity", "notset") == "quiet" ||
         (Sys.getenv ("GITHUB_ACTIONS") == "true" &&
             Sys.getenv ("GITHUB_JOB") != "test-coverage.yaml")
-}
-
-#' @title Get the URL for local ollama API
-#'
-#' @description Return the URL of the specified ollama API. Default is
-#' "127.0.0.1:11434"
-#'
-#' @return The ollama API URL
-#'
-#' @seealso `set_ollama_url`
-#'
-#' @family ollama
-#' @examples
-#' u <- get_ollama_url ()
-#'
-#' @export
-get_ollama_url <- function () {
-
-    # Need to return fixed value in tests, because tests create
-    # sub-environments which do not inherit withr::local_env:
-    if (identical (Sys.getenv ("PKGMATCH_TESTS"), "true")) {
-        return ("127.0.0.1:11434")
-    }
-
-    op <- options ()
-    if (!"pkgmatch.ollama.url" %in% names (op)) {
-        stop ("ollama URL can not be retrieved")
-    }
-    getOption ("pkgmatch.ollama.url")
-}
-
-
-#' @title Set the URL for local ollama API
-#'
-#' @param ollama_url The desired ollama API URL
-#'
-#' @return The ollama API URL
-#'
-#' @seealso [get_ollama_url()]
-#'
-#' @family ollama
-#' @examples
-#' u <- "127.0.0.1:11434"
-#' set_ollama_url (u)
-#' get_ollama_url ()
-#' @export
-set_ollama_url <- function (ollama_url) {
-
-    check_ollama_url (ollama_url)
-
-    op <- options ()
-    op.pkgmatch <- list (pkgmatch.ollama.url = ollama_url)
-    options (op.pkgmatch)
-
-    invisible (ollama_url)
-}
-
-check_ollama_url <- function (ollama_url) {
-
-    ollama_url <- ifelse (
-        grepl ("^http", ollama_url),
-        ollama_url,
-        paste0 ("https://", ollama_url)
-    )
-    check <- tryCatch (
-        curl::curl_parse_url (ollama_url),
-        error = function (e) e
-    )
-    if (methods::is (check, "error")) {
-        cli::cli_abort (check$message)
-    }
 }
