+++ man/pkgmatch_embeddings_from_text.Rd
@@ -1,36 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/embeddings.R
-\name{pkgmatch_embeddings_from_text}
-\alias{pkgmatch_embeddings_from_text}
-\title{Return raw embeddings from a vector of text strings.}
-\usage{
-pkgmatch_embeddings_from_text(input = NULL)
-}
-\arguments{
-\item{input}{A vector of one or more text strings for which embeddings are
-to be extracted.}
-}
-\value{
-A matrix of embeddings, one column for each \code{input} item, and a
-fixed number of rows defined by the embedding length of the language models.
-}
-\description{
-This function accepts a vector of character strings,
-packages, or paths to local source code directories, and calculates language
-model (LM) embeddings for each string within the vector.
-
-The embeddings are currently retrieved from a local 'ollama' server
-(\url{https://ollama.com}) running Jina AI text embeddings
-(\url{https://ollama.com/jina/jina-embeddings-v2-base-en}).
-}
-\examples{
-\dontrun{
-input <- "Download open spatial data from NASA"
-emb <- pkgmatch_embeddings_from_text (input = input)
-}
-}
-\seealso{
-Other embeddings: 
-\code{\link{pkgmatch_embeddings_from_pkgs}()}
-}
-\concept{embeddings}
