+++ NAMESPACE
@@ -3,19 +3,14 @@
 S3method(head,pkgmatch)
 S3method(print,pkgmatch)
 export(generate_pkgmatch_example_data)
-export(get_ollama_url)
-export(ollama_check)
 export(pkgmatch_bm25)
 export(pkgmatch_bm25_fn_calls)
 export(pkgmatch_browse)
-export(pkgmatch_embeddings_from_pkgs)
-export(pkgmatch_embeddings_from_text)
 export(pkgmatch_load_data)
 export(pkgmatch_similar_pkgs)
 export(pkgmatch_treesitter_fn_tags)
 export(pkgmatch_update_cache)
 export(pkgmatch_update_data)
-export(set_ollama_url)
 export(text_is_code)
 importFrom(Rcpp,sourceCpp)
 importFrom(memoise,memoise)
