### 2026-03-04 7c0c024
rm embeddings from tests; update test-data-update #254
