### 2026-03-04 7818d26
rm embeddings from 'test-similar' for #254
