+++ tests/testthat/test-similar.R
@@ -1,27 +1,15 @@
 test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") ||
     identical (Sys.getenv ("GITHUB_JOB"), "test-coverage"))
 
-skip_if (ollama_is_running ())
-
 test_that ("similar pkgs text input", {
 
     withr::local_envvar (list (
-        "PKGMATCH_TESTS" = "true",
-        "pkgmatch.ollama.url" = "127.0.0.1:11434"
+        "PKGMATCH_TESTS" = "true"
     ))
 
     input <- "A similar package"
     n <- 5L
-    embeddings <- get_test_embeddings (
-        npkgs = 10,
-        nfns = 10,
-        embedding_len = 768
-    )
-    txt <- c (
-        "a not so very similar package",
-        "a test package",
-        "a function to test"
-    )
+    txt <- get_sample_input_text ()
     idfs <- get_test_idfs (txt)
 
     expect_error (
@@ -29,18 +17,15 @@ test_that ("similar pkgs text input", {
         "'corpus' must be specified."
     )
 
-    out <- httptest2::with_mock_dir ("sim_pkgs_txt", {
-        pkgmatch_similar_pkgs (
-            input,
-            corpus = "ropensci",
-            embeddings = embeddings,
-            idfs = idfs,
-            n = n
-        )
-    })
+    out <- pkgmatch_similar_pkgs (
+        input,
+        corpus = "ropensci",
+        idfs = idfs,
+        n = n
+    )
     expect_s3_class (out, "pkgmatch")
     expect_type (out, "list")
-    expect_true (all (out$package %in% colnames (embeddings$text_with_fns)))
+    expect_true (all (out$package %in% names (idfs$token_lists$with_fns)))
     expect_equal (attr (out, "n"), n)
 
     # print method:
@@ -67,93 +52,28 @@ test_that ("similar pkgs text input", {
     expect_equal (as.integer (row1), seq_along (row1))
 })
 
-# without ollama fails on windows, because it reverts to key input for
-# downloading
-skip_on_os ("windows")
-
-test_that ("without ollama", {
-
-    input <- "A similar package"
-    txt <- c (
-        "a not so very similar package",
-        "a test package",
-        "a function to test"
-    )
-    idfs <- get_test_idfs (txt)
-
-    # Text input:
-    expect_warning (
-        out <- pkgmatch_similar_pkgs (
-            input,
-            corpus = "ropensci",
-            idfs = idfs
-        ),
-        paste0 (
-            "ollama is not available. Matches will be ",
-            "based on word frequencies only"
-        )
-    )
-    expect_s3_class (out, "pkgmatch")
-    expect_equal (nrow (out), length (txt))
-
-    # package input:
-    path <- pkgmatch_test_skeleton (pkg_name = "demo")
-    roxygen2::roxygenise (path)
-    expect_warning (
-        out <- pkgmatch_similar_pkgs (
-            path,
-            corpus = "ropensci",
-            idfs = idfs,
-        ),
-        paste0 (
-            "ollama is not available. Matches will be ",
-            "based on word frequencies only"
-        )
-    )
-    expect_s3_class (out, "pkgmatch")
-    expect_equal (nrow (out), length (txt))
-
-    detach ("package:demo", unload = TRUE)
-    fs::dir_delete (path)
-})
-
-
 test_that ("similar pkgs text input cran", {
 
     withr::local_envvar (list (
-        "PKGMATCH_TESTS" = "true",
-        "pkgmatch.ollama.url" = "127.0.0.1:11434"
[…]
