### 2026-03-04 3e7a4cd
rm unused test snapshots and httptest2 mock files for #254
