### 2026-03-04 79658d1
rm couple refs to 'embeddings' #254
