+++ tests/testthat/test-bm25.R
@@ -1,6 +1,3 @@
-test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") ||
-    identical (Sys.getenv ("GITHUB_JOB"), "test-coverage"))
-
 test_that ("bm25", {
 
     pkgs <- c ("cli", "checkmate", "rappdirs")
@@ -18,10 +15,6 @@ test_that ("bm25", {
     expect_equal (ncol (res1), 2L)
     expect_identical (names (res1), c ("token", "idf"))
     expect_true (nrow (res0) != nrow (res1))
-
-    # Thie mocked call is only for embeddings from 'path', not for the
-    # cran-specific bits:
-    skip_if (!test_all)
 
     # Fails on gha windows machines at `cli::has_keypress_support()`
     input <- "A package which does some stuff"
