### 2026-03-04 778b087
improve error catching in cache fns and switch off one test for #254
