### 2026-03-04 717cd5d
tidy namespace for #254 + doc tweaks
