+++ R/cache.R
@@ -8,10 +8,9 @@
 #' \item "calls" for frequency tables for actual function calls.
 #' }
 #'
-#' This function is called within the main
-#' \link{pkgmatch_similar_pkgs} and \link{pkgmatch_similar_fns} functions
-#' to load required data there, and should not generally need to be explicitly
-#' called.
+#' This function is called within the main \link{pkgmatch_similar_pkgs}
+#' function to load required data there, and should not generally need to be
+#' explicitly called.
 #'
 #' @inheritParams pkgmatch_similar_pkgs
 #' @param what One of the three data types described above: "idfs",
