+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: pkgmatch
 Title: Find R Packages Matching Either Descriptions or Other R Packages
-Version: 0.5.1.034
+Version: 0.5.1.035
 Authors@R: c(
     person("Mark", "Padgham", , "mark.padgham@email.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2172-5265")),
@@ -21,8 +21,6 @@ Imports:
     fs,
     httr2,
     memoise,
-    methods,
-    pbapply,
     piggyback,
     Rcpp,
     rvest,
