### 2026-03-04 518fe6b
rm ollama and embeddings from pkgdown groups for #254
