### 2026-03-04 bc05716
rm refs to embeddings and ollama from readme for #254
