+++ README.md
@@ -18,16 +18,7 @@ on [CRAN](https://cran.r-project.org).
 
 ## Installation
 
-This package relies on a locally-running instance of
-[ollama](https://ollama.com). Procedures for setting that up are
-described in a [separate
-vignette](https://docs.ropensci.org/pkgmatch/articles/B_ollama.html)
-(`vignette("ollama", package = "pkgmatch")`). Although some
-functionality of this package may be used without ollama, the main
-functions require ollama to be installed.
-
-Once ollama is running, the easiest way to install this package is via
-the [associated
+The easiest way to install this package is via the [associated
 `r-universe`](https://ropensci-review-tools.r-universe.dev/). As shown
 there, simply enable the universe with
 
@@ -62,17 +53,11 @@ The package can then loaded for use with
 library (pkgmatch)
 ```
 
-The [`ollama_check()`
-function](https://docs.ropensci.org/pkgmatch/reference/ollama_check.html)
-can then be used to confirm that [ollama](https://ollama.com) is up and
-running as expected.
-
 ## Using the `pkgmatch` package
 
 The ‘pkgmatch’ package takes input either from a text description or
-local path to an R package, and finds matching packages based on both
-Language Model (LM) embeddings, and more traditional text and code
-matching algorithms.
+local path to an R package, and finds matching packages based on text
+and code matching algorithms.
 
 The package has two main functions:
 
@@ -99,8 +84,7 @@ on visualising inter-relationships among distinct trees.
 pkgmatch_similar_pkgs (input, corpus = "ropensci")
 ```
 
-    ## [1] "unifir"          "waywiser"        "medrxivr"        "tidyqpcr"       
-    ## [5] "googleLanguageR"
+    ## [1] "ssarp"   "rotl"    "stats19" "phruta"  "stplanr"
 
 The corpus parameter must be specified as one of “ropensci”, “cran”, or
 “bioc” (for [BioConductor](https://bioconductor.org); all
@@ -145,41 +129,14 @@ pkgmatch_similar_pkgs (path, corpus = "cran")
 ```
 
     ## $text
-    ## [1] "httr2"     "crul"      "httptest"  "httr"      "httptest2"
+    ## [1] "httr2"     "ellmer"    "geocodebr" "thisplot"  "httr"     
     ## 
     ## $code
-    ## [1] "PlotContour" "httr2"       "ellmer"      "paperplanes" "roundyh"
+    ## [1] "r4ds.tutorials"  "outliers.ts.oga" "ReliaShiny"      "secrfunc"       
+    ## [5] "strider"
 
 The result includes the top five matches based from both text and code
-of the input package. The input package itself is not the first-placed match in
-both cases. This happens because embeddings are “chunked” or randomly permuted,
-and because matches are statistical and not deterministic. See the [vignette on
-*Why are the results not what I
-expect?*](https://docs.ropensci.org/pkgmatch/articles/F_why-are-the-results-not-what-i-expect.html)
-for more detail on how matches are generated.
-
-## Finding functions
-
-There is an additional function to find functions within packages which
-best match a text description.
-
-``` r
-input <- "A function to label a set of geographic coordinates"
-pkgmatch_similar_fns (input)
-```
-
-    ## [1] "GSODR::nearest_stations"          "refsplitr::plot_addresses_points"
-    ## [3] "rnoaa::meteo_nearby_stations"     "charlatan::CoordinateProvider"   
-    ## [5] "quadkeyr::grid_to_polygon"
-
[…]
