### 2026-03-04 86377e9
rm ollama-related vignettes for #254
