### 2026-03-04 524f88f
rm refs to embeddings and ollama from main vignette for #254
