### 2026-03-04 b9d32a3
rm dockerfile and docker workflow for #254
