### 2026-03-04 70ba490
rm refs to embeddings in all vignettes for #254
