### 2026-03-04 df1811d
rm all final mentions of 'embeddings' and 'ollama' for #254
