+++ R/example-data.R
@@ -15,10 +15,7 @@
 #' @examples
 #' generate_pkgmatch_example_data ()
 #' input <- "curl" # Name of a single installed package
-#' # This generates a warning about "ollama" not being available:
-#' suppressWarnings (
-#'     pkgmatch_similar_pkgs (input, corpus = "cran")
-#' )
+#' pkgmatch_similar_pkgs (input, corpus = "cran")
 generate_pkgmatch_example_data <- function () {
 
     cli::cli_inform ("This function resets the cache directory used by 'pkgmatch'")
