### 2026-03-05 d5ce92a
rm ollama install from update workflow for #254
