+++ .github/workflows/update.yaml
@@ -32,15 +32,5 @@ jobs:
       - name: Install this package
         run: R CMD INSTALL .
 
-      - name: ollama
-        run: |
-          curl -fsSL https://ollama.com/install.sh | sh
-          ollama serve &
-          sleep 5
-          ollama pull jina/jina-embeddings-v2-base-en
-          sleep 5
-          ollama pull unclemusclez/jina-embeddings-v2-base-code
-          sleep 5
-
       - name: Call the update function
         run: Rscript -e "pkgmatch::pkgmatch_update_data(upload = TRUE)"
