+++ data-raw/release-data-script.R
@@ -100,6 +100,25 @@ options ("rlib_message_verbosity" = "verbose")
 path <- "/<path>/<to>/<cran-mirror>/tarballs"
 packages <- fs::dir_ls (path, regexp = "\\.tar\\.gz$")
 
+move_obsolete_pkgs <- function (path) {
+
+    packages <- fs::dir_ls (path, regexp = "\\.tar\\.gz$")
+    pkgs_tarballs <- grep ("\\.tar\\.gz$", packages, value = TRUE)
+    pkgs_names <- gsub ("\\_.*$", "", basename (pkgs_tarballs))
+    # Rm any old pkg versions:
+    nms_dup <- pkgs_names [which (duplicated (pkgs_names))]
+    if (length (nms_dup) > 0L) {
+        dups_to_rm <- unlist (lapply (nms_dup, function (n) {
+            tb <- pkgs_tarballs [which (pkgs_names == n)]
+            tb [tb < max (tb)]
+        }))
+        cli::cli_alert_danger ("Moving {length (dups_to_rm)} obsolete tarballs to Archive ...")
+        fs::file_move (dups_to_rm, fs::path (path, "Archive"))
+    }
+    fs::dir_ls (path, regexp = "\\.tar\\.gz$")
+}
+packages <- move_obsolete_pkgs (path)
+
 extract_packages <- function (packages) {
     path <- fs::path_common (packages)
     path_exdir <- fs::path (fs::path_dir (path), "temp")
@@ -109,12 +128,13 @@ extract_packages <- function (packages) {
 
     pkgs_tarballs <- grep ("\\.tar\\.gz$", packages, value = TRUE)
     pkgs_names <- gsub ("\\_.*$", "", basename (pkgs_tarballs))
+    pkgs_versions <- gsub ("^.*\\_|\\.tar\\.gz$", "", basename (pkgs_tarballs))
     pkgs_exdir <- fs::path (path_exdir, pkgs_names)
     pkgs_exdir_ls <- fs::dir_ls (path_exdir, type = "directory")
     index <- which (!pkgs_exdir %in% pkgs_exdir_ls)
     pkgs_todo <- pkgs_exdir [index]
     npkgs <- format (length (pkgs_todo), big.mark = ",")
-    if (npkgs > 0L) {
+    if (length (pkgs_todo) > 0L) {
         cli::cli_inform ("Extracting {npkgs} CRAN tarballs ...")
     }
 
@@ -128,31 +148,31 @@ extract_packages <- function (packages) {
         return (exdir)
     })
 
-    fs::dir_ls (path_exdir, type = "directory")
+    pkgs <- fs::dir_ls (path_exdir, type = "directory")
+    index <- which (pkgs_names %in% basename (pkgs))
+    data.frame (dir = pkgs, version = pkgs_versions [index], row.names = NULL)
 }
 
 cli::cli_h1 ("CRAN BM25")
 f <- "bm25-cran.Rds"
 if (!fs::file_exists (f)) {
-    pkg_version <- gsub ("^.*\\_|\\.tar\\.gz$", "", basename (packages))
     packages <- extract_packages (packages)
-    npkgs <- format (length (packages), big.mark = ",")
+    npkgs <- format (nrow (packages), big.mark = ",")
 
     cli::cli_inform ("Extract text from {npkgs} CRAN packages ...")
     num_cores <- parallel::detectCores () - 2L
     cl <- parallel::makeCluster (num_cores)
 
     txt_with_fns <- pbapply::pblapply (
-        packages,
+        packages$dir,
         function (p) pkgmatch:::get_pkg_text (p),
         cl = cl
     )
     parallel::stopCluster (cl)
 
-    index <- which (packages %in% names (txt_with_fns))
-    packages <- packages [index]
-    pkg_version <- pkg_version [index]
[…]
