### 2026-03-05 278d9c2
fix 'list_new_cran_updates' for #254
