+++ R/data-update-cran.R
@@ -159,18 +159,13 @@ list_new_cran_updates <- function (flist, latest_only = TRUE) {
     if (latest_only) {
         published <- as.Date (cran_db$Published [index])
         flist_remote <- list_remote_files ()
-        i <- which (flist_remote$file_name == basename (f_idf))
+        i <- which (flist_remote$file_name == basename (f_bm25))
         idfs_date <- as.Date (flist_remote$timestamp [i])
         dt <- difftime (idfs_date, published, units = "days")
         max_days <- 2L # allow published up to this many days before last update
         index <- index [which (dt <= max_days)]
     } # Otherwise update all pkgs regardless of dates ...
     cran_new <- cran_tarball [index]
-
-    # And include any which do not have data in all 5 structures:
-    extra <- names (pkgs) [which (pkgs < 5L)]
-    index <- match (extra, gsub ("\\_[0-9].*$", "", cran_tarball))
-    cran_new <- unique (c (cran_new, cran_tarball [index]))
 
     # Remove old versions from all data
     cran_new_pkg <- gsub ("\\_.*$", "", cran_new)
@@ -187,7 +182,6 @@ list_new_cran_updates <- function (flist, latest_only = TRUE) {
             nms <- gsub ("\\_.*$", "", names (bm25$token_list [[what]]))
             index <- which (!nms %in% pkgs_rm)
             bm25$token_lists [[what]] <- bm25$token_lists [[what]] [index]
-
         }
         n <- vapply (bm25$token_lists, length, integer (1L))
         if (!identical (n0, n)) {
