### 2026-03-05 413c3a1
minor tweak to update-cran fn
