### 2026-03-05 b34ef6d
catch error in base::parseNamespaceFile
That function sometimes errors in the pkgcheck Docker container
