### 2026-03-06 e338976
fix 'extract_tarball' when pkg has no R dir
