### 2026-03-06 ddfb949
fix update-ropensci fn for #254
