### 2026-03-06 81bfcf8
version bump -> 0.5.2
Also align release tag with this; for #254
