### 2026-03-11 3e0e9be
fix bioc bm25-fns structure for #280
