### 2026-03-11 f7280d5
reinstate similar-fns using bm25 only for #254
