+++ NAMESPACE
@@ -7,6 +7,7 @@ export(pkgmatch_bm25)
 export(pkgmatch_bm25_fn_calls)
 export(pkgmatch_browse)
 export(pkgmatch_load_data)
+export(pkgmatch_similar_fns)
 export(pkgmatch_similar_pkgs)
 export(pkgmatch_treesitter_fn_tags)
 export(pkgmatch_update_cache)
