### 2026-03-11 128a28c
fix browse for similar fns, and also for bioc #280
