### 2026-03-11 39bbac3
update readme with similar_fns reinstated
