+++ README.Rmd
@@ -92,6 +92,8 @@ Packages for analysing evolutionary trees, with a particular focus
 on visualising inter-relationships among distinct trees.
 "
 pkgmatch_similar_pkgs (input, corpus = "ropensci")
+input <- "Visualise evolutionary trees"
+pkgmatch_similar_fns (input, corpus = "ropensci")
 ```
 
 The corpus parameter must be specified as one of "ropensci", "cran", or "bioc"
@@ -99,20 +101,26 @@ The corpus parameter must be specified as one of "ropensci", "cran", or "bioc"
 corpus is much larger than the rOpenSci or BioConductor corpora, and matching
 for `corpus = "cran"` will generally take notably longer.
 
-Websites of packages returned by [the `pkgmatch_similar_pkgs()`
-function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
+Websites of packages or functions returned by either [the `pkgmatch_similar_pkgs()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
+of [`pkgmatch_similar_fns()`
+functions](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_fns.html)
 can be automatically opened, either by calling the function with `browse =
-TRUE`, or by storing the return value of [the `pkgmatch_similar_pkgs()`
-function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
+TRUE`, or by storing the return value of these functions
 as an object and passing that to [the `pkgmatch_browse()`
 function](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_browse.html).
 
 ### Matching entire packages
 
 The `input` parameter can also specify an entire package, either as a local
-path to a package directory, or the name of an installed package. To
-demonstrate that, the following code downloads a `.tar.gz` file of the `httr2`
-package from CRAN:
+path to a package directory, or the name of an installed package. For any
+locally-installed packages, you can just use the name of the package:
+```{r, similar-pkgs-installed}
+pkgmatch_similar_pkgs (input = "httr2", corpus = "cran")
+```
+
+Alternatively, the `input` parameter can be a local path to a package
+repository. To demonstrate that, the following code downloads a `.tar.gz` file
+of the same `httr2` package from CRAN:
 
 ```{r download-httr2}
 pkg <- "httr2"
@@ -130,7 +138,7 @@ passed to the
 [`pkgmatch_similar_pkgs()`](https://docs.ropensci.org/pkgmatch/reference/pkgmatch_similar_pkgs.html)
 function:
 
-```{r pkgmatch-similar-cran}
+```{r pkgmatch-similar-tarball}
 pkgmatch_similar_pkgs (path, corpus = "cran")
 ```
 
@@ -153,11 +161,11 @@ which describes several different example applications of `pkgmatch`, and
 illustrates the ways by which this package provides different kind of results
 to search engines and to general language model interfaces.
 - [_How does pkgmatch
-work?_](https://docs.ropensci.org/pkgmatch/articles/C_how-does-it-work.html)
+work?_](https://docs.ropensci.org/pkgmatch/articles/B_how-does-it-work.html)
 which provides detailed explanations of the matching algorithms implemented in
 the package.
 - [_Data caching and
-updating_](https://docs.ropensci.org/pkgmatch/articles/D_data-caching-and-updating.html)
+updating_](https://docs.ropensci.org/pkgmatch/articles/C_data-caching-and-updating.html)
 which describes how `pkgmatch` caches and updates the language model results
 for the individual corpora.
 
