### 2026-03-11 e81bbe0
reduce bm25 for sim_pkgs to only text_wo_fns
Entirely remove calculation with 'text_with_fns'
