### 2026-03-11 fb877c4
reduce similar-pkgs fn to use single bm25 from full pkg text only
no text with fns, and no bm25 from fn call freq tables
