### 2026-03-12 8232926
rm idf clean-up lines in similar-pkgs
