### 2026-03-12 7552a16
simplify similar_pkgs by removing embedding dls for fn de #254
