### 2026-03-12 91db759
improve pkg text extraction and include .Rnw files #254
