+++ R/get-pkg-data.R
@@ -1,181 +1,4 @@
-get_Rd_metadata <- utils::getFromNamespace (".Rd_get_metadata", "tools") # nolint
-
-get_pkg_text <- function (pkg_name) {
-
-    m_get_pkg_text (pkg_name)
-}
-
-get_pkg_text_internal <- function (pkg_name) {
-
-    if (pkg_is_installed (pkg_name) && !fs::dir_exists (pkg_name)) {
-        txt <- get_pkg_text_namespace (pkg_name)
-    } else {
-        txt <- get_pkg_text_local (pkg_name)
-    }
-
-    return (txt)
-}
-m_get_pkg_text <- memoise::memoise (get_pkg_text_internal)
-
-get_pkg_text_namespace <- function (pkg_name) {
-
-    # Suppress no visible binding notes:
-    Package <- NULL
-
-    stopifnot (length (pkg_name) == 1L)
-
-    desc <- utils::packageDescription (pkg = pkg_name, fields = "Description")
-    desc <- gsub ("\\n", " ", desc)
-    desc <- gsub ("\\s+", " ", desc)
-    desc <- c (
-        desc_template (pkg_name, desc),
-        "## Functions",
-        ""
-    )
-
-    fns <- get_fn_descs_from_ns (pkg_name)
-    fns <- lapply (seq_len (nrow (fns)), function (i) {
-        c (
-            paste0 ("### ", gsub ("\\.Rd$", "", fns$rd_name [i])),
-            "",
-            fns$desc [i],
-            ""
-        )
-    })
-
-    ip <- data.frame (utils::installed.packages ()) |>
-        dplyr::filter (Package == pkg_name)
-    rmds <- NULL
-    if (nrow (ip) > 0L) {
-        pkg_path <- fs::path (ip$LibPath, pkg_name)
-        rmd_files <- fs::dir_ls (pkg_path, recurse = TRUE, regexp = "\\.(R)?md")
-        rmd_files <- rmd_files [which (!duplicated (fs::path_file (rmd_files)))]
-        rmds <- unname (unlist (lapply (rmd_files, extract_one_md)))
-    }
-
-    paste0 (c (
-        desc,
-        "",
-        "## Vignettes",
-        rmds,
-        unlist (fns)
-    ), collapse = "\n ")
-}
-
-get_fn_descs_from_ns <- function (pkg_name) {
-
-    rd <- tools::Rd_db (package = pkg_name)
-    descs <- vapply (rd, function (i) {
-        get_Rd_metadata (i, "description")
-    }, character (1L))
-    descs <- gsub ("\\\\n", " ", descs)
-    descs <- gsub ("\\n", " ", descs)
-    descs <- gsub ("\\", "", descs, fixed = TRUE)
-    descs <- gsub ("\\", "", descs, fixed = TRUE)
-    descs <- gsub ("\\s+", " ", descs)
-
-    index <- which (!is.na (descs))
-    data.frame (
-        desc = unname (descs),
-        rd_name = names (descs)
-    ) [index, ]
-}
-
-desc_template <- function (pkg_name, desc) {
-    c (
-        paste0 ("# ", pkg_name, "\n"),
-        "",
-        "## Description",
-        "",
-        desc,
-        ""
-    )
-}
-
-get_pkg_text_local <- function (path) {
-
-    stopifnot (length (path) == 1L)
-
-    path <- fs::path_norm (path)
-
-    is_tarball <- fs::path_ext (path) == "gz"
-    if (is_tarball) {
-        path <- extract_tarball (path)
-        on.exit ({
-            fs::dir_delete (path)
-        })
-    }
-
-    stopifnot (fs::dir_exists (path))
-
-    desc_file <- fs::path (path, "DESCRIPTION")
[…]
