### 2026-03-12 14a9939
update 'release-data-script' to only analyse pkg text wo fns #254
