### 2026-03-12 b221a24
simplify 'data-update.R' for #254
