### 2026-03-13 748edd0
fix function name+desc extraction for #254
