+++ tests/testthat/test-data-update.R
@@ -23,38 +23,18 @@ test_that ("data update extract from local dir", {
     expect_equal (names (dat$bm25), c ("idfs", "token_lists"))
 
     # -------- test bm25 idfs --------
-    expect_type (dat$bm25$idfs, "list")
-    expect_length (dat$bm25$idfs, 2L)
-    expect_equal (names (dat$bm25$idfs), c ("with_fns", "wo_fns"))
-
-    expect_s3_class (dat$bm25$idfs$with_fns, "data.frame")
-    expect_equal (names (dat$bm25$idfs$with_fns), c ("token", "idf"))
-    expect_true (nrow (dat$bm25$idfs$with_fns) > 10L)
-
-    expect_s3_class (dat$bm25$idfs$wo_fns, "data.frame")
-    expect_equal (names (dat$bm25$idfs$wo_fns), c ("token", "idf"))
-    expect_true (nrow (dat$bm25$idfs$wo_fns) > 10L)
-    expect_true (nrow (dat$bm25$idfs$wo_fns) <
-        nrow (dat$bm25$idfs$with_fns))
+    expect_s3_class (dat$bm25$idfs, "data.frame")
+    expect_equal (ncol (dat$bm25$idfs), 2L)
+    expect_named (dat$bm25$idfs, c ("token", "idf"))
+    expect_true (nrow (dat$bm25$idfs) > 10L)
 
     # -------- test bm25 token_lists --------
     expect_type (dat$bm25$token_lists, "list")
-    expect_length (dat$bm25$token_lists, 2L)
-    expect_equal (names (dat$bm25$token_lists), c ("with_fns", "wo_fns"))
+    expect_length (dat$bm25$token_lists, 1L)
 
-    expect_type (dat$bm25$token_lists$with_fns, "list")
-    expect_length (dat$bm25$token_lists$with_fns, 1L)
-    expect_s3_class (dat$bm25$token_lists$with_fns [[1]], "data.frame")
-    expect_equal (names (dat$bm25$token_lists$with_fns [[1]]), c ("token", "n"))
-    expect_true (nrow (dat$bm25$token_lists$with_fns [[1]]) > 10L)
-
-    expect_type (dat$bm25$token_lists$wo_fns, "list")
-    expect_length (dat$bm25$token_lists$wo_fns, 1L)
-    expect_s3_class (dat$bm25$token_lists$wo_fns [[1]], "data.frame")
-    expect_equal (names (dat$bm25$token_lists$wo_fns [[1]]), c ("token", "n"))
-    expect_true (nrow (dat$bm25$token_lists$wo_fns [[1]]) > 10L)
-    expect_true (nrow (dat$bm25$token_lists$wo_fns [[1]]) <
-        nrow (dat$bm25$token_lists$with_fns [[1]]))
+    expect_s3_class (dat$bm25$token_lists [[1]], "data.frame")
+    expect_equal (names (dat$bm25$token_lists [[1]]), c ("token", "n"))
+    expect_true (nrow (dat$bm25$token_lists [[1]]) > 10L)
 
     # -------- test bm25_fns --------
     expect_type (dat$bm25_fns, "list")
@@ -93,19 +73,13 @@ test_that ("data update append to bm25", {
 
     pkgs <- c ("cli", "checkmate", "rappdirs")
     txt_with_fns <- lapply (pkgs, get_pkg_text)
-    txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
-    code <- lapply (pkgs, get_pkg_code)
-    names (txt_with_fns) <- names (txt_wo_fns) <- names (code) <- pkgs
+    txt <- rm_fns_from_pkg_txt (txt_with_fns)
[…]
