### 2026-03-13 ddde2f5
fix 'append_data_to_bm25' fn for #254
