### 2026-03-13 7b054bd
fix test-bm25.R for #254
