### 2026-03-13 bc679b5
fix 'test-similar' for #254
