### 2026-03-13 6490b1f
fix test-pkg-data.R for #254
