### 2026-03-13 5c50cc4
fix example-data generation for #254
