### 2026-03-13 ebeaefe
reinstate missing httptest2 mock call for #254
Mocking is needed to get the 'pkg_fns_from_r_search' result for fn call namespaces
