### 2026-03-13 630bc83
rm local paths from all object names for #254
