+++ data-raw/release-data-script.R
@@ -21,6 +21,7 @@ if (!all (fs::file_exists (f))) {
     txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
     idfs <- bm25_idf (txt_wo_fns)
     token_lists <- bm25_tokens_list (txt_wo_fns)
+    names (token_lists) <- basename (names (token_lists))
     bm25_data <- list (idfs = idfs, token_lists = token_lists)
     saveRDS (bm25_data, f [1])
 
@@ -171,6 +172,7 @@ if (!fs::file_exists (f)) {
     txt_wo_fns <- rm_fns_from_pkg_txt (txt_with_fns)
     idfs <- bm25_idf (txt_wo_fns)
     token_lists <- bm25_tokens_list (txt_wo_fns)
+    names (token_lists) <- basename (names (token_lists))
 
     bm25_data <- list (idfs = idfs, token_lists = token_lists)
     saveRDS (bm25_data, f)
