### 2026-03-13 776d084
fix matching cran pkg names with versions in bm25
