### 2026-03-13 497bbd2
fix bm25 pkg names in cran data
