+++ tests/testthat/_snaps/cache.md
@@ -1,0 +1,11 @@
+# dl messages
+
+    Code
+      withr::with_envvar(list(PKGMATCH_TESTS = "true"), msg <- send_dl_message(flist))
+    Message
+      Data already exist in 'tempdir' and will now be overwritten.
+      This directory may be safely deleted at any time.
+      See the pkgmatch 'Data caching and updating' vignette for details.
+      Data for the cran corpus comprises 3 files totalling around 0.47MB.
+      ! Your environment does not support key entry, downloading will now proceed.
+
