### 2026-03-13 9f190d3
more test snapshot redaction
