### 2026-03-13 cf444d4
test .Rnw extraction
