### 2026-03-13 022e6e2
text 'get_fn_defs_local' fn
