### 2026-03-13 cc6a194
fix 'data-update-bioc' to close #260; ref #254
