### 2026-03-13 15ae8ff
also improve cran update fn for #254
