### 2026-03-13 c25fa75
entirely rm now unused rerank.R routines for #254
