+++ tests/testthat/_snaps/cache.md
@@ -8,3 +8,4 @@
       See the pkgmatch 'Data caching and updating' vignette for details.
       Data for the cran corpus comprises 3 files totally around 1MB.
       ! Your environment does not support key entry, downloading will now proceed.
+
