### 2026-03-16 ae0b55f
add 'minchars' param to all tokens fns #254
To strip tokens down to only those with >= that many characters, and remove punctuation and other junk
