### 2026-03-16 e3cc51b
add 'minchar' pars to all data-update fns #254
