### 2026-03-16 badf68a
update tests with new fns data structure for #254
