### 2026-03-16 ac670c0
update 'release-data-script' with new data structures #254
