### 2026-03-17 b0396e3
fix update fn-calls data for #254
