### 2026-03-17 2296779
static-cast one cpp var
