### 2026-03-17 80f5c87
exclude news by default
And re-design pkg text extraction to ensure similar results from local vs installed
