### 2026-03-17 a0aa3d5
separate 'pkg_text_from_local_rds' fn for #254
