### 2026-03-17 13aeb2e
add 'sec_separator' sub-fn for systematically greppable sections #254
