+++ tests/testthat/test-pkg-data.R
@@ -10,15 +10,17 @@ test_that ("get pkg local text", {
     txt <- get_pkg_text (path)
     expect_type (txt, "character")
     expect_length (txt, 1L)
-    expect_identical (fns_separator, "## ---- Functions ----")
-    expect_true (grepl (fns_separator, txt, fixed = TRUE))
-    expect_false (grepl (fns_separator_ptn, txt, fixed = FALSE))
+    ptn0 <- sec_separator ("Functions", regex = FALSE)
+    ptn <- sec_separator ("Functions", regex = TRUE)
+    expect_identical (ptn0, "## ---- Functions ----")
+    expect_true (grepl (ptn0, txt, fixed = TRUE))
+    expect_false (grepl (ptn, txt, fixed = FALSE))
     expect_true (grepl ("#\\s*demo", txt))
     expect_true (nchar (txt) < 1000) # small test package
 
     txt <- strsplit (txt, "\\n") [[1]]
-    expect_true (length (grep (fns_separator, txt, fixed = TRUE)) == 1L)
-    expect_false (length (grepl (fns_separator_ptn, txt, fixed = FALSE)) == 1L)
+    expect_true (length (grep (ptn0, txt, fixed = TRUE)) == 1L)
+    expect_false (length (grepl (ptn, txt, fixed = FALSE)) == 1L)
 
     # detach is critical here, because httptest2 uses `utils::sessionInfo()`,
     # which checks namespaces and tries to load DESC file from pkg location.
