### 2026-03-17 c61d047
unify description format from internal vs source #254
