### 2026-03-17 2ead411
properly exclude news from text output by default #254
