### 2026-03-17 2212693
fix rnw gsub's #254 when encoding errors
