+++ R/get-pkg-data.R
@@ -102,12 +102,20 @@ extract_one_rnw <- function (rnw_file) {
     if (length (index) > 0L) {
         rnw <- rnw [-index]
     }
-    rnw <- gsub ("^\\\\item", "", rnw)
 
-    # Remove any direct "\...{" commands:
-    rnw <- gsub ("\\\\[^\\\\]*\\{", "", rnw, perl = TRUE)
-    # And the terminal "}" after the command:
-    rnw <- gsub ("(?<=[[:alpha:]])\\}", "", rnw, perl = TRUE)
+    # Rm: (1) items; (2) Direct "\...{" commands; (3) terminal "}" after
+    # commands:
+    ptns <- c ("^\\\\(item|hline|line)", "\\\\[^\\\\]*\\{", "(?<=[[:alpha:]])\\}")
+    perl <- c (FALSE, TRUE, TRUE)
+    for (p in seq_along (ptns)) {
+        rnw_sub <- tryCatch (
+            gsub (ptns [p], "", rnw, perl = perl [p]),
+            error = function (e) NULL
+        )
+        if (!is.null (rnw_sub)) {
+            rnw <- rnw_sub
+        }
+    }
 
     return (rnw)
 }
