### 2026-03-17 f191ca0
also exclude license file from parsed pkg text #254
