### 2026-03-18 e74bbaa
fix collapse par in 'rm_fns_from_pkg_txt'
