### 2026-03-18 3bc136e
add 'get_pkg_desc_from_pkg_txt' fn for #254
In prep to generate 2 BM25 scores for full text, and description only
