+++ data-raw/release-data-script.R
@@ -7,6 +7,14 @@ minchar <- 3L
 path <- "/<path>/<to>/ropensci/"
 packages <- fs::dir_ls (path, type = "directory")
 
+make_bm25 <- function (txt, minchar = 3L) {
+
+    idfs <- bm25_idf (txt, minchar = minchar)
+    token_lists <- bm25_tokens_list (txt, minchar = minchar)
+    names (token_lists) <- basename (names (token_lists))
+    list (idfs = idfs, token_lists = token_lists)
+}
+
 # -------------------- BM25 FOR ROPENSCI --------------------
 cli::cli_h1 ("rOpenSci BM25")
 f <- c ("bm25-ropensci.Rds", "bm25-ropensci-fns.Rds")
@@ -21,11 +29,13 @@ if (!all (fs::file_exists (f))) {
     parallel::stopCluster (cl)
 
     txt <- rm_fns_from_pkg_txt (txt_with_fns)
-    idfs <- bm25_idf (txt, minchar = minchar)
-    token_lists <- bm25_tokens_list (txt, minchar = minchar)
-    names (token_lists) <- basename (names (token_lists))
-    bm25_data <- list (idfs = idfs, token_lists = token_lists)
-    saveRDS (bm25_data, f [1])
+    descs <- get_pkg_desc_from_pkg_txt (txt_with_fns)
+    bm25 <- list (
+        full = make_bm25 (txt),
+        descs_only = make_bm25 (descs)
+    )
+
+    saveRDS (bm25, f [1])
 
     txt_fns <- get_all_fn_descs (txt_with_fns)
     fns_idfs <- bm25_idf (txt_fns$desc, minchar = minchar)
@@ -187,12 +197,12 @@ if (!fs::file_exists (f)) {
     txt_with_fns <- txt_with_fns [index]
 
     txt <- rm_fns_from_pkg_txt (txt_with_fns)
-    idfs <- bm25_idf (txt, minchar = minchar)
-    token_lists <- bm25_tokens_list (txt, minchar = minchar)
-    names (token_lists) <- basename (names (token_lists))
-
-    bm25_data <- list (idfs = idfs, token_lists = token_lists)
-    saveRDS (bm25_data, f)
+    descs <- get_pkg_desc_from_pkg_txt (txt_with_fns)
+    bm25 <- list (
+        full = make_bm25 (txt),
+        descs_only = make_bm25 (descs)
+    )
+    saveRDS (bm25, f)
 } else {
     cli::cli_inform ("skipping coz already done.")
 }
