### 2026-03-18 815f5ad
rm 'R/data-update.bioc'; update 'release-data-script'
Bioc generation is identical to rOpenSci, but needs no update, so simple
flags added in release-data-script to generate either full rOpenSci or Bioc
in identical ways.
