### 2026-03-18 4966ea2
use updated full+descs_only data structures in similar_pkgs fn #254
Actually returning ranks from full only at this point; next step is to reinstate
'rerank.R', and combine ranks again
