+++ tests/testthat/helper-idfs.R
@@ -5,12 +5,18 @@ mknm <- function (nchar = 12) {
 
 npkgs <- nfns <- 10L
 
-get_test_idfs <- function (txt, seed = 1L) {
+get_test_idfs1 <- function (txt) {
 
-    set.seed (seed)
     pkg_nms <- vapply (seq_along (txt), function (i) mknm (), character (1L))
 
     token_lists <- bm25_tokens_list (txt)
     names (token_lists) <- pkg_nms
     list (idfs = bm25_idf (txt), token_lists = token_lists)
+}
+
+get_test_idfs <- function (txt, seed = 1L) {
+
+    set.seed (seed)
+    i <- get_test_idfs1 (txt)
+    list (full = i, descs_only = i)
 }
