### 2026-03-18 c88a506
reinstate 'rerank.R', and combine pkg-match full+desc-only #254
