### 2026-03-18 eca32fc
re-render readme with updates from #254
