### 2026-03-18 872eba4
update 'example-data' and all fn exs with new data structures #254
