+++ man/text_is_code.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
+% Please edit documentation in R/input-assertions.R
 \name{text_is_code}
 \alias{text_is_code}
 \title{Estimate whether input text is code or English prose text.}
