### 2026-03-18 b32781c
fix wrong fn name -> 'sec_separator'
