### 2026-03-18 2508246
update 'cran-update' fns with new 'full/descs-only' data structures #254
