### 2026-03-18 7982a37
memoise both main fns
