### 2026-03-18 d93f40a
fix cran version column in sim-pkgs output for #254
