### 2026-03-18 2db6ebf
update fn docs for #254
