### 2026-03-18 689c6f1
rm 'text_is_code' enitrely for#254
