### 2026-03-18 99b8bed
rm one 'no vis binding' note
