### 2026-03-18 a6cd945
update return val docs for sim-pkgs #254
