### 2026-03-18 05bb3ca
update all vignettes for #254
