### 2026-03-20 3f0ad2f
improve verbose timing in 'update_cran()'
