### 2026-03-24 9115765
fix descr formatting
