### 2026-03-24 1ef17cf
rm angle brackets from sim_fns docs
win-builder rejects those outright
