### 2026-03-24 eab08c3
fix problematic gh url -> cran pkg version
