### 2026-03-27 #671 @MargaretSiple-NOAA
Hi @mpadge , this looks awesome! @agricolamz and @Selbosh I think these updates are good -- if you approve them can you approve with the [template](https://github.com/ropensci/software-review/issues/671#issuecomment-3697233936)? If not, we can continue to hash it out here. If possible, I would like to move this along before I leave for summer survey season at the end of May. Thank you all for your hard work and patience.
