### 2026-03-27 #671 @agricolamz
Hi all, I'll try to review the result at the end of April, when I'll be back from the field trip.
