### 2026-04-02 5868c30
fix rbuildignore to build tarball for cran
