### 2026-04-02 ce77003
add cran-comments for 1st cran release
