+++ cran-comments.md
@@ -1,0 +1,15 @@
+# CRAN notes for pkgmatch_0.5.3 submission
+
+This is a first submission.
+
+## Test environments
+
+The package has been checked on all environments listed below, and generates only the single note identifying the package as a new submission.
+
+GitHub actions:
+* Linux: R-release, R-devel
+* OSX: R-release
+* Windows: R-release
+
+CRAN win-builder:
+* R-oldrelease, R-release, R-devel
