### 2026-04-09 b0cc261
update desc for #290
