+++ man/pkgmatch_treesitter_fn_tags.Rd
@@ -42,11 +42,9 @@ fn_tags <- pkgmatch_treesitter_fn_tags ("curl") # Name of installed package
 fn_tags <- pkgmatch_treesitter_fn_tags ("cli") # Name of installed package
 
 # Or get calls from full source code:
-u <- "https://cran.r-project.org/src/contrib/odbc_1.5.0.tar.gz"
+u <- "https://cran.r-project.org/src/contrib/Archive/odbc/odbc_1.5.0.tar.gz"
 path <- file.path (tempdir (), basename (u))
-\dontrun{
-download.file (u, destfile = path)
+download.file (u, destfile = path, quiet = TRUE)
 fn_tags <- pkgmatch_treesitter_fn_tags (path)
-}
 }
 \concept{treesitter}
