### 2026-04-14 b97e656
fix vignette formatting
