### 2026-04-14 26554e5
fix 'update_cache' fn for #290
