### 2026-04-14 bddc256
link to update_cache fn in caching vignette
