### 2026-04-14 fb9edda
update example data for #292
