+++ R/example-data.R
@@ -32,11 +32,15 @@ generate_pkgmatch_example_data <- function () {
     Sys.setenv ("PKGMATCH_CACHE_DIR" = ex_dir)
     options ("pkgmatch.example_env" = "true")
 
-    corpus <- "cran"
-    fnames <- c ("bm25", "idfs-fn-calls", "fn-calls")
-    fnames_full <- fs::path (ex_dir, paste0 (fnames, "-", corpus, ".Rds"))
+    corpus <- "ropensci"
+    fnames <- c ("bm25", "fn-calls")
+    fnames_full <- fs::path (ex_dir, c (
+        paste0 (fnames, "-", corpus, ".Rds"),
+        paste0 ("bm25-", corpus, "-fns.Rds")
+    ))
+    fn_names <- c ("bm25", "idfs_fn_calls", "fn_calls")
     index <- which (!fs::file_exists (fnames_full))
-    fnames <- data.frame (name = fnames, path = fnames_full) [index, ]
+    fnames <- data.frame (name = fn_names, path = fnames_full) [index, ]
 
     # Best matching packages against "curl" package for text and code:
     pkg_nms <- c (
@@ -96,10 +100,28 @@ ex_words <- function () {
 }
 
 ex_idfs_fn_calls <- function (pkg_nms, fname) {
-    ip <- data.frame (utils::installed.packages ())
-    fns <- lapply (seq_len (nrow (ip)), function (i) {
+
+    sample_pkgs <- c (
+        "curl",
+        "cli",
+        "fs",
+        "httr",
+        "memoise",
+        "tokenizers"
+    )
+
+    fns <- lapply (sample_pkgs, function (p) {
+
+        lp <- .libPaths ()
+        index <- vapply (.libPaths (), function (i) {
+            fs::dir_exists (fs::path (i, p))
+        }, logical (1L))
+        if (!any (index)) {
+            return (NULL)
+        }
+        lp <- lp [which (index)] [1L]
         ns <- tryCatch (
-            parseNamespaceFile (ip$Package [i], ip$LibPath [i]),
+            parseNamespaceFile (p, lp),
             error = function (e) NULL
         )
         if (!is.null (ns)) {
@@ -108,7 +130,7 @@ ex_idfs_fn_calls <- function (pkg_nms, fname) {
         res <- NULL
         if (length (ns) > 0) {
             res <- data.frame (
-                token = paste0 (ip$Package [i], "::", ns),
+                token = paste0 (p, "::", ns),
                 idf = 10 - stats::rgamma (length (ns), shape = 1)
             )
         }
@@ -121,25 +143,44 @@ ex_idfs_fn_calls <- function (pkg_nms, fname) {
 }
 
 ex_fn_calls <- function (pkg_nms, fname) {
-    ip <- data.frame (utils::installed.packages ())
 
-    tags <- NULL
-    while (is.null (tags)) {
-        i <- sample (nrow (ip), size = 1L)
-        pkg <- ip$Package [i]
-        tags <- tryCatch (
+    sample_pkgs <- c (
+        "curl",
+        "cli",
+        "fs",
+        "httr",
+        "memoise",
+        "tokenizers"
+    )
+
+    tags <- lapply (sample_pkgs, function (p) {
+        tryCatch (
             suppressMessages (
-                pkgmatch_treesitter_fn_tags (pkg)
+                pkgmatch_treesitter_fn_tags (p)
             ),
             error = function (e) NULL
         )
-    }
-    fns <- tags$name [which (!grepl (paste0 ("^", pkg), tags$name))]
-    fn_tbl <- table (fns)
+    })
 
[…]
