### 2026-04-14 9249441
fix test-cache with updated example data #292
