### 2026-05-04 #671 @Selbosh
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 6
